### 2024-07-13 #6980 @editorialbot
[PRE REVIEW]: DistrictGenerator: Generating Building-Specific Load Profiles for Residential Districts
**Submitting author:** @JoelSchoelzel (<a href="http://orcid.org/0009-0000-5053-3161">Joel Schoelzel</a>)
**Repository:** https://github.com/RWTH-EBC/districtgenerator
**Branch with paper.md** (empty if default branch): JOSS_submission
**Version:** v0.1.0
