+++ EHDO/load_params.py
@@ -494,101 +494,224 @@ def load_params(data):
     ################################################################
     # LOAD MODEL PARAMETERS
 
-    # load economic and ecologic data (of the district generator)
-    ecoData = {}
-    with open(os.path.join(os.path.dirname(srcPath), 'data', 'eco_data.json')) as json_file:
-        jsonData = json.load(json_file)
-        for subData in jsonData:
-            ecoData[subData["name"]] = subData["value"]
 
-    ### Energy costs ###
-    # Electricity costs
-    if param["enable_supply_el"] == "True":
-        param["enable_supply_el"] = True
-    elif param["enable_supply_el"] == "False":
-        param["enable_supply_el"] = False
+
+    model_parameter = {}
+    model_parameter["webtool"] = {}
+    if model_parameter["webtool"] == {}:
+
+        # load economic and ecologic data (of the district generator)
+        ecoData = {}
+        with open(os.path.join(os.path.dirname(srcPath), 'data', 'eco_data.json')) as json_file:
+            jsonData = json.load(json_file)
+            for subData in jsonData:
+                ecoData[subData["name"]] = subData["value"]
+
+        ### Energy costs ###
+        # Electricity costs
+        if param["enable_supply_el"] == "True":
+            param["enable_supply_el"] = True
+        elif param["enable_supply_el"] == "False":
+            param["enable_supply_el"] = False
+        else:
+            sys.exit("Invalid input!")
+        param["price_supply_el"] = ecoData["C_dem_electricity"]
+
+        param["price_cap_el"] = 0  # kEUR/MW
+        if param["enable_feed_in_el"] == "True":
+            param["enable_feed_in_el"] = True
+        elif param["enable_feed_in_el"] == "False":
+            param["enable_feed_in_el"] = False
+        else:
+            sys.exit("Invalid input!")
+        param["revenue_feed_in_el"] = ecoData["C_feed_electricity"]
+
+        param["enable_cap_limit_el"] = False
+        param["cap_limit_el"]        = 1000  # kW
+
+        param["enable_supply_limit_el"] = False
+        param["supply_limit_el"]        = 100000 # kWh/a
+
+
+        # Natural gas
+        if param["enable_supply_gas"] == "True":
+            param["enable_supply_gas"] = True
+        elif param["enable_supply_gas"] == "False":
+            param["enable_supply_gas"] = False
+        else:
+            sys.exit("Invalid input!")
+        param["price_supply_gas"] = ecoData["C_dem_gas"]
+        param["price_cap_gas"] = 0
+        param["enable_feed_in_gas"]      = False
+        param["revenue_feed_in_gas"]     = 0.01
+
+        param["enable_cap_limit_gas"]    = False
+        param["cap_limit_gas"]           = 100  # kW
+        param["enable_supply_limit_gas"] = False
+        param["supply_limit_gas"]        = 100000 # kWh/a
+
+
+        # Biomass
+        param["enable_supply_biomass"]    = False
+        param["price_biomass"]            = 0.2  # EUR/kWh
+        param["enable_supply_limit_biom"] = False
+        param["supply_limit_biomass"]     = 100000  #kWh
+
+
+        # Hydrogen
+        param["enable_supply_hydrogen"]       = False
+        param["price_hydrogen"]               = 0.4  # EUR/kWh
+        param["enable_supply_limit_hydrogen"] = False
+        param["supply_limit_hydrogen"]        = 100000  #kWh
+
+
+        # Waste
+        param["enable_supply_waste"]       = False
[…]
