### 2024-07-22 34ef1da
update weather data depending on postal code
