### 2024-08-08 302cb0e
Number of Occupants Electricity consumption
