+++ classes/users.py
@@ -8,7 +8,7 @@ import richardsonpy
 import richardsonpy.classes.stochastic_el_load_wrapper as wrap
 import richardsonpy.classes.appliance as app_model
 import richardsonpy.classes.lighting as light_model
-import functions.heating_profile_5R1C as heating
+import districtgenerator.functions.heating_profile_5R1C as heating
 
 
 class Users:
@@ -125,60 +125,60 @@ class Users:
         """
 
         if self.building == "SFH":
-            # choose random number of occupants (2-5) for single family houses (assumption)
-
-            # loop over all flats of current building
-            for j in range(self.nb_flats):
+            # choose random number of occupants (1-5) for single family houses  (assumption)
+            probabilities = (0.245114, 0.402323, 0.154148, 0.148869, 0.049623) # Probabilities of having 1, 2, 3, 4 or 5 occupants in a single-family house, assuming a maximum of 5 occupants. Sources: https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Wohnen/Tabellen/tabelle-wo2-mietwohnungen.html and https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Wohnen/Tabellen/tabelle-wo2-eigentuemerwohnungen.html
+            # loop over all flats of current single family house
+            for k in range(self.nb_flats):
                 random_nb = rd.random()  # picking random number in [0,1)
-                j = 1  # staring with one (additional) occupant
-                # the random number decides how many occupants are chosen (2-5)
-                while j <= 4:
-                    if random_nb < j / 4:
-                        self.nb_occ.append(1 + j)  # minimum is 2 occupants
+                j = 1  # staring with one occupant
+                # the random number decides how many occupants are chosen (1-5)
+                while j <= 5 :
+                    if random_nb < sum(probabilities[:j]) :
+                        self.nb_occ.append(j)  # minimum is 1 occupant
                         break
                     j += 1
 
         if self.building == "TH":
-            # choose random number of occupants (2-5) for terraced houses (assumption)
-
-            # loop over all flats of current building
-            for j in range(self.nb_flats):
+            # choose random number of occupants (1-5) for terraced houses  (assumption)
+            probabilities = (0.236817, 0.400092, 0.157261, 0.154371, 0.051457) # Probabilities of having 1, 2, 3, 4 or 5 occupants in a terraced house, assuming a maximum of 4 occupants. Sources: https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Wohnen/Tabellen/tabelle-wo2-mietwohnungen.html and https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Wohnen/Tabellen/tabelle-wo2-eigentuemerwohnungen.html
+            # loop over all flats of current terraced house
+            for k in range(self.nb_flats):
                 random_nb = rd.random()  # picking random number in [0,1)
-                j = 1  # staring with one (additional) occupant
-                # the random number decides how many occupants are chosen (2-5)
-                while j <= 4:
-                    if random_nb < j / 4:
-                        self.nb_occ.append(1 + j)  # minimum is 2 occupants
+                j = 1  # staring with one occupant
+                # the random number decides how many occupants are chosen (1-5)
+                while j <= 5 :
+                    if random_nb < sum(probabilities[:j]) :
+                        self.nb_occ.append(j)  # minimum is 1 occupant
                         break
                     j += 1
 
         if self.building == "MFH":
-            # choose random number of occupants (1-4) for each flat in the multifamily house (assumption)
-
-            # loop over all flats of current building
-            for j in range(self.nb_flats):
+            # choose random number of occupants (1-5) for each flat in the multi family house  (assumption)
[…]
