### 2024-08-08 bd9c107
Temperature difference dhw
