### 2024-08-08 6be2a21
Temperature difference dhw
