### 2024-08-09 a2c94d9
fix: fixes for the quartiersenergieausweis calculation
