+++ functions/opti_central.py
@@ -8,9 +8,12 @@ import os, json
 import gurobipy as gp
 import time
 
-def run_opti_central(model, buildingData, site, cluster, srcPath):
-    now = time.time()
+def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {}):
 
+    if optiData == {}:
+        optiData["webtool"] = False
+
+    now = time.time()
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
     # Setting up the model
     # number of buildings in neighborhood
@@ -33,6 +36,7 @@ def run_opti_central(model, buildingData, site, cluster, srcPath):
     STC_heat = {}   # electricity generation of PV [W]
     EV_dem = {}     # electricity demand electric vehicle (EV) [Wh]
     elec_dem = {}   # electricity demand for appliances and lighting [W]
+    occ = {}
     for n in range(nb):
         Q_DHW[n] = buildingData[n]["user"].dhw_cluster[cluster]
         Q_heating[n] = buildingData[n]["user"].heat_cluster[cluster]
@@ -40,6 +44,7 @@ def run_opti_central(model, buildingData, site, cluster, srcPath):
         STC_heat[n] = buildingData[n]["generationSTC_cluster"][cluster]
         EV_dem[n] = buildingData[n]["user"].car_cluster[cluster]
         elec_dem[n] = buildingData[n]["user"].elec_cluster[cluster]
+        occ[n] = buildingData[n]["user"].occ[:len(elec_dem[n])]
 
     ecoData = {}
     timeData = {}
@@ -64,7 +69,7 @@ def run_opti_central(model, buildingData, site, cluster, srcPath):
     ecs_power = ("HP", "EH", "CHP", "FC", "PV")  # power consuming/producing devices (photovoltaic (PV))
     ecs_sell = ("CHP", "FC", "PV", "BAT")
     ecs_gas = ("CHP", "FC", "BOI")  # gas consuming devices
-    ecs_storage = ("BAT", "TES") #, "EV")  # battery (BAT), thermal energy storage (TES), electric vehicle (EV)
+    ecs_storage = ("BAT", "TES", "EV")  # battery (BAT), thermal energy storage (TES), electric vehicle (EV)
     hp_modi = ("HP35", "HP55")  # modi of the HP with different HP supply temperatures in °C
 
     # %% TECHNICAL PARAMETERS
@@ -220,13 +225,13 @@ def run_opti_central(model, buildingData, site, cluster, srcPath):
     days = [0,1,2,3,4,5,6]
     daily_peak = {}
     for d in days:
-        daily_peak[d] = model.addVar(vtype="c",lb=-gp.GRB.INFINITY,name="peak_network_load")
+        daily_peak[d] = model.addVar(vtype="C",lb=-gp.GRB.INFINITY,name="peak_network_load")
     peaksum = model.addVar(vtype="c",lb=-gp.GRB.INFINITY,name="sum_peak_daily_network_load")
 
     # Total operational costs
     operational_costs = model.addVar(vtype="C",lb=-gp.GRB.INFINITY,name="Cost_total")
     # Total gross CO2 emissions
-    co2_total = model.addVar(vtype="c",lb=-gp.GRB.INFINITY,name="Emission_total")
+    co2_total = model.addVar(vtype="C",lb=-gp.GRB.INFINITY,name="Emission_total")
 
     # Objective function
     obj = model.addVar(vtype="C",lb=-gp.GRB.INFINITY,name="obj")
@@ -283,22 +288,7 @@ def run_opti_central(model, buildingData, site, cluster, srcPath):
[…]
