+++ examples/e1.4_generate first_district.py
@@ -14,10 +14,10 @@ def example1_4_generate_first_district():
     data = Datahandler()
 
     # Generate Environment for the District
-    data.generateEnvironment()
+    data.generateEnvironment(plz="0")
 
     # Initialize Buildings to the District
-    data.initializeBuildings(scenario_name="example")
+    data.initializeBuildings(scenario_name="Quartier_3")
 
     # Generate a more detailed Building
     data.generateBuildings()
