### 2024-08-14 ada91d8
fix: fixed quartiersenergieausweis optimization
