+++ classes/datahandler.py
@@ -212,7 +212,7 @@ class Datahandler:
                                         albedo=self.site["albedo"])
 
     # def initializeBuildings(self, scenario_name='example'):
-    def initializeBuildings(self, list_all: list = None, scenario_name='example'):
+    def initializeBuildings(self, list_all: list = None, scenario_name='example', path=None):
         """
         Fill district with buildings from scenario file.
 
@@ -229,31 +229,67 @@ class Datahandler:
         num_sfh = 0
         num_mfh = 0
         self.scenario_name = scenario_name
-        if list_all is not None:
+        if list_all is not None and len(list_all) != 0:
             self.scenario = pd.DataFrame(list_all, columns=['id', 'type', 'year', 'condition', 'area'])
         else:
-            self.scenario = pd.read_csv(os.path.join(self.filePath, 'scenarios')
-                                        + "/"
-                                        + self.scenario_name + ".csv",
-                                        header=0, delimiter=";")
-
-        for id, entry in enumerate(list_all):
-            list = list_all[id]
-            building = {}
-            building["buildingFeatures"] = pd.Series(
-                [list[0], list[1], list[2], list[3], list[4], 3, 1, 2, 3, 1, 2, 3, 1, 2, 3, 1],
-                index=['id', 'building', 'year', 'retrofit', 'area', 'heater', 'PV',
-                       'STC', 'EV', 'BAT', 'f_TES', 'f_BAT', 'f_EV', 'f_PV', 'f_STC',
-                       'gamma_PV', ])
-
-            self.district.append(building)
-
-            if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"]["building"] == 'TH':
-                num_sfh += 1
-            elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"]["building"] == 'AB':
-                num_mfh += 1
+            if path is None:
+                self.scenario = pd.read_csv(os.path.join(self.filePath, 'scenarios')
+                                            + "/"
+                                            + self.scenario_name + ".csv",
+                                            header=0, delimiter=";")
+            else:
+                self.scenario = pd.read_csv(os.path.join(path)
+                                            + "/"
+                                            + self.scenario_name + ".csv",
+                                            header=0, delimiter=";")
 
         # Calculate calculation time for the whole district generation
+        if list_all:
+            building_df_list = []
+            for id, entry in enumerate(list_all):
+                list = list_all[id]
+                building = {}
+                building["buildingFeatures"] = pd.Series(
+                    [list[0], list[1], list[2], list[3], list[4], "BOI", 0, 0, 0, 0, 20, 0, "S", 0.4, 0.04, 0],
+                    index=['id', 'building', 'year', 'retrofit', 'area', 'heater', 'PV',
+                           'STC', 'EV', 'BAT', 'f_TES', 'f_BAT', 'f_EV', 'f_PV', 'f_STC',
+                           'gamma_PV', ])
+
+                self.district.append(building)
+                # building_series_list.append(building["buildingFeatures"])
+                df = pd.DataFrame(data=building["buildingFeatures"]).transpose()
+                df.set_index("id")
+                # df.drop(df.columns[0], axis=1)
+                building_df_list.append(df)
+
+                if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"][
+                    "building"] == 'TH':
+                    num_sfh += 1
+                elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"][
+                    "building"] == 'AB':
+                    num_mfh += 1
+            file_path = os.path.join(path, 'files') + "/" + self.scenario_name + ".csv"
[…]
