+++ functions/opti_central.py
@@ -8,8 +8,8 @@ import os, json
 import gurobipy as gp
 import time
 
-def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {}):
 
+def run_opti_central(model, buildingData, site, cluster, srcPath, optiData={}):
     if optiData == {}:
         optiData["webtool"] = False
 
@@ -30,12 +30,12 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
                 param_dec_devs[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
 
     T_e = site["T_e_cluster"][cluster]  # ambient temperature [°C]
-    Q_DHW = {}      # DHW (domestic hot water) demand [W]
+    Q_DHW = {}  # DHW (domestic hot water) demand [W]
     Q_heating = {}  # space heating [W]
-    PV_gen = {}     # electricity generation of PV [W]
-    STC_heat = {}   # electricity generation of PV [W]
-    EV_dem = {}     # electricity demand electric vehicle (EV) [Wh]
-    elec_dem = {}   # electricity demand for appliances and lighting [W]
+    PV_gen = {}  # electricity generation of PV [W]
+    STC_heat = {}  # electricity generation of PV [W]
+    EV_dem = {}  # electricity demand electric vehicle (EV) [Wh]
+    elec_dem = {}  # electricity demand for appliances and lighting [W]
     occ = {}
     for n in range(nb):
         Q_DHW[n] = buildingData[n]["user"].dhw_cluster[cluster]
@@ -89,7 +89,7 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
         # Initial state of charge
         soc_init["TES"][n] = soc_nom["TES"][n] * 0.5  # Wh
         soc_init["BAT"][n] = soc_nom["BAT"][n] * 0.5  # Wh
-        soc_init["EV"][n] = soc_nom["EV"][n] * 0.5    # Wh
+        soc_init["EV"][n] = soc_nom["EV"][n] * 0.5  # Wh
 
         # %% TECHNICAL CONSTRAINTS
 
@@ -105,7 +105,8 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
         for n in range(nb):
             power_dom[device][n] = {}
             for t in time_steps:
-                power_dom[device][n][t] = model.addVar(vtype="C",name="power_" + device + "_n" + str(n) + "_t" + str(t))
+                power_dom[device][n][t] = model.addVar(vtype="C",
+                                                       name="power_" + device + "_n" + str(n) + "_t" + str(t))
 
     # Heat to/from devices
     heat_dom = {}
@@ -114,15 +115,16 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
         for n in range(nb):
             heat_dom[device][n] = {}
             for t in time_steps:
-                heat_dom[device][n][t] = model.addVar(vtype="C",name="heat_" + device + "_n" + str(n) + "_t" + str(t))
+                heat_dom[device][n][t] = model.addVar(vtype="C", name="heat_" + device + "_n" + str(n) + "_t" + str(t))
 
-    power_mode= {}
+    power_mode = {}
     for device in hp_modi:
         power_mode[device] = {}
         for n in range(nb):
             power_mode[device][n] = {}
             for t in time_steps:
-                power_mode[device][n][t] = model.addVar(vtype="C",name="power_mode_" + device + "_n" + str(n) + "_t" + str(t))
+                power_mode[device][n][t] = model.addVar(vtype="C",
[…]
