### 2024-08-15 033709c
Fixed invaldi escape squence
