+++ districtgenerator/datahandler.py
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 # -*- coding: utf-8 -*-
 
 import json
@@ -1031,639 +1030,3 @@ class Datahandler():
         self.KPIs = KPIs(self)
         # calculate KPIs
         self.KPIs.calculateAllKPIs(self)
-
-=======
-# -*- coding: utf-8 -*-
-
-import json
-import pickle
-import os
-from typing import Any
-import numpy as np
-import pandas as pd
-from itertools import count
-from teaser.project import Project
-from districtgenerator.envelope import Envelope
-from districtgenerator.solar import Sun
-from districtgenerator.users import Users
-from districtgenerator.plots import DemandPlots
-from districtgenerator.non_residential import NonResidential
-from districtgenerator.non_residential_users import NonResidentialUsers
-import functions.weather_handling as weather_handling 
-
-class Datahandler():
-    """
-    Abstract class for data handling
-    Collects data from input files, TEASER, User and Enevelope.
-
-    Attributes
-    ----------
-    site:
-        dict for site data, e.g. weather
-    time:
-        dict for time settings
-    district:
-        list of all buildings within district
-    scenario_name:
-        name of scenario file
-    scenario:
-        scenario data
-    counter:
-        dict for counting number of equal building types
-    srcPath:
-        source path
-    filePath:
-        file path
-    weatherFile::
-        path to weather file, which is not a TRY file. If None, the TRY file is used. 
-        file should follow the following format, according to DWD naming conventions 
-        temp_sunDirect = B  Direkte Sonnenbestrahlungsstaerke (horiz. Ebene)
-        temp_sunDiff = D Diffuse Sonnenbetrahlungsstaerke (horiz. Ebene)
-        temp_temp = t Lufttemperatur in 2m Hoehe ueber Grund 
-        direct_illuminance = direct illuminance, only present in epw file
-        diffuse_illuminance = diffuse illuminance, only present in epw file
-
-    sheetFile:
-        path to scenario file, which is a csv file. If None, the csv file is used.
-        Can be used to provide more detailed information for gerneation of the buildings models in the district
-       
-    """
-
-    def __init__(self, weather_file=None, sheet_file=None):
-        self.site = {}
-        self.time = {}
-        self.district = []
-        self.scenario_name = None
-        self.scenario = None
-        self.counter = {}
-        self.building_dict = {} # Dictionary to store Residential Building IDs
-        self.srcPath = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        self.filePath = os.path.join(self.srcPath, 'data')
-        self.resultPath = os.path.join(self.srcPath, 'results', 'demands')
-        self.weatherFile = weather_file
-        self.sheetFile = sheet_file
-        self.advancedModel = None
-        if weather_file is not None:
-            self.timestamp = weather_handling.get_time_horizon(self.weatherFile)
-
-
-    def setResultPath(self, new_path=None):
-            """
-            Sets the path where the results will be saved.
-
-            Args:
-                new_path (str, optional): The new path to set. If not provided, the default path will be used.
-
-            Returns:
-                None
-            """
-            self.resultPath = new_path if new_path is not None else os.path.join(self.srcPath, 'results', 'demands')
-    
-    def setAdvancedModel(self, pathAdvancedModel=None):
-            """
-            Sets the path and loads data for advanded modelling
-
-            Args:
-                new_path (str, optional): The new path to set. If not provided, the default path will be used.
-
-            Returns:
-                None
-            """
-            self.advancedModel = pathAdvancedModel if pathAdvancedModel is not None else None
-
-    def setWeatherFile(self, pathWeatherFile=None):
-            """
-            Sets the path and loads data for weather file
-
[…]
