### 2024-08-20 ca83770
fix: added new plz loading function and validation
