### 2024-08-21 93a5c2a
fix: added webapp interface for  weather conditions and try choice
