### 2024-08-28 1475c00
Type hints
