+++ districtgenerator/non_residential_users.py
@@ -15,6 +15,7 @@ import functions.heating_profile_5R1C as heating
 import functions.schedule_reader as schedules
 import functions.light_demand as light_demand
 from districtgenerator.solar import SunIlluminance
+from typing import Dict, List, Any, Optional
 
 
 class NonResidentialUsers():
@@ -99,7 +100,7 @@ class NonResidentialUsers():
         Loads profiles from previously saved CSV files.
 
     """
-    def __init__(self, building_usage, area, file, envelope, site, time, nb_of_days):
+    def __init__(self, building_usage: str, area: float, file: str, envelope: Any, site: Dict[str, Any], time: Dict[str, Any], nb_of_days: int) -> None:
         """
         Constructor of Users Class
         """
@@ -148,14 +149,13 @@ class NonResidentialUsers():
         # self.create_el_wrapper()
 
     
-    def load_json_data(self, json_path):
+    def load_json_data(self, json_path: str) -> Dict[str, Any]:
         with open(json_path, 'r') as file:
-            data = json.load(file)
-            return data
+            return json.load(file)
 
 
 
-    def generate_number_occupants(self):
+    def generate_number_occupants(self) -> None:
         '''
         Generate number of occupants for different of building types.
         According to the data in  data\occupancy_schedules\average_occupancy.json
@@ -182,7 +182,7 @@ class NonResidentialUsers():
         else:
             print(f"No data about number of occupants available for building type: {self.usage}")
 
-    def generate_schedules(self):
+    def generate_schedules(self) -> None:
         # get schedules occupancy
         df_schedules, schedule = schedules.get_schedule(self.usage)
         self.occupancy_schedule = schedules.adjust_schedule(inital_day= 0, schedule=df_schedules[["DAY", "HOUR", "OCCUPANCY"]], nb_days=self.nb_of_days)
@@ -190,10 +190,10 @@ class NonResidentialUsers():
         self.lighntning_schedule = schedules.adjust_schedule(inital_day= 0, schedule=df_schedules[["DAY", "HOUR", "LIGHTING"]], nb_days=self.nb_of_days)
 
     
-    def generate_occupancy(self):
+    def generate_occupancy(self) -> None:
         self.occ = self.occupancy_schedule["OCCUPANCY"] * self.nb_occ
         
-    def generate_annual_el_consumption_equipment(self, equipment="Mittel"):
+    def generate_annual_el_consumption_equipment(self, equipment: str = "Mittel") -> None:
         '''
         Generate annual elictricity consumption in dependency of the building type and the average area. 
         
@@ -219,8 +219,8 @@ class NonResidentialUsers():
         # such as calculation bases on users, tasks, e.g. 
         # Here the calculation is done, based on the average equipment in the building type 
         # For more information see: 
-        # [1] S. Henning and K. Jagnow, “Statistische Untersuchung der Flächen- und Nutzstromanateile von Zonen in Nichtwohngebäuden (Fortführung),” 51/2023, Jul. 2023. [Online]. Available: https://www.bbsr.bund.de/BBSR/DE/veroeffentlichungen/bbsr-online/2023/bbsr-online-51-2023-dl.pdf?__blob=publicationFile&v=3
-        # [2] K. Jagnow and S. Henning, “Statistische Untersuchung der Flächen- und Nutzstromanteile von Zonen in Nichtwohngebäuden,” Hochschule	Magdeburg-Stendal, Mar. 2020. [Online]. Available: https://www.h2.de/fileadmin/user_upload/Fachbereiche/Bauwesen/Forschung/Forschungsberichte/Endbericht_SWD-10.08.18.7-18.29.pdf
[…]
