+++ pyproject.toml
@@ -2,5 +2,5 @@
 max-line-length = 80
 
 [tool.pylint.basic]
-variable-naming-style = "snake_case"
-function-naming-style = "snake_case"
+variable-naming-style = "camelCase"
+function-naming-style = "camelCase"
