### 2024-08-28 595e1a9
Fixed building types
