### 2024-08-28 6947f53
Fixed building types
