### 2024-08-29 da77007
Changed archetype name to full name
