### 2024-08-29 4c6f7e1
Changed archetype name to full name
