### 2024-08-29 16ee1d1
Fixed building assignement
