### 2024-08-29 1ba3699
Fixed building assignement
