### 2024-09-10 a0254bd
Fixed pandas export of data
