### 2024-09-10 aae9610
Fixed pandas export of data
