+++ examples/e2_generate_custom_district.py
@@ -47,7 +47,7 @@ def example_generate_custom_district():
     # Generate district with energy profiles
     # We now use the function generateDistrictComplete() to generate a complete district all at once.
     data = Datahandler()
-    data.generateDistrictComplete(scenario_name='example', calcUserProfiles=True, saveUserProfiles=False, plz='34117')
+    data.generateDistrictComplete(scenario_name='example', calcUserProfiles=True, saveUserProfiles=True, plz='52064')
 
     # Let's look into our custom district:
 
@@ -71,7 +71,7 @@ def example_generate_custom_district():
     max_electricity_power = max(data.district[0]['user'].elec)
     print("The maximum electricity power of building 0 is " + str(round(max_electricity_power)) + "W.\n")
 
-    data.designDevicesComplete(saveGenerationProfiles=True)
+    data.designDevicesComplete("central_devices_test", saveGenerationProfiles=True)
 
     return data
 
