### 2024-09-11 b9bbe3f
Vectorized heating profile
