### 2024-09-11 63aa5d0
Vectorized heating profile
