### 2024-09-11 10487ab
Fixed Bug in light demand calculation
