### 2024-09-11 ad526c5
Fixed building assignement
