### 2024-09-11 7e3a9ba
Improved documentation for time dependen values.
