+++ data/info.txt
@@ -1,6 +1,42 @@
 ## Information
 
-To enable the simulation of Mix-Used Districts a variety of data needs to be gathered, from archetypes for different building types, to assumptions about usage, e.g. 
+To enable the simulation of Mix-Used Districts a variety of data needs to be gathered, from archetypes for different building types, to assumptions about usage, e.g.. 
+
+## Data Structure
+
+The data folder contains essential information for simulating mixed-use districts. Here's an overview of the key components:
+
+1. building_types.csv:
+   This file serves as a central mapping for various building type classifications. It includes:
+   - districtgenerator: Internal classification used by the tool
+   - IWU: Classification according to the Institut Wohnen und Umwelt for non-residential buildings
+   - SIA: Swiss standard classification 
+   - CEA: City Energy Analyst classification
+   - TEK: German energy efficiency standard classification
+   - 18599_lightning: Classification for multi-zone average usage profiles
+   - 18599 classification of archetypes 
 
 
-This 
+2. occupancy_schedules:
+   Contains CSV files with occupancy schedules for different building types, based on the SIA classification.
+   Data obtained from City Energy Analyst: https://github.com/architecture-building-systems/CityEnergyAnalyst 
+
+3. TEKs:
+   Includes TEK_districtgenerator.csv, which provides energy consumption benchmarks for different building types.
+   Data obtained from DIBS: https://github.com/IWUGERMANY/DibsData/blob/main/src/dibs_data/data/auxiliary/TEKs/TEK_NWG_Vergleichswerte_zuweisung.csv 
+
+4. multi_zone_average:
+   Contains Non-domestic-multi-zone-average-usage-profiles-for-Germany.csv, providing average usage profiles for non-residential buildings.
+   Data obtained from IWU:  https://github.com/IWUGERMANY/Non-domestic-multi-zone-average-usage-profiles-for-Germany 
+
+5. weather:
+   Stores weather data files (e.g., EPW files) used for simulations.
+
+6. scenarios:
+   Contains CSV files describing different district scenarios, used as input for simulations.
+   Data is generated by TECDEM:  https://github.com/TUB-DVG/TECDEM  
+
+This structure allows for flexible and comprehensive modeling of various building types and their characteristics, enabling accurate simulation of mixed-use districts.
+
+
+ 
