### 2024-09-11 d4d4b13
Fixed errors in simulation
