### 2024-09-11 a313c06
Fixed errors in simulation
