+++ data/info.txt
@@ -1,6 +1,7 @@
 ## Information
 
-To enable the simulation of Mix-Used Districts a variety of data needs to be gathered, from archetypes for different building types, to assumptions about usage, e.g. 
+To enable the simulation of Mix-Used Districts a variety of data needs to be gathered, from archetypes for different building types, 
+to assumptions about usage, e.g. 
 
 
-This 
+This information is structed in the repository.
