### 2024-09-12 f468c05
Seperated cooling demand from heating demand in results
