### 2024-09-12 ec5ae51
Seperated cooling demand from heating demand in results
