### 2024-09-13 #6980 @kyleniemeyer
@openjournals/dev any ideas what is causing this error in the `invite as editor` command?
