### 2024-09-13 #6980 @AdamRJensen
@kyleniemeyer There's hardly any documentation for this package (there's a readme file and some example .py files) which makes me question whether or not this should pass scope review.

There are also no continuous integration tests and there has only been a single release (no track record of development).
