### 2024-09-13 #6980 @AdamRJensen
> @openjournals/dev any ideas what is causing this error in the `invite as editor` command?

It's probably because of my case sensitive github username.
