### 2024-09-15 #6980 @JoelSchoelzel
Dear all,

I apologise for the problems and the late reply. Unfortunately, I did not receive an email notification about this issue in July. If there is still the possibility, I will solve the described problems in the coming days! The DistrictGenerator was created in the first version as a finished tool from a research project and then published as an open source tool on github, which is why there is no track record of development.
What did you mean by "the authors have made efforts to make the paper render"?
We are still very interested in the submission and your review, as we hope that this will significantly improve the quality of our tool. We therefore hope for your patience.

Kind regards
