### 2024-09-17 79b0a9d
Deletion duplicated code
