### 2024-09-17 741e169
fixed f strings for server
