### 2024-09-17 0096915
fixed f strings for server
