### 2024-09-18 1c7a670
Modification of the Richardson tool to make it suitable for non-residential buildings
