+++ classes/lighting.py
@@ -1,0 +1,230 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+Created on Thu May 21 21:24:29 2015
+
+@author: Thomas
+"""
+from __future__ import division
+
+import random
+import math
+import csv
+
+
+# The Excel Sheet has a fairly complicated configuration file
+# (which I suppose most people have ignored so far)
+# This class provides the standard inputs.
+# If required, other values can be entered.
+class LightingModelConfiguration():
+
+    def __init__(self,
+                 external_irradiance_threshold=[60, 10],
+                 calibration_scalar=0.00815368639667705,
+                 effective_occupancy=[0,
+                                      1,
+                                      1.52814569536424,
+                                      1.69370860927152,
+                                      1.98344370860927,
+                                      2.09437086092715],
+                 lighting_event_lower_value=[1, 2, 3, 5, 9, 17, 28, 50, 92],
+                 lighting_event_upper_value=[1, 2, 4, 8, 16, 27, 49, 91, 259]):
+        """
+        Constructor of lighting class object instance
+
+        Parameters
+        ----------
+        external_irradiance_threshold : list, optional
+            List holding house external global irradiance threshold values
+            in W/m2 [index 0: mean value; index 1: standard deviation value]
+            (default: [60, 10])
+        calibration_scalar : float, optional
+            Original tool description: This calibration scaler is used to
+            calibrate the model to so that it provides a particular average
+            output over a large number of runs.
+            (default: 0.00815368639667705)
+        effective_occupancy : list (of floats), optional
+            List holding effective occupancy based on number of active o
+            ccupants (default: [0,
+                                1,
+                                1.52814569536424,
+                                1.69370860927152,
+                                1.98344370860927,
+                                2.09437086092715])
+        lighting_event_lower_value : list (of ints), optional
+            Lower number of lighting event duration in minutes
+            (default: [1, 2, 3, 5, 9, 17, 28, 50, 92])
+        lighting_event_upper_value : list (of ints), optional
+            Upper number of lighting event duration in minutes
+            (default: [1, 2, 4, 8, 16, 27, 49, 91, 259])
+        """
+
+        # House external global irradiance threshold
+        self.ext_irr_threshold_mean = external_irradiance_threshold[0]
+        self.ext_irr_threshold_std_dev = external_irradiance_threshold[1]
+
+        # This calibration scaler is used to calibrate the model to so that
+        # it provides
+        # a particular average output over a large number of runs.
+        self.calib_scalar = calibration_scalar
+
+        # Effective occupancy represents the sharing of light use.
+        self.eff_occupancy = effective_occupancy
+
+        # This model defines how long a bulb will stay on for, if a
+        # switch-on event occurs.
+        self.light_event_lower_value = lighting_event_lower_value
+        self.light_event_upper_value = lighting_event_upper_value
+
+    def relative_bulb_use_weighting(self):
+        """
+        This represents the concept that some bulbs are used more
+        frequently than others in a house.
+
+        The return value is [-ln(random_variable)]
+
+        Returns
+        -------
+        -math.log(random.random())
+        """
+        return -math.log(random.random())
+
+def run_lighting_simulation(vOccupancyArray, vBulbArray, vIrradianceArray,
+                            light_mod_config):
+    """
+
+    Parameters
+    ----------
+    vOccupancyArray : array-like
+        Occupancy for one day (10 minute resolution)
+    vBulbArray : array-like
+        Bulb data
+    vIrradianceArray : array-like
+        Irradiance data
+    light_mod_config : object
+        LightingModelConfiguration object instance
+
+    Returns
+    -------
+    result : float
+        Power in Watt
+    """
[…]
