### 2024-09-18 cf9a76e
Adding profiles for dhw, electricity, occupancy and internal gains for non-residential buildings
