+++ classes/envelope.py
@@ -34,7 +34,7 @@ class Envelope:
         SFH: single family house; TH: terraced house; MFH: multifamily house; AP: apartment block.
     """
 
-    def __init__(self, prj, building_params, construction_type, file_path):
+    def __init__(self, prj, building_params, construction_type, window_layout, file_path):
         """
         Constructor of Envelope class.
 
@@ -59,6 +59,7 @@ class Envelope:
         self.id = building_params["id"]
         self.construction_year = building_params["year"]
         self.construction_type = construction_type
+        self.window_layout = window_layout
         self.retrofit = building_params["retrofit"]
         self.usage_short = building_params["building"]
         self.file_path = file_path
@@ -244,134 +245,237 @@ class Envelope:
         comp = "wall"
         # WALLS: Materials and U-value
         for name, elem in element_bind.items():
-            if "OuterWall" in name:
-                if elem["building_age_group"][0] <= self.construction_year <= \
-                        elem["building_age_group"][1] and \
-                        elem["construction_type"] == self.construction_type \
-                        + "_1_" + self.usage_short:
+            if self.construction_type in {"heavy", "light"}:   # so it's not a residential building
+
+                if "OuterWall" in name:
+                    dummy = min(2015,
+                                self.construction_year)  # data available until 2015
+                    if elem["building_age_group"][0] <= dummy <= \
+                        elem["building_age_group"][1] and elem["construction_type"] == self.construction_type:
+
+                        for lay in elem["layer"].items():
+                            self.d["opaque"][comp] = np.append(self.d["opaque"][comp], lay[1]["thickness"])
+                            material_prop = self.loadMaterialID(lay[1]["material"]["material_id"], material_bind)
+                            self.rho["opaque"][comp] = np.append(self.rho["opaque"][comp], material_prop[1])
+                            self.Lambda["opaque"][comp] = np.append(self.Lambda["opaque"][comp], material_prop[2])
+                            self.cp["opaque"][comp] = np.append(self.cp["opaque"][comp], material_prop[3] * 1000)
+            else:
+                if "OuterWall" in name and elem["building_age_group"][0] <= self.construction_year <= \
+                        elem["building_age_group"][1] and elem["construction_type"] == self.construction_type + "_1_" + self.usage_short:
+
                     for lay in elem["layer"].items():
-                        self.d["opaque"][comp] = np.append(self.d["opaque"][comp],
-                                                           lay[1]["thickness"])
-                        material_prop = self.loadMaterialID(
-                            lay[1]["material"]["material_id"], material_bind)
-                        self.rho["opaque"][comp] = np.append(self.rho["opaque"][comp],
-                                                             material_prop[1])
-                        self.Lambda["opaque"][comp] = np.append(self.Lambda["opaque"][comp],
-                                                                material_prop[2])
-                        self.cp["opaque"][comp] = np.append(self.cp["opaque"][comp],
-                                                            material_prop[3] * 1000)
+                        self.d["opaque"][comp] = np.append(self.d["opaque"][comp], lay[1]["thickness"])
+                        material_prop = self.loadMaterialID(lay[1]["material"]["material_id"], material_bind)
+                        self.rho["opaque"][comp] = np.append(self.rho["opaque"][comp], material_prop[1])
+                        self.Lambda["opaque"][comp] = np.append(self.Lambda["opaque"][comp], material_prop[2])
+                        self.cp["opaque"][comp] = np.append(self.cp["opaque"][comp], material_prop[3] * 1000)
 
         comp = "roof"
         # ROOF: Materials and U-value
         for name, elem in element_bind.items():
-            if "Rooftop" in name:
-                if elem["building_age_group"][0] <= self.construction_year <= \
[…]

[…]
