### 2024-09-18 0f504ac
Modification of the OpenDHW tool to make it suitable for non-residential buildings
