+++ functions/OpenDHW.py
@@ -107,7 +107,7 @@ def import_from_dhwcalc(s_step, daylight_saving, categories,
     return timeseries_df
 
 
-def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy,  weekend_weekday_factor=1.2, initial_day=0):
+def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy, building_type,  weekend_weekday_factor, initial_day=0):
     """
     Generates a DHW profile. The generation is split up in different
     functions and generally follows the methodology described in the DHWcalc
@@ -131,17 +131,20 @@ def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy
     mean_drawoff_vol_per_day *= occupancy
 
     # --- holds statistic info about the drawoffs
+
     cats_df = get_data_drawoff_categories(
         s_step=s_step,
         categories=categories,
-        mean_drawoff_vol_per_day=mean_drawoff_vol_per_day
+        mean_drawoff_vol_per_day=mean_drawoff_vol_per_day,
+        building_type=building_type
     )
 
     # --- deterministic function
     timeseries_df = generate_yearly_probability_profile(
         s_step=s_step,
-        weekend_weekday_factor=1.2,
-        initial_day=0
+        weekend_weekday_factor=weekend_weekday_factor,
+        initial_day=0,
+        building_type=building_type
     )
 
     # --- empty drawoffs list, will be filled afterwards
@@ -165,42 +168,58 @@ def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy
     return timeseries_df
 
 
-def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day):
+def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day, building_type):
     """
     Get some data for each drawoff category. If only one category is chosen,
     a simplified datafarme is returned.
 
     :param s_step:                      int:    seconds in a timestep. f.e 900
-    :param categories:                  int:    1 or 4, see DHWcalc
+    :param categories:                  int:    1 or 4, 1: short laod (washing hands, etc.), 2: medium load (dish-washer, etc.), 3: bath, 4:shower (see DHWcalc)
     :param mean_drawoff_vol_per_day:    int:    volume per day used in house
     :return: cats_df:                   df:     Categores Data
     """
-    if categories == 4:
-        cats_data_60 = {'mean_flow_rate_per_drawoff_LperH': [60, 360, 840, 480],
-                        'drawoff_duration_min': [1, 1, 10, 5],
-                        'portion': [0.14, 0.36, 0.1, 0.4],
-                        'stddev_flow_rate_per_drawoff_LperH': [120, 120, 12,
-                                                               24],
-                        'min_flow_rate_per_drawoff_LperH': [1, 1, 1, 1]
-                        }
+    if building_type in {"SFH", "TH", "MFH", "AB"}:
 
-        cats_df = pd.DataFrame(data=cats_data_60)
+        if categories == 4:
+            cats_data_60 = {'mean_flow_rate_per_drawoff_LperH': [60, 360, 840, 480],
+                            'drawoff_duration_min': [1, 1, 10, 5],
+                            'portion': [0.14, 0.36, 0.1, 0.4],
+                            'stddev_flow_rate_per_drawoff_LperH': [120, 120, 12,24],    #woher diese Werte? es ist in der paper und tool mit 120 immer gegeben
+                            'min_flow_rate_per_drawoff_LperH': [1, 1, 1, 1]
+                            }
+
[…]
