### 2024-09-18 67cc635
Changed paths to work with docker
