### 2024-09-18 e725868
Changed paths to work with docker
