### 2024-09-19 #6980 @AdamRJensen
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/9922029643?check_suite_focus=true).

@JoelSchoelzel the paper isn't rendering, most likely because you haven't formatted the paper correctly.

Also, is it correct that there's no separate documentation site for the package?
