### 2024-09-19 #6980 @kyleniemeyer
@JoelSchoelzel I think the issue might be missing the `---` after the YAML block at the top of the paper. If you add this below the `bibliography` line, that might fix the problem.

I also noticed that you have some authors with empty `orcid` fields. I'd suggest just removing those, if those coauthors do not have ORCIDs.
