### 2024-09-19 fcae9a7
Fixed Bug in Schedule calculation
