+++ districtgenerator/non_residential_users.py
@@ -114,6 +114,8 @@ class NonResidentialUsers():
         #self.nb_flats = None
         self.annual_appliance_demand = None
         self.annual_lightning_demand = None
+        self.appliance_demand = None
+        self.lightning_demand = None
         self.nb_occ = []
         self.occ = None
         self.dhw = None
@@ -142,10 +144,8 @@ class NonResidentialUsers():
         self.generate_number_occupants()
         self.generate_annual_el_consumption_equipment()
         self.generate_annual_el_consumption_lightning()
+        self.generate_el_demand()
         # To-Do: Validate Sum of electricity
-        self.elec = self.annual_appliance_demand + self.annual_lightning_demand
-        # self.generate_lighting_index()
-        # self.create_el_wrapper()
 
     
     def load_json_data(self, json_path: str) -> Dict[str, Any]:
@@ -189,12 +189,15 @@ class NonResidentialUsers():
 
 
     def generate_schedules(self) -> None:
-        # get schedules occupancy
+        """
+        Generate occupancy, appliance, and lighting schedules for the building.
+        """
+     
         df_schedules, schedule = schedules.getSchedule(self.usage)
         if df_schedules is not None:
-            self.occupancy_schedule = schedules.adjust_schedule(inital_day= 0, schedule=df_schedules[["DAY", "HOUR", "OCCUPANCY"]], nb_days=self.nb_of_days)
-            self.appliance_schedule =  schedules.adjust_schedule(inital_day= 0, schedule=df_schedules[["DAY", "HOUR", "APPLIANCES"]], nb_days=self.nb_of_days)
-            self.lighntning_schedule = schedules.adjust_schedule(inital_day= 0, schedule=df_schedules[["DAY", "HOUR", "LIGHTING"]], nb_days=self.nb_of_days)
+            self.occupancy_schedule = schedules.adjust_schedule(initial_day= 0, schedule=df_schedules[["DAY", "HOUR", "OCCUPANCY"]], nb_days=self.nb_of_days)
+            self.appliance_schedule =  schedules.adjust_schedule(initial_day= 0, schedule=df_schedules[["DAY", "HOUR", "APPLIANCES"]], nb_days=self.nb_of_days)
+            self.lighntning_schedule = schedules.adjust_schedule(initial_day= 0, schedule=df_schedules[["DAY", "HOUR", "LIGHTING"]], nb_days=self.nb_of_days)
         else:
             print(f"No schedules found for {self.usage} and {schedule}")
 
@@ -237,12 +240,13 @@ class NonResidentialUsers():
         if self.usage in self.electricity_data:
             electricity_values = self.electricity_data[self.usage]
             try: 
-                # To-Do: Check if data is atually in w 
                 # To-Do: Check if data is series
-                annual_el_demand_temp = electricity_values[equipment] * self.area 
+                # Units: electricity_values[equipment] in kWh/m2*a , self.area in m2
+                annual_el_demand_temp = electricity_values[equipment] * self.area
+                # assumption: standard deviation 20% of mean value 
                 self.annual_appliance_demand = rd.gauss(annual_el_demand_temp,
-                                                        annual_el_demand_temp * 0.10)  # assumption: standard deviation 20% of mean value
-
+                                                        annual_el_demand_temp * 0.10)
+                self.appliance_demand = self.annual_appliance_demand * self.appliance_schedule["APPLIANCES"]
             except KeyError:
                 print(f"No data about annual electrical consumption available for building type: {self.usage}")
             # To Do 
@@ -279,8 +283,10 @@ class NonResidentialUsers():
 
         # To-Do: 
[…]
