### 2024-09-19 af1f9df
Fixed Bug in Schedule calculation
