+++ districtgenerator/datahandler.py
@@ -547,6 +547,8 @@ class Datahandler():
         set = []
         if not os.path.exists(self.resultPath):
             os.makedirs(self.resultPath)
+        if os.path.exists(self.resultPath):
+            print("The folder already exists. Will overwrite existing files.")
         for building in self.district:
             # %% create unique building name
             # needed for loading and storing data with unique name
