+++ functions/schedule_reader.py
@@ -56,14 +56,27 @@ def getSchedule(building_type):
         print(f"File not found: {data_path}")
         return None, None
 
-def adjust_schedule(inital_day, schedule, nb_days):
+def adjust_schedule(initial_day, schedule, nb_days):
     """
-    Function returns the schedule, 
-    adjusted to the initial_day and the last 
+    Adjusts and expands a schedule based on an initial day and a specified number of days.
+
+    This function performs the following operations:
+    1. Rotates the days of the week to start from the specified initial day.
+    2. Creates a copy of the input schedule to avoid modifying the original.
+    3. Converts the 'DAY' column to a categorical type with the rotated order.
+    4. Sorts the schedule by the new day order and hour.
+    5. Expands the schedule to cover the specified number of days.
+
+    Parameters:
+    initial_day (int): The day of the week to start the schedule (0-6, where 0 is Monday).
+    schedule (pd.DataFrame): The original schedule DataFrame with 'DAY' and 'HOUR' columns.
+    nb_days (int): The total number of days the schedule should cover.
+
+    Returns:
+    pd.DataFrame: An adjusted and expanded copy of the input schedule.
     """
     # Create a custom sorter index
-    sorter = rotate_list(initial_day=inital_day)
-    sorter_index = {day: index for index, day in enumerate(sorter)}
+    sorter = rotate_list(initial_day=initial_day)
 
     # Create a copy of the schedule DataFrame
     schedule_copy = schedule.copy()
@@ -74,15 +87,13 @@ def adjust_schedule(inital_day, schedule, nb_days):
     schedule_copy = expand_dataframe(schedule_copy, total_days=nb_days)
     return schedule_copy
 
-
 def rotate_list(initial_day):
     # Number of days in the schedule
-    lst = [0 , 1, 2, 3, 4, 5, 6]
+    lst = [0, 1, 2, 3, 4, 5, 6]
     # Find the index of the start day
     start_index = lst.index(initial_day)
     # Rotate the list from that index
     return lst[start_index:] + lst[:start_index]
-
 
 def expand_dataframe(df, total_days):
     unique_days = df['DAY'].unique()
