### 2024-09-19 a008edc
Update light_demand.py
