### 2024-09-20 #6980 @JoelSchoelzel
@AdamRJensen, @kyleniemeyer
we have discussed your comments with our team. First of all, you are absolutely right that the documentation is not sufficient. There is currently no separate documentation page. Our idea is to build a detailed documentation page and improve the readme. Furthermore, we would like to revise the examples so that they can be used as end-to-end functional tests for the main functions of the DistrictGenerator.

As the DistrictGenerator is an important product of the German accompanying scientific research in the field of energy supply of neighbourhoods (https://www.energiewendebauen.de/en/research/accompanying_scientific_research) and has been published as an open webtool for science and practice actors (https://districtgenerator.eonerc.rwth-aachen.de/de/), we would like to ask for the possibility to improve and complete the mentioned points.  With view on our project work, we will probably be able to do this by 16 October. We hope that this will still be possible during the pre-review phase.
