### 2024-09-20 #6980 @AdamRJensen
@JoelSchoelzel This should be fine. Note, a separate documentation page using for example Sphinx and readthedocs is often preferred. That way you can also render the examples with some output plots.

I also don't see any testing, which is necessary to make sure that functions don't accidentally change over time.
