### 2024-09-20 #6980 @kyleniemeyer
@JoelSchoelzel that timing should be fine—we will pause your submission until you've been able to make those changes. As @AdamRJensen said, testing is also required; you can see the full review criteria for JOSS here: https://joss.readthedocs.io/en/latest/review_criteria.html
