### 2024-09-23 74861b4
insert external certificate function
