+++ classes/KPIs.py
@@ -4,6 +4,14 @@ import sys
 import numpy as np
 import os
 import json
+import reportlab
+from reportlab.pdfgen import canvas
+from reportlab.lib import colors
+from reportlab.graphics.shapes import *
+from reportlab.graphics.charts.piecharts import Pie
+from datetime import datetime
+from reportlab.lib.styles import ParagraphStyle
+from reportlab.platypus import Paragraph
 
 
 class KPIs:
@@ -367,3 +375,495 @@ class KPIs:
         self.calculateOperationCosts(data)
         self.calculateCO2emissions(data)
         self.calculateAutonomy()
+
+    def create_kpi_pdf(self,result_path):
+        """
+        Generate a PDF file with a list of KPIs.
+
+        Parameters:
+        - filename: The name of the PDF file to create.
+        - title: The title of the document.
+        - kpis: A list of strings, where each string is a KPI to be written in the document.
+        """
+
+    def create_certificate(kennwerte, opt_ergebnisse, struktur, gebaeudeliste):
+        """
+        Creates an energy certificate in PDF format for a neighborhood created via the EBC Quartiersgenerator.
+
+        Args:
+            kennwerte: a dictionary with the following keys (in order, formatted as strings), and all values formatted as
+            strings with the corresponding units (unless otherwise specified):
+                Primärenergiebedarf: primary energy demand of the district
+                Endenergiebedarf: end energy demand of the district
+                Norm-Heizlast insgesamt: the overall heat demand of the district
+                Solltemperatur: set-point temperature of the buildings in the district
+                Bedarfe: a TUPLE containing three values (float/int) for demands of electricity, heat, and water heating (in that order)
+                Max. Leistungen: a TUPLE containing three values (float/int) for the maximum power of each energy type, in the order above
+
+            opt_ergebnisse: a dictionary with the following keys (in order, formatted as strings), and all values formatted
+            as strings with the corresponding units:
+                CO2-äqui. Emissionen: CO2-equivalent emissions of the district
+                Energiekosten: energy cost for the district
+                Spitzenlast (el.) gesamt: peak load for the district
+                Max. Einspeiseleistung gesamt: maximum feed-in power of the district
+                Supply-Cover-Faktor: supply cover factor
+                Demand-Cover-Faktor: demand cover factor
+
+            struktur: a dictionary with the following keys (in order, formatted as strings) and and all values formatted as
+            strings with the corresponding units (unless otherwise specified):
+                EFH: a DICTIONARY containing the following keys and values pertaining to single-family homes in the district
+                (keys formatted as strings, values formatted as strings including the relevant units):
+                    Anzahl: the number of buildings of this type in the district
+                    Gesamtfläche: the total floor space of these buildings (without unit!)
+                    vor 1968: the total floor space of the buildings of this type built before 1968, in m^2 (without unit in string!)
+                    1968-1979: the total floor space of the buildings of this type built between 1968 and 1979, in m^2 (without unit in string!)
+                    1979-1983: the total floor space of the buildings of this type built between 1979 and 1983, in m^2 (without unit in string!)
+                    1984-1994: the total floor space of the buildings of this type built between 1984 and 1994, in m^2 (without unit in string!)
[…]
