### 2024-09-24 fcfc2d9
Fixed computation of daily demand
