+++ functions/dhw_stochastical.py
@@ -73,6 +73,7 @@ def load_profiles(filename):
     return profiles
 
 
+
 def compute_daily_demand(probability_profiles, average_profile, occupancy,
                          current_day, temperature_difference=35):
     """
