### 2024-09-24 dba7942
Fixed computation of daily demand
