### 2024-09-27 6858296
add preprocessing methods to calc further KPIs
