+++ classes/KPIs.py
@@ -356,6 +356,64 @@ class KPIs:
         #    self.EnergyAutonomy_year += EnergyAutonomy[c] * (self.inputData["clusterWeights"][self.inputData["clusters"][c]]
         #                                                  / sum_ClusterWeights)
 
+    def calcareas(self, data):
+        """
+        Returns
+        -------
+        None.
+        """
+        total_net_leased_area = 0
+        total_number_flats = 0
+        total_number_occ = 0
+
+        for building in data.district:
+            total_net_leased_area += building["buildingFeatures"]["area"]
+            total_number_flats += building["user"].nb_flats
+            total_number_occ += building["user"].nb_occ
+
+        self.totalarea = total_net_leased_area
+        self.totalnumberflats = total_number_flats
+        self.totalnumberocc = total_number_occ
+
+    def calctotalenergies(self, data):
+        """
+        Returns
+        -------
+        None.
+        """
+        total_heatload = 0
+        total_el_demand = 0
+        total_heat_demand = 0
+        total_dhw_demand = 0
+        total_cooling_demand = 0
+        total_el_profile = data.district["0"]["user"].elec * 0
+        total_heat_profile = data.district["0"]["user"].elec * 0
+        total_dhw_profile = data.district["0"]["user"].elec * 0
+        total_cooling_profile = data.district["0"]["user"].elec * 0
+
+
+        for building in data.district:
+            total_heatload += building["heatload"]
+            total_el_demand += building["user"].annual_el_demand
+            total_heat_demand += building["user"].annual_heat_demand
+            total_dhw_demand += building["user"].annual_dhw_demand
+            total_cooling_demand += building["user"].annual_cooling_demand
+            total_el_profile += building["user"].elec
+            total_heat_profile += building["user"].heat
+            total_dhw_profile += building["user"].dhw
+            total_cooling_profile += building["user"].cooling
+
+
+        self.totalheatload = total_heatload
+        self.total_el_demand = total_el_demand
+        self.total_heat_demand = total_heat_demand
+        self.total_dhw_demand = total_dhw_demand
+        self.total_cooling_demand = total_cooling_demand
+        self.total_el_peak = max(total_el_profile)
+        self.total_heat_peak = max(total_heat_profile)
+        self.total_dhw_peak = max(total_dhw_profile)
+        self.total_cooling_peak = max(total_cooling_profile)
+
     def calculateAllKPIs(self, data):
         """
         Calculate all KPIs.
@@ -375,8 +433,11 @@ class KPIs:
         self.calculateOperationCosts(data)
         self.calculateCO2emissions(data)
         self.calculateAutonomy()
+        self.calcareas(data)
+        self.calctotalenergy(data)
+        self.calctotalenergies(data)
 
-    def create_kpi_pdf(self,result_path):
+    def create_kpi_pdf(self, data, result_path):
         """
         Generate a PDF file with a list of KPIs.
 
@@ -386,7 +447,135 @@ class KPIs:
         - kpis: A list of strings, where each string is a KPI to be written in the document.
         """
 
-    def create_certificate(kennwerte, opt_ergebnisse, struktur, gebaeudeliste):
+        # preprocessing buildinglist
+        template_dict = {"Anzahl": 0,
+               "Gesamtfläche": 0,
+               "vor 1968": 0,
+               "1968-1978": 0,
+               "1979-1983": 0,
+               "1984-1994": 0,
+               "1995-2001": 0,
+               "2002-2009": 0,
+               "2010-2015": 0,
+               "ab 2016": 0,
+               }
+        SFH = dict(template_dict)
[…]
