+++ docs/source/index.rst
@@ -1,0 +1,24 @@
+.. districtgenerator documentation master file, created by
+   sphinx-quickstart on Thu Jul 11 08:20:35 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+About districtgenerator
+-------------------
+
+.. mdinclude::  ../../README.md
+
+.. toctree::
+   :maxdepth: 2
+
+   code/modules
+   Contribution
+   version_his
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
