+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+sphinx==6.2.1
+m2r2
+sphinx-material
+autodoc_pydantic
+sphinx-rtd-theme
