+++ classes/datahandler.py
@@ -1,389 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import json
-import pickle
-import os
-import numpy as np
-import pandas as pd
-from itertools import count
-from teaser.project import Project
-from classes.envelope import Envelope
-from classes.solar import Sun
-from classes.users import Users
-from classes.plots import DemandPlots
-
-class Datahandler():
-    """
-    Abstract class for data handling
-    Collects data from input files, TEASER, User and Enevelope.
-
-    Attributes
-    ----------
-    site:
-        dict for site data, e.g. weather
-    time:
-        dict for time settings
-    district:
-        list of all buildings within district
-    scenario_name:
-        name of scenario file
-    scenario:
-        scenario data
-    counter:
-        dict for counting number of equal building types
-    srcPath:
-        source path
-    filePath:
-        file path
-
-    """
-
-    def __init__(self):
-        self.site = {}
-        self.time = {}
-        self.district = []
-        self.scenario_name = None
-        self.scenario = None
-        self.counter = {}
-        self.srcPath = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-        self.filePath = os.path.join(self.srcPath, 'data')
-        self.resultPath = os.path.join(self.srcPath, 'results', 'demands')
-
-
-    def generateEnvironment(self):
-        """
-        Load physical district environment - site and weather
-
-        Parameters
-        ----------
-
-        Returns
-        -------
-
-        """
-
-        # %% load information about of the site under consideration
-        # important for weather conditions
-        with open(os.path.join(self.filePath, 'site_data.json')) as json_file:
-            jsonData = json.load(json_file)
-            for subData in jsonData:
-                self.site[subData["name"]] = subData["value"]
-
-        # %% load weather data for site
-        # extract irradiation and ambient temperature
-        first_row = {}
-        if self.site["TRYYear"]=="TRY2015":
-            first_row = 35
-        elif self.site["TRYYear"]=="TRY2045":
-            first_row = 37
-
-        weatherData = np.loadtxt(os.path.join(self.filePath, 'weather')
-                                 + "/"
-                                 + self.site["TRYYear"] + "_Zone"
-                                 + str(self.site["climateZone"]) + "_"
-                                 + self.site["TRYType"] + ".txt",
-                                 skiprows=first_row - 1)
-
-
-        # weather data starts with 1st january at 1:00 am. Add data point for 0:00 am to be able to perform interpolation.
-        weatherData_temp = weatherData[-1:, :]
-        weatherData = np.append(weatherData_temp, weatherData, axis=0)
-
-        # get weather data of interest
-        [temp_sunDirect,  temp_sunDiff, temp_temp] = [weatherData[:, 12], weatherData[:, 13], weatherData[:, 5]]
-
-        # %% load time information and requirements
-        # needed for data conversion into the right time format
-        with open(os.path.join(self.filePath, 'time_data.json')) as json_file:
-            jsonData = json.load(json_file)
-            for subData in jsonData :
-                self.time[subData["name"]] = subData["value"]
-        self.time["timeSteps"] = int(self.time["dataLength"] / self.time["timeResolution"])
-
-        # interpolate input data to achieve required data resolution
-        # transformation from values for points in time to values for time intervals
-        self.site["SunDirect"] = np.interp(np.arange(0, self.time["dataLength"]+1, self.time["timeResolution"]),
-                                           np.arange(0, self.time["dataLength"]+1, self.time["dataResolution"]),
-                                           temp_sunDirect)[0:-1]
-        self.site["SunDiffuse"] = np.interp(np.arange(0, self.time["dataLength"]+1, self.time["timeResolution"]),
[…]
