+++ setup.py
@@ -6,8 +6,8 @@ setuptools.setup(name='districtgenerator',
                  description='Energy profile generation and '
                              'optimization of districts',
                  url='https://github.com/RWTH-EBC/districtgenerator',
-                 author='Sarah Henn',
-                 author_email='shenn@eonerc.rwth-aachen.de',
+                 author='Joel Schölzel',
+                 author_email='joel.schoelzel@eonerc.rwth-aachen.de',
                  license='MIT License',
                  packages=setuptools.find_packages(),
                  install_requires=['numpy', 'pandas', 'matplotlib', 'scipy',
