+++ docs/about/history.rst
@@ -1,5 +1,0 @@
-===============
-Version History
-===============
-
-.. mdinclude::  ../CHANGELOG.md

[…]
