+++ classes/profils.py
@@ -144,7 +144,7 @@ class Profiles:
 
         return self.occ_profile
 
-    def generate_profiles_non_residential(self):
+    def generate_profiles_non_residential(self,holidays):
         """
          Generate stochastic peaople profiles, devices profiles and month profiles
          for every zone of the non-residential building
@@ -163,48 +163,78 @@ class Profiles:
             sia_week_profile_devices_zone = []
 
             for i in range(7):
-                is_weekend = (i + self.initial_day) % 7 in (0, 6)
+                is_not_working_day = (i + self.initial_day) % 7 in (0, 6)
 
-                sia_day_profile_people_zone = [0] * 24 if is_weekend else data['profile_people']
-                sia_day_profile_devices_zone = [min(data['profile_devices'])] * 24 if is_weekend else data['profile_devices']
+                sia_day_profile_people_zone = [0] * 24 if is_not_working_day else data['profile_people']
+                sia_day_profile_devices_zone = [min(data['profile_devices'])] * 24 if is_not_working_day else data['profile_devices']
                 sia_week_profile_people_zone.extend(sia_day_profile_people_zone)
                 sia_week_profile_devices_zone.extend(sia_day_profile_devices_zone)
 
-                repetitions = total_hours // len(sia_week_profile_people_zone)
-                remaining_hours = total_hours % len(sia_week_profile_people_zone)
+            repetitions = total_hours // len(sia_week_profile_people_zone)
+            remaining_hours = total_hours % len(sia_week_profile_people_zone)
 
-                # Repeat the weekly pattern to cover the entire year
-                sia_profile_people_zone = sia_week_profile_people_zone * repetitions + sia_week_profile_people_zone[:remaining_hours]
-                sia_profile_devices_zone = sia_week_profile_devices_zone * repetitions + sia_week_profile_devices_zone[:remaining_hours]
+            # Repeat the weekly pattern to cover the entire year
+            sia_profile_people_zone = sia_week_profile_people_zone * repetitions + sia_week_profile_people_zone[:remaining_hours]
+            sia_profile_devices_zone = sia_week_profile_devices_zone * repetitions + sia_week_profile_devices_zone[:remaining_hours]
 
-                # Apply random variation to the people and devices profiles
-                profile_people_zone = [min(max(np.random.normal(value, value * 0.1), 0), 1) for value in sia_profile_people_zone]
-                profile_devices_zone = [min(max(np.random.normal(value, value * 0.1), 0), 1) for value in sia_profile_devices_zone]
+            # Consider holidays
+            for day in range(self.nb_days):
+                if (day + self.initial_day) in holidays:
+                    sia_profile_people_zone[24 * day: 24 * (day + 1)] = [0] * 24
+                    sia_profile_devices_zone[24 * day: 24 * (day + 1)] = [min(data['profile_devices'])] * 24
 
-                # Change resolution
-                profile_people_zone = chres.changeResolution(profile_people_zone, 3600, self.time_resolution, "mean")
-                profile_devices_zone = chres.changeResolution(profile_devices_zone, 3600, self.time_resolution, "mean")
+            # Apply random variation to the people and devices profiles
+            profile_people_zone = [min(max(np.random.normal(value, value * 0.1), 0), 1) for value in sia_profile_people_zone]
+            profile_devices_zone = [min(max(np.random.normal(value, value * 0.1), 0), 1) for value in sia_profile_devices_zone]
 
-                # Apply random variation to the monthly profile
-                profile_month_zone = [min(max(np.random.normal(value, value * 0.1), 0), 1) for value in data['profile_month']]
+            # Apply random variation to the monthly profile
[…]
