### 2024-10-14 1db1e52
updating lighting for non-residential
