### 2024-10-14 6277650
adding new non-residential buildings
