+++ classes/users.py
@@ -129,10 +129,10 @@ class Users:
         elif self.building == "TH":
             self.nb_flats = 1
         elif self.building == "MFH":
-            if area <= 6 * 100:
-                self.nb_flats = 6
-            elif area > 6 * 100:
-                self.nb_flats = 8
+            if area <= 4 * 100:
+                self.nb_flats = 4
+            elif area > 4 * 100:
+                self.nb_flats = rd.randint((area//100)-1, (area//100)+1)
         elif self.building == "AB":
             self.nb_flats = 8
 
@@ -170,6 +170,22 @@ class Users:
                 self.nb_main_rooms = round(self.total_main_area / rd.gauss(mean_area_per_main_room,
                                                                            mean_area_per_main_room * 0.1))  # A random number of classrooms following a Gaussian distribution with a mean of mean_area_per_main_room (70 m²) and a standard deviation of 10%. The number of offices is calculated in order to later determine the number of occupants of the building depending on it
 
+            if self.building == "Grocery_store":
+                proportion_selling_space = 0
+                nb_rooms = 0
+                mean_area_per_main_room = 400
+                for number, data in self.SIA2024.items():
+                    zone_name = data.get('Zone_name_GER')
+                    if zone_name:
+                        proportion = self.building_zones[zone_name]
+                        nb_rooms += area * proportion / data['area_room']
+                self.nb_rooms = round(nb_rooms)
+
+                for zone_name in {"Lebensmittelverkauf"}:
+                    proportion_selling_space += self.building_zones[zone_name]
+                self.total_main_area = area * proportion_selling_space
+                self.nb_main_rooms = round(self.total_main_area / rd.gauss(mean_area_per_main_room,
+                                                                           mean_area_per_main_room * 0.1))
     def generate_number_occupants(self,area):
         """
         Generate number of occupants for different of building types.
@@ -243,13 +259,19 @@ class Users:
             if self.building == "OB":
                 # loop over all office rooms of current office building
                 for k in range(self.nb_main_rooms):
+                    # self.nb_occ is the number of occupants in every main room
                     self.nb_occ.append(round(rd.gauss(self.total_main_area/self.nb_main_rooms/12,area/self.nb_main_rooms/12 * 0.15)))  # 12 m² area per occupant (source: SIA); assumption: random number based on a Gaussian distribution with a standard deviation of 15%
 
             if self.building == "School":
                 # loop over all School main rooms of current School building
                 for k in range(self.nb_main_rooms):
+                    # self.nb_occ is the number of occupants in every main room
                     self.nb_occ.append(round(rd.gauss(self.total_main_area/self.nb_main_rooms/3,area/self.nb_main_rooms/3 * 0.15)))  # 3 m² area per occupant (source: SIA); assumption: random number based on a Gaussian distribution with a standard deviation of 15%
 
+            if self.building == "Grocery_store":
+                for k in range(self.nb_main_rooms):
+                    # self.nb_occ is the number of occupants in every main room
+                    self.nb_occ.append(round(rd.gauss(self.total_main_area/self.nb_main_rooms/8,area/self.nb_main_rooms/8 * 0.15)))
     def generate_annual_el_consumption_residential(self):
         """
         Generate annual electricity consumption
@@ -401,7 +423,7 @@ class Users:
 
         elif self.building in {"OB"}:
             for j in range(self.nb_rooms):
-                room_illuminance =rd.randint(300, 1000)
+                room_illuminance =rd.randint(300, 750)
 
                 room_area = area/self.nb_rooms
[…]
