+++ functions/OpenDHW.py
@@ -107,7 +107,7 @@ def import_from_dhwcalc(s_step, daylight_saving, categories,
     return timeseries_df
 
 
-def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy, building_type,  weekend_weekday_factor, initial_day=0):
+def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy, holidays, building_type,  weekend_weekday_factor, initial_day=0):
     """
     Generates a DHW profile. The generation is split up in different
     functions and generally follows the methodology described in the DHWcalc
@@ -143,6 +143,7 @@ def generate_dhw_profile(s_step, categories, mean_drawoff_vol_per_day, occupancy
     timeseries_df = generate_yearly_probability_profile(
         s_step=s_step,
         weekend_weekday_factor=weekend_weekday_factor,
+        holidays = holidays,
         initial_day=0,
         building_type=building_type
     )
@@ -184,7 +185,7 @@ def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day, bu
             cats_data_60 = {'mean_flow_rate_per_drawoff_LperH': [60, 360, 840, 480],
                             'drawoff_duration_min': [1, 1, 10, 5],
                             'portion': [0.14, 0.36, 0.1, 0.4],
-                            'stddev_flow_rate_per_drawoff_LperH': [120, 120, 12,24],    #woher diese Werte? es ist in der paper und tool mit 120 immer gegeben
+                            'stddev_flow_rate_per_drawoff_LperH': [120, 120, 12,24],
                             'min_flow_rate_per_drawoff_LperH': [1, 1, 1, 1]
                             }
 
@@ -208,10 +209,10 @@ def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day, bu
 
 
     else:
-        cats_data = {'mean_flow_rate_per_drawoff_LperH': [60, 360],
+        cats_data = {'mean_flow_rate_per_drawoff_LperH': [100, 360],
                      'drawoff_duration_min': [1, 1],
                      'portion': [0.28, 0.72],
-                     'stddev_flow_rate_per_drawoff_LperH': [120, 120],
+                     'stddev_flow_rate_per_drawoff_LperH': [200, 240],
                      'min_flow_rate_per_drawoff_LperH': [1, 1]
                      }
 
@@ -243,8 +244,7 @@ def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day, bu
         cats_df['mean_flow_rate_per_drawoff_LperH'] \
         / 60 * cats_df['drawoff_duration_min']
 
-    cats_df['mean_vol_per_day'] = mean_drawoff_vol_per_day * cats_df[
-        'portion']
+    cats_df['mean_vol_per_day'] = mean_drawoff_vol_per_day * cats_df['portion']
 
     cats_df['mean_vol_per_year'] = cats_df['mean_vol_per_day'] * 365
 
@@ -262,9 +262,9 @@ def get_data_drawoff_categories(s_step, categories, mean_drawoff_vol_per_day, bu
             = max(cats_df['mean_flow_rate_per_drawoff_LperH'].max(), 1200)
 
     else:
-        # add max flow rate: Max(500, highest category mean flow rate)
+        # add max flow rate: Max(1600, highest category mean flow rate)
         cats_df['max_flow_rate_per_drawoff_LperH'] \
-            = max(cats_df['mean_flow_rate_per_drawoff_LperH'].max(), 500)
+            = max(cats_df['mean_flow_rate_per_drawoff_LperH'].max(), 1600)
 
     return cats_df
 
[…]
