+++ examples/e5_generate_demands.py
@@ -10,8 +10,10 @@ Then choose 'Modify Run Configuration' and tick 'Run with Python Console'.
 
 # Import classes of the districtgenerator to be able to use the district generator.
 from districtgenerator.classes import *
+import pandas as pd
+import matplotlib.pyplot as plt
 
-def example1_4_generate_first_district():
+def example5_generate_demands():
 
     # Initialize District
     data = Datahandler()
@@ -30,43 +32,49 @@ def example1_4_generate_first_district():
     # if we want to calculate and save the demand profiles: If "calcUserProfiles=True", the datahandler
     # generates the profiles and saves them in the directory "results/demands/".
     # Alternatively we can load existing profiles. To do so, we put "calcUserProfiles=False".
+    # The Richardson tool is used to generate stochastic occupancy, internal heat gain and electric load profiles.
+    # With an 5R1C thermal building model the space heat profiles and a stochastic model the drinking hot water
+    # profiles are calculated.
     data.generateDemands(calcUserProfiles=True, saveUserProfiles=True)
 
     ### ===========================================  Output  =========================================== ###
     # The (demand) profiles for electricity demand of appliances and lighting (elec),
     # for heat demand of space heating (heat),domestic hot water (dhw), internal heat gains (gains)
     # and the time series for the presents of occupants (occ) are now calculated
-    # We can access them under data.district.id.user.
+    # We can access them under data.district.id.user or in the results folder.
 
-
-
-    # Here are some examples:
-
-    print("\nThe heating power of building 0 in timestep 0 is "+ str(round(data.district[0]['user'].heat[0])) + "W.")
-    print("The electricity power for plug-in devices of building 0 in timestep 0 is "
-          + str(round(data.district[0]['user'].elec[0])) + "W.\n")
-
-    # We can also use the profiles, to get the total demand or the maximun power.
-    # The given numbers in the demand profile are the power in that timestep. So if we have 15 Minute timesteps,
-    # the demand in each timestep equals the given power times 1/4 hour.
-    # To get the right units, we see how long one timestep is defined. As this value had to be inserted in seconds,
-    # we convert it into hours:
-    len_timestep = data.time["timeResolution"]/3600
-    total_heating_demand_0 = sum(
-        data.district[0]['user'].heat[t] for t in range(len(data.district[0]['user'].heat))) * len_timestep
-    total_heating_demand_1 = sum(
-        data.district[1]['user'].heat[t] for t in range(len(data.district[1]['user'].heat))) * len_timestep
-    max_heating_power_0 = max(data.district[0]['user'].heat)
-    print("The annual heating demand of building 0 is " + str(round(total_heating_demand_0)/1000) + "kWh.")
-    print("The maximum heating power of building 0 is " + str(round(max_heating_power_0)) + "W.")
-
-    print("The maximum electricity power of building 0 is " + str(round(max(data.district[0]['user'].elec))) + "W.\n")
+    # We can now use the profiles for exemplary analyses like monthly demands or peak loads.
+    # We plot the district space heat demand in kWh
+    exemplary_plot(data)
 
 
     return data
 
+def exemplary_plot(data):
+
+    # Sum heat demand of buildings
+    heat = data.district[0]["user"].heat + data.district[1]["user"].heat
+    # Unit conversion [kWh]
+    heat = heat / (data.time["dataResolution"] / data.time["timeResolution"]) / 1000
+
+    # Create a dataframe that contains the timestamps
+    date_range = pd.date_range(start='2023-01-01', periods=data.time["timeSteps"], freq='15T')
[…]
