+++ examples/e4_generate_buildings.py
@@ -11,7 +11,7 @@ Then choose 'Modify Run Configuration' and tick 'Run with Python Console'.
 # Import classes of the districtgenerator to be able to use the district generator.
 from districtgenerator.classes import *
 
-def example1_3_generate_first_district() :
+def example4_generate_buildings() :
     # Initialize District
     data = Datahandler()
 
@@ -39,4 +39,4 @@ def example1_3_generate_first_district() :
 
 
 if __name__ == '__main__' :
-    data = example1_3_generate_first_district()
+    data = example4_generate_buildings()
