+++ districtgenerator/classes/__init__.py
@@ -5,6 +5,5 @@ from .envelope import Envelope
 from .solar import Sun
 from .users import Users
 from .profils import Profiles
-from .plots import DemandPlots
 # import of participants as houses and the aggregator
 from .participants import *
