+++ examples/e2_generate_environment.py
@@ -5,7 +5,7 @@
 This is the second example to create the environment of the district.
 
 If you run the examples with Python console, you can see the output file.
-To do this right-hand click the example.py file (e1.0_generate_first_district.py).
+To do this right-hand click the example.py file (e2_generate_environment.py).
 Then choose 'Modify Run Configuration' and tick 'Run with Python Console'.
 """
 
@@ -22,16 +22,11 @@ def example2_generate_environment():
 
     ### =====================================  Output  ===================================== ###
     # Based on the location of the district the environment includes outside temperatures and
-    # sun radiation. When you open data.site, you see the "location" is [51.0,6.55], the
-    # "climateZone" 0 and the "altitude" 0. The weather data is taken from a Test Reference
-    # Year Database from the DWD.
-    # You find the time resolution in data.time. For this example it is 900 seconds,
-    # which is equal to 15 Minutes.
+    # sun radiation. When you open "data.site", you see the "location" is [51.0,6.55], the
+    # "climateZone" 0 and the "altitude" 0. The weather data is taken from a test reference
+    # year database from the DWD. You find the time resolution in "data.time". For this example
+    # it is 900 seconds, which is equal to 15 Minutes.
 
-    print("\nOur district now looks like this:")
-    print("District:" + str(data.district))
-    print("There is still no information about our district.\n")
-    print("But we added information about our environment into 'site'.\n")
 
     return data
 
