+++ docs/EXAMPLES.md
@@ -10,42 +10,43 @@ order of the examples represents the workflow of the DistrictGenerator.
 
 Within this example we explain how the user initializes the datahandler.
 
-- [Initialize Datahandler](https://github.com/RWTH-EBC/districtgenerator/blob/master/examples/e1.0_generate_first_district.py) 
+- [Initialize Datahandler](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e1_initialize_datahandler.py) 
 
 #### Generate environment for the district
     
 Within this example we generate the environment of the district.
 This includes data respectively time series on solar radiation 
 and outdoor temperature for a given location and test reference year.
-- [Generate Environment](https://github.com/RWTH-EBC/districtgenerator/blob/master/examples/e1.1_generate_first_district.py) 
+- [Generate Environment](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e2_generate_environment.py) 
 
 
 #### Initialize buildings of the district
 
 Within this example we initialize the buildings by the building type,
 the construction year, the retrofit level and the floor area.
-- [Initialize Buildings](https://github.com/RWTH-EBC/districtgenerator/blob/master/examples/e1.2_generate_first_district.py)
+- [Initialize Buildings](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e3_initialize_buildings.py)
 
 
 #### Extension of the building models
 
 Within this example we add more detailed information of the building components 
 and their materials to the building models.
-- [Extension Buildung Models](https://github.com/RWTH-EBC/districtgenerator/blob/master/examples/e1.3_generate_fist_district.py)
+- [Extension Buildung Models](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e4_generate_buildings.py)
 
 
 #### Generation and plotting of demand profiles
 
-- [Demand Profiles](./ngsi_v2/e11_ngsi_v2_semantics)
+Within this example we generate the (demand) profiles and exemplary plot the monthly space heat
+demand of the district.
+
+- [Demand Profiles](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e5_generate_demands.py)
 
 
 #### Customisation of specific assumptions for individual districts
 
-location, TRY
-The location can be changed in the site_data.json file. You find it in \data.
-The time resolution can be changed in time_data.json. You also find it in \data.
+Within this example we change specific assumptions like the location of the district,
+the minimal indoor temperature of the buildings, the weather test reference year and
+the time resolution.
 
-
-
-- [Semantics](./ngsi_v2/e11_ngsi_v2_semantics)
+- [Specify Assumptions](https://github.com/RWTH-EBC/districtgenerator/blob/15-joss-documentation/examples/e6_individual_district.py)
 

[…]
