+++ districtgenerator.egg-info/SOURCES.txt
@@ -1,0 +1,20 @@
+LICENSE.md
+README.md
+setup.py
+districtgenerator/__init__.py
+districtgenerator.egg-info/PKG-INFO
+districtgenerator.egg-info/SOURCES.txt
+districtgenerator.egg-info/dependency_links.txt
+districtgenerator.egg-info/requires.txt
+districtgenerator.egg-info/top_level.txt
+districtgenerator/classes/__init__.py
+districtgenerator/classes/datahandler.py
+districtgenerator/classes/envelope.py
+districtgenerator/classes/profils.py
+districtgenerator/classes/solar.py
+districtgenerator/classes/users.py
+districtgenerator/functions/__init__.py
+districtgenerator/functions/change_resolution.py
+districtgenerator/functions/dhw_stochastical.py
+districtgenerator/functions/heating_profile_5R1C.py
+tests/test_examples.py
