+++ joss/paper.md
@@ -1,0 +1,107 @@
+title: 'Districtgenerator: Generating building-specific load profiles for residential districts.'
+tags:
+  - Python
+  - District
+  - Building
+  - Load profile generation
+authors:
+  - name: Sarah Henn
+    orcid: 0000-0002-3423-7511
+    affiliation: 1
+  - name: Marius Brünjes
+    affiliation: 1
+  - name: Joel David Schölzel
+    orcid: 0009-0000-5053-3161
+    affiliation: 1
+  - name: Tobias Beckhölter
+    affiliation: 1
+  - name: Dirk Müller
+    orcid: 0000-0002-6106-6607
+    affiliation: 1
+affiliations:
+ - name: Institute for Energy Efficient Buildings and Indoor Climate, RWTH Aachen University
+   index: 1
+date: 01 June 2023
+bibliography: paper.bib
+
+# Summary
+
+The `Districtgenerator` enables an automated generation of time-resolved load profiles for residential districts. 
+The profiles generated for each building within the district are the demands for electricity, space heating demand, domestic hot water, and the occupancy profiles. 
+Additionally, a heating load calculation is carried out for each building.
+The `Districtgenerator` is conceptualized using a minimum amount of input data. 
+Thus, the tool is valuable for researchers and planners to obtain information  needed, for example, for the energy system or energy management system design in an early planning phase of a district.
+
+# Statement of need
+
+An essential part of facing global warming is to reduce the carbon footprint of the building sector, especially with a growing world population and a trend of urbanization in mind [@IEA_2021].
+The `Districtgenerator` offers knowledge, shows synergy potentials and supports scalability regarding building and district solutions.  
+
+* **Knowledge** of the energy demands is fundamental for the proper design and operation of any energy system.
+With the integration of distributed energy resources (DER), district energy systems and building energy systems get more complex.
+The energy supply of highly variable generators and user-dependent consumption need to be aligned. 
+Thus, detailed time-resolved profiles of energy demand and generation are required for a successful and future-oriented energy system design and operation. 
+
+* **Synergy** potentials between buildings of different usage structures can be exploited through joint concepts.
+While joint heating concepts, such as local heating networks, have been the focus of research for a long time, joint electricity concepts have only recently emerged.
+For example, the European Union proposed and introduced the concept of energy communities [@REDII_2018; @IEMD_2019].
+According to the @REDII_2018 directive, a district can also form an energy community where energy is shared or traded Peer-to-Peer.
+By this, district concepts lead to financial benefits for consumers and prosumers. 
+Moreover, self-consumption increases, peaks are evened out and the superordinate grid might be decongested. 
+
+* **Scalability** is a key issue for a resource-efficient transformation of the energy system, especially as numerous existing districts will have to be retrofitted and newly equipped in the next few years [@IEA_2021].
+The `Districtgenerator` enables its users to easily gain information about districts in order to identify promising concepts and conduct comparative evaluations.
+By using representative building types, data for representative districts is created. 
+Thus, the data applies to a wide range of residential districts and can be included in guidelines or representative examples.
+
+# Functional principle 
+
+The minimum input data required is the number of buildings  and basic information about each building, namely the building type, year of construction, retrofit level, and net floor area.
+The district size is freely choosable and the number of buildings to compute is not limited by the program. 
+The default file format for the building information is .csv because of its compatibility and ease of use.
+Optionally, the site of the district, the time resolution of the profiles and the test reference year (TRY) for weather data [@dwd_2023; @bbsr_2020] can be modified. 
+There are 16 choosable sites, which refer to the representative locations of the German climate zones. 
[…]
