+++ joss/paper.bib
@@ -1,0 +1,146 @@
+@article{remmen_teaser_2018,
+	title = {{TEASER}: an open tool for urban energy modelling of building stocks},
+	volume = {11},
+	issn = {1940-1493},
+	shorttitle = {{TEASER}},
+	doi = {10.1080/19401493.2017.1283539 },
+	number = {1},
+	journal = {Journal of Building Performance Simulation},
+	author = {Remmen, Peter and Lauster, Moritz and Mans, Michael and Fuchs, Marcus and Osterhage, Tanja and Müller, Dirk},
+	month = jan,
+	year = {2018},
+	keywords = {dynamic urban scale energy simulation, urban energy modelling, workflow automation},
+	pages = {84--98},
+}
+
+@article{wirtz_ehdo_2020,
+	title = {{EHDO}: A free and open-source webtool for designing and optimizing multi-energy systems based on MILP},
+	volume = {29},
+	issn = {1061-3773, 1099-0542},
+	shorttitle = {{EHDO}},
+	doi = {10.1002/cae.22352 },
+	number = {5},
+	journal = {Computer Applications in Engineering Education},
+	author = {Wirtz, Marco and Remmen, Peter and Müller, Dirk},
+	month = oct,
+	year = {2020},
+	keywords = {Energy hub, Hydrogen, MILP, Multi‐energy system, Webtool},
+	pages = {983--993},
+}
+
+@article{schiefelbein_automated_2019,
+title = {Automated urban energy system modeling and thermal building simulation based on OpenStreetMap data sets},
+journal = {Building and Environment},
+volume = {149},
+pages = {630 - 639},
+year = {2019},
+doi = {10.1016/j.buildenv.2018.12.025},
+author = {Jan Schiefelbein and Jana Rudnick and Anna Scholl and Peter Remmen and Marcus Fuchs and Dirk Müller},
+keywords = {City district, Urban energy models, Geographic information systems, Data enrichment, Archetypes, Workflow automation},
+}
+
+
+@article{richardson_domestic_2010,
+	title = {Domestic electricity use: {A} high-resolution energy demand model},
+	volume = {42},
+	issn = {03787788},
+	shorttitle = {Domestic electricity use},
+	doi = {10.1016/j.enbuild.2010.05.023},
+	language = {en},
+	number = {10},
+	urldate = {2022-06-26},
+	journal = {Energy and Buildings},
+	author = {Richardson, Ian and Thomson, Murray and Infield, David and Clifford, Conor},
+	month = oct,
+	year = {2010},
+	pages = {1878--1887},
+}
+
+@book{loga_deutsche_2015,
+	address = {Darmstadt},
+	edition = {2.},
+	title = {Deutsche {Wohngebäudetypologie} - {Beispielhafte} {Maßnahmen} zur {Verbesserung} der {Energieeffizienz} von typischen {Wohngebäuden}},
+	isbn = {978-3-941140-47-9},
+	shorttitle = {Deutsche {Wohngebäudetypologie}},
+	language = {de},
+	publisher = {{Institut Wohnen und Umwelt (IWU)}},
+	author = {Loga, Tobias and Stein, Britta and Diefenbach, Nikolaus and Born, Rolf},
+	year = {2015},
+}
+
+@misc{co2online_stromspiegel_2022,
+	title = {Stromspiegel für {Deutschland}},
+	url = {https://www.stromspiegel.de/fileadmin/ssi/stromspiegel/Downloads/Stromspiegel-2019-web.pdf},
+	author = {{co2online}},
+	year = {2019},
+}
+
+@misc{din_en_12831,
[…]
