### 2024-10-17 #6980 @JoelSchoelzel
Hello @AdamRJensen,
we have now added the following points: under docs you will find a Sphinx documentation (docs/build_/html/index to open it in the browser). We have also adapted the examples and added a test_examples.py for the functional testing of the examples and thus the main functions of the DistrictGenerator that the user can use. The bug in the paper has now also been fixed according to @kyleniemeyer advice.
We hope that this fulfils all requirements. Let me know if there are any other inadequacies.
Best regards
Joel
