### 2024-10-18 b2605e8
bugfix paper.md
