### 2024-10-18 720a2ad
Deletion of print statements
