### 2024-10-30 2bbb314
Design temperatures for non residential buildings
