+++ data/info.txt
@@ -38,6 +38,15 @@ The data folder contains essential information for simulating mixed-use district
    Contains CSV files describing different district scenarios, used as input for simulations.
    Data is generated by TECDEM:  https://github.com/TUB-DVG/TECDEM  
 
+
+7. design_building_data.json:
+   Contains design data for residential buildings.
+
+8. non_residential_design_building_data.json:
+   Contains design data for non-residential buildings.
+   Data is obtained from: https://github.com/IWUGERMANY/Non-domestic-multi-zone-average-usage-profiles-for-Germany for heating and cooling limits. 
+
+
 This structure allows for flexible and comprehensive modeling of various building types and their characteristics, enabling accurate simulation of mixed-use districts.
 
 
