### 2024-10-30 d3ab670
Design temperatures for non residential buildings
