### 2024-10-30 34ff89b
Adoption of C:m factors
