### 2024-10-30 f1fe70b
Residential builidngs test
