### 2024-10-31 8b51dc6
Updates type to float
