### 2024-10-31 c21a85a
Unused variable deleted
