### 2024-10-31 57a9ca9
Adapted tests to provide energy demand per square meter
