+++ tests/test_residential.py
@@ -8,54 +8,44 @@ from functions import path_checks
 # Should create simulations for residential buildings
 # Write a test that all functions run through without errors
 # Check if the results exist and are not zero
+
 class TestResidential(unittest.TestCase):
     def setUp(self):
         # Create test scenario
         self.test_data = { 
             'id': [0, 1, 2, 3],
-            'building': [
-                "SFH",
-                "SFH", 
-                "TH",
-                "TH"
-            ],
+            'building': ["SFH", "SFH", "TH", "TH"],
             'year': [rd.randint(1900, 2024) for _ in range(4)],
-            'retrofit': [0] * 4,
+            'retrofit': [rd.randint(0 , 2 ) for _ in range(4)]
             'area': [rd.randint(10, 100) for _ in range(4)]
         }
-
         df = pd.DataFrame(self.test_data, columns=self.test_data.keys())
         folder_path = os.getcwd()
         self.scenario_path = os.path.join(folder_path, 'data', 'scenarios', 'test_scenario.csv')
         df.to_csv(self.scenario_path, index=False, sep=';')
         print(df.head())
+        # Define a consistent results path
+        #C:\Users\felix\Programmieren\tecdm\src\districtgenerator\test_scenario
+        self.results_path = os.path.join(folder_path, 'test_scenario')
+        
 
     def test_non_residential(self):
-        # Test all functions in the non_residential module
         data = Datahandler()
         data.setWeatherFile('data/weather/EPW/DEU_BE_Berlin-Schonefeld.AP.103850_TMYx.2004-2018.epw')
-
-
-        # Generate Environment for the District
         data.generateEnvironment()
-
-        # Use the created scenario for testing
         data.initializeBuildings('test_scenario')
-        data.setResultPath()
+        data.setResultPath('test_scenario')
         data.generateBuildings()
-
         data.generateDemands()
-
-        # Add assertions to check if the results exist and are not zero
-        self.assertTrue(os.path.exists(self.scenario_path), f"Scenario file not found: {self.scenario_path}")
-        self.assertGreater(os.path.getsize(self.scenario_path), 0, f"Scenario file is empty: {self.scenario_path}")
+        # Verify that results are generated
+        self.assertTrue(os.path.exists(self.results_path), f"Results directory not found: {self.results_path}")
+        self.assertTrue(any(fname.endswith('.csv') for fname in os.listdir(self.results_path)), "No CSV files found in results directory.")
         folder_path = os.getcwd()
-        results_path = os.path.join(folder_path, 'results', 'demands', 'test_scenario')
+        results_path = os.path.join(folder_path, 'test_scenario')
         for file in os.listdir(results_path):
             if file.endswith('.csv'):
                 file_path = os.path.join(results_path, file)
                 df = pd.read_csv(file_path, sep=',')
-                # Check if the required columns are present
                 required_columns = ["elec", "dhw", "occ", "gains", "heat"]
                 for column in required_columns:
                     self.assertIn(column, df.columns, f"Column {column} not found in file: {file_path}")
@@ -63,37 +53,44 @@ class TestResidential(unittest.TestCase):
                 self.assertGreaterEqual(df["elec"].sum(), 0, f"Electricity sum is negative in file: {file_path}")
                 self.assertGreaterEqual(df["dhw"].sum(), 0, f"DHW sum is negative in file: {file_path}")
                 self.assertGreaterEqual(df["occ"].sum(), 0, f"Occupancy sum is negative in file: {file_path}")
-                # Gains can be negative due to cooling
                 self.assertTrue(df["gains"].notna().any(), f"Gains column has no values in file: {file_path}")
+        
+        scenario_data = pd.read_csv(self.scenario_path, sep=';')
+        for index, row in scenario_data.iterrows():
[…]
