### 2024-10-31 f857091
Adapted tests to provide energy demand per square meter
