### 2024-11-04 #6980 @AdamRJensen
@JoelSchoelzel could you recommend some additional reviewers. The first two you suggested are already reviewing other JOSS articles.
