### 2024-11-05 e1817dc
add cooling KPIs
