### 2024-11-11 3efd731
Fix of appliance demand calculation
