### 2024-11-12 f3b5156
Fixed bug in electrical demand calcluation
