### 2024-11-12 5bfcc4f
random electrical setup
