### 2024-11-12 f46eb11
Gauss variation occupancy
