### 2024-11-12 e379b5c
Gauss variation electrical demand
