### 2024-11-12 20d99d2
Fix electrical demand gym
