+++ .github/workflows/pylint.yml
@@ -18,6 +18,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pylint
-    - name: Analysing the code with pylint
-      run: |
-        pylint $(git ls-files '*.py')
+    - name: Run pylint with verbose output
+      run: pylint --verbose .
+
