### 2024-11-20 9b550ee
extend opti_central with energy hub
