+++ classes/system.py
@@ -4,6 +4,8 @@ import json
 import os
 import functions.opti_dimensioning_central_devices as opti_dimensioning_central_devices
 import functions.load_params_central_devices as load_params_central_devices
+from classes.solar import Sun
+import functions.wind_turbines as wind_turbines
 
 
 class BES:
@@ -202,5 +204,39 @@ class CES:
         # Run optimization
         capacities_centralDevices = opti_dimensioning_central_devices.run_optim(devs, param, dem, result_dict)
 
-
         return capacities_centralDevices
+
+    def generation(self, filePath, time, site):
+
+        global sun
+        sun = Sun(filePath=filePath)
+        # calculate theoretical PV generation
+        potentialPV, defaultSTC = \
+            sun.calcPVAndSTCProfile(time=time,
+                                    site=site,
+                                    area_roof=10,
+                                    # todo: auf internetseite ergänzen
+                                    # In Germany, this is a roof pitch between 30 and 35 degrees
+                                    beta=[35],
+                                    # surface azimuth angles (Orientation to the south: 0°)
+                                    gamma=[0],
+                                    usageFactorPV=1,
+                                    usageFactorSTC=0)
+
+        # calculate theoretical STC generation
+        defaultPV, pontentialSTC = \
+            sun.calcPVAndSTCProfile(time=time,
+                                    site=site,
+                                    area_roof=10,
+                                    # todo: auf internetseite ergänzen
+                                    # In Germany, this is a roof pitch between 30 and 35 degrees
+                                    beta=[35],
+                                    # surface azimuth angles (Orientation to the south: 0°)
+                                    gamma=[0],
+                                    usageFactorPV=0,
+                                    usageFactorSTC=1)
+
+        potentialWIND = wind_turbines.WT_generation(site["wind_speed"])
+
+
+        return (potentialPV, pontentialSTC, potentialWIND)
