+++ classes/optimizer.py
@@ -107,7 +107,7 @@ class Optimizer:
         None.
         """
 
-        results = opti_central.run_opti_central(self.model, self.data.district, self.data.site, self.cluster,
-                                                self.srcPath, optiData)
+        results = opti_central.run_opti_central(self.model, self.data.district, self.data.centralDevices, self.data.site,
+                                                self.cluster, self.srcPath, optiData)
 
         return results
