+++ functions/opti_central.py
@@ -8,7 +8,7 @@ import os, json
 import gurobipy as gp
 import time
 
-def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {}):
+def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath, optiData = {}):
 
     if optiData == {}:
         optiData["webtool"] = False
@@ -64,13 +64,21 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
     # %% Sets of energy conversion systems
     # heat generation devices (heat pump (HP), electric heating (EH),
     # combined heat and power (CHP), fuel cell (FC), boiler (BOI), solar thermal collector (STC))
-    # TODO: add STC
-    ecs_heat = ("HP", "EH", "CHP", "FC", "BOI", "STC")
+    ecs_heat = ("HP", "EH", "CHP", "FC", "BOI", "STC", "heat_grid")
     ecs_power = ("HP", "EH", "CHP", "FC", "PV")  # power consuming/producing devices (photovoltaic (PV))
     ecs_sell = ("CHP", "FC", "PV", "BAT")
     ecs_gas = ("CHP", "FC", "BOI")  # gas consuming devices
     ecs_storage = ("BAT", "TES", "EV")  # battery (BAT), thermal energy storage (TES), electric vehicle (EV)
     hp_modi = ("HP35", "HP55")  # modi of the HP with different HP supply temperatures in °C
+
+    # Create set for energy hub devices
+    eh_devs = ["PV", "WT", "STC", "WAT",
+                "HP", "EB", "CC", "AC",
+                "CHP", "BOI", "GHP",
+                "BCHP", "BBOI", "WCHP", "WBOI",
+                "ELYZ", "FC", "H2S", "SAB",
+                "TES", "CTES", "BAT", "GS",
+                ]
 
     # %% TECHNICAL PARAMETERS
     soc_nom = {}
@@ -201,6 +209,79 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
     # activation variable for trafo load
     yTrafo = model.addVars(time_steps,vtype="B",name="yTrafo_" + str(t))
 
+    # Device's capacity (i.e. rated power)
+    eh_cap = {}
+    for device in eh_devs:
+        eh_cap[device] = model.addVar(vtype="C", name="eh_nominal_capacity_" + str(device))
+
+    # Area used for PV and solar thermal collector installation
+    eh_area = {}
+    for device in ["PV", "STC"]:
+        eh_area[device] = model.addVar(vtype = "C", name="eh_roof_area_" + str(device))
+
+    # Gas flow to/from devices
+    eh_gas = {}
+    for device in ["CHP", "BOI", "GHP", "SAB", "from_grid", "to_grid"]:
+        eh_gas[device] = {}
+        for t in time_steps:
+            eh_gas[device][t] = model.addVar(vtype="C", name="gas_" + device + "_t" + str(t))
+
+    # Electric power to/from devices
+    eh_power = {}
+    for device in ["PV", "WT", "WAT", "HP", "EB", "CC", "CHP", "BCHP", "WCHP", "ELYZ",
+                   "FC", "from_grid", "to_grid"]:
+        eh_power[device] = {}
+        for t in time_steps:
+            eh_power[device][t] = model.addVar(vtype="C", name="eh_power_" + device + "_t" + str(t))
+
+    # Heat to/from devices
+    eh_heat = {}
[…]
