+++ requirements.txt
@@ -2,6 +2,6 @@ numpy>=1.20.1
 pandas>=1.2.3
 matplotlib>=3.3.4
 scipy>=1.6.1
-teaser
+teaser==0.7.7
 richardsonpy
 pylightxl
