+++ classes/KPIs.py
@@ -4,7 +4,7 @@ import sys
 import numpy as np
 import os
 import json
-
+import math
 
 class KPIs:
 
@@ -38,6 +38,7 @@ class KPIs:
         self.Gas_year = None
         self.dcf_year = None
         self.scf_year = None
+        self.annual_investment_total = None
 
         # initialize input data for calculation of KPIs
         inputData = {}
@@ -67,7 +68,7 @@ class KPIs:
         self.calculateEnergyExchangeWithinDistrict(data)
         self.calculateAutonomy()
         self.calculateCoverFactors(data)
-
+        self.calc_annual_cost_total(data.scenario)
 
     def prepareData(self, data):
         """
@@ -261,6 +262,108 @@ class KPIs:
             self.scf_year += self.supplyCoverFactor[c] * (self.inputData["clusterWeights"][self.inputData["clusters"][c]]
                                                           / sum_ClusterWeights)
 
+    def calc_annual_cost_total(self, scenario):
+
+        optiData = {}
+        optiData["HP"] = {}
+        optiData["HP"]["cap"] = 20 # kW
+        optiData["HP"]["life_time"] = 15
+        optiData["HP"]["inv_var"] = 900 # €/kW
+        optiData["HP"]["cost_om"] = 0.05
+
+        optiData["param"] = {}
+        optiData["param"]["observation_time"] = 20
+        optiData["param"]["interest_rate"] = 0.05
+
+        # Count occurrences of "BOI", "HP", and "CHP" in the 'heater' column
+        heater_counts = scenario['heater'].value_counts()
+
+        # Sum the values in the 'PV', 'STC', 'EV', and 'BAT' columns
+        heater_counts["PV"] = scenario['PV'].sum()
+        heater_counts["STC"] = scenario['STC'].sum()
+        heater_counts["EV"] = scenario['EV'].sum()
+        heater_counts["BAT"] = scenario['BAT'].sum()
+
+        calc_annual_investment = {}
+        for dev in ["BOI", "HP", "CHP", "PV", "STC", "EV", "BAT"]:
+            calc_annual_investment[dev] = {}
+            try:
+                if heater_counts[dev] > 0:
+                    calc_annual_investment[dev] = self.calc_annual_cost_device(optiData["HP"], optiData["param"])
+                else: calc_annual_investment[dev] = 0
+            except:
+                calc_annual_investment[dev] = 0
+                heater_counts[dev] = 0
+
+        self.annual_investment_total = calc_annual_investment["BOI"] * heater_counts["BOI"] \
+                                  + calc_annual_investment["HP"] * heater_counts["HP"] \
+                                  + calc_annual_investment["CHP"] * heater_counts["CHP"] \
+                                  + calc_annual_investment["PV"] * heater_counts["PV"] \
+                                  + calc_annual_investment["STC"] * heater_counts["STC"] \
+                                  + calc_annual_investment["EV"] * heater_counts["EV"] \
+                                  + calc_annual_investment["BAT"] * heater_counts["BAT"]
+    def calc_annual_cost_device(self, dev, param):
+        """
+        Calculation of total investment costs including replacements (based on VDI 2067-1, pages 16-17).
+
+        Parameters
+        ----------
+        dev : dictionary
+            technology parameter
+        param : dictionary
+            economic parameters
+
+        Returns
+        -------
+        annualized fix and variable investment
+        """
+
+        # Projektlaufzeit (observation_time)
+        # Zinssatz (für Kapitalkosten) (interest_rate)
+        # Anschaffungskosten inkl. Installation, etc. [€/kW]
+        # Lebensdauer [Jahre]
+        # Betriebs- und Wartungskosten [% of Invest]
[…]
