### 2024-12-09 #6980 @AdamRJensen
👋 @Haitham-ghaida, @jasondegraw & @eldemet, reaching out again one more time in the hope you might be able to review this submission for JOSS?

At JOSS our mission is to further open source science and for that we depend on reviewers like you!
