### 2024-12-13 8da0e91
fix: merged central + decentral devices optimization
