+++ functions/opti_central.py
@@ -8,8 +8,8 @@ import os, json
 import gurobipy as gp
 import time
 
+def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath, optiData = {}):
 
-def run_opti_central(model, buildingData, site, cluster, srcPath, optiData={}):
     if optiData == {}:
         optiData["webtool"] = False
 
@@ -19,23 +19,52 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData={}):
     # number of buildings in neighborhood
     nb = len(buildingData)
 
+
+    ecoData = {}
+    timeData = {}
+    for data in ("eco", "time"):
+        with open(srcPath + "/data/" + data + "_data.json") as json_file:
+            jsonData = json.load(json_file)
+            for subData in jsonData:
+                if data == "eco":
+                    ecoData[subData["name"]] = subData["value"]
+                else:
+                    timeData[subData["name"]] = subData["value"]
+    time_steps = range(int(timeData["clusterLength"] / timeData["timeResolution"]))
+    dt = timeData["timeResolution"] / timeData["dataResolution"]
+    last_time_step = len(time_steps) - 1
+
     # load parameters of decentral energy conversion devices
-    path = srcPath
     param_dec_devs = {}
-    with open(path + "/data/" + "decentral_device_data.json") as json_file:
+    with open(srcPath + "/data/" + "decentral_device_data.json") as json_file:
         jsonData = json.load(json_file)
         for subData in jsonData:
             param_dec_devs[subData["abbreviation"]] = {}
             for subsubData in subData["specifications"]:
                 param_dec_devs[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
 
+    # load parameters of central energy conversion devices
+    model_param_eh = {}
+    with open(srcPath + "/data/" + "model_parameters_EHDO.json") as json_file:
+        jsonData = json.load(json_file)
+        for subData in jsonData:
+            model_param_eh[subData["name"]] = subData["value"]
+    technical_param_eh = {}
+    with open(srcPath + "/data/" + "technical_parameters_EHDO.json") as json_file:
+        jsonData = json.load(json_file)
+        for subData in jsonData:
+            technical_param_eh[subData["abbreviation"]] = {}
+            for subsubData in subData["specifications"]:
+                technical_param_eh[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
+
+
     T_e = site["T_e_cluster"][cluster]  # ambient temperature [°C]
-    Q_DHW = {}  # DHW (domestic hot water) demand [W]
+    Q_DHW = {}      # DHW (domestic hot water) demand [W]
     Q_heating = {}  # space heating [W]
-    PV_gen = {}  # electricity generation of PV [W]
-    STC_heat = {}  # electricity generation of PV [W]
-    EV_dem = {}  # electricity demand electric vehicle (EV) [Wh]
-    elec_dem = {}  # electricity demand for appliances and lighting [W]
+    PV_gen = {}     # electricity generation of PV [W]
+    STC_heat = {}   # electricity generation of PV [W]
+    EV_dem = {}     # electricity demand electric vehicle (EV) [Wh]
+    elec_dem = {}   # electricity demand for appliances and lighting [W]
     occ = {}
     for n in range(nb):
[…]
