+++ examples/e8_scenario_evaluation.py
@@ -11,16 +11,16 @@ from classes import *
 
 def example8_scenario_evaluation():
 
-
-    import time
-    A_1 = time.time()
-
     # Initialize District
     data = Datahandler()
 
     # We directly generate a complete district.
-    data.generateDistrictComplete(scenario_name='Typquartier_1_2', calcUserProfiles=False, saveUserProfiles=False,
-                                  fileName_centralSystems="BF_Strategie_central_devices", saveGenProfiles=False)
+    #data.generateDistrictComplete(scenario_name='Quartier_3', calcUserProfiles=False, saveUserProfiles=False,
+    #                              fileName_centralSystems="BF_Strategie_central_devices", saveGenProfiles=False)
+    data.initializeBuildings(scenario_name='example')
+    data.generateEnvironment(plz=50672)
+    data.generateBuildings()
+    data.generateDemands(calcUserProfiles=True, saveUserProfiles=False)
 
 
     centralEnergySupply = False
@@ -28,13 +28,14 @@ def example8_scenario_evaluation():
     # data.designDevicesComplete(fileName_centralSystems="BF_Strategie_central_devices", saveGenerationProfiles=True)
     #input_Quartiersausweis = dataframe
     if centralEnergySupply == True:
-        data.initializeCentralDevices(fileName_centralSystems="BF_Strategie_central_devices")
         data.designDecentralDevices(saveGenerationProfiles=True)
-        data.designCentralDevices(saveGenerationProfiles=True)
+        data.designCentralDevices()
 
     else:
         data.designDecentralDevices(saveGenerationProfiles=True)
+        data.centralDevices = {}
         # data.designDecentralDevices(saveGenerationProfiles=True, input_webtool)
+
 
     # Within a clustered time series, data points are aggregated across different time periods
     # based on the k-medoids method
@@ -45,16 +46,13 @@ def example8_scenario_evaluation():
 
 
     # Calculation of the key performance indicators using the devices' operation profiles of clustered time periods
+    webtool = {}
     data.calulateKPIs()
 
     # TODO: erzeuge Energieausweis
 
 
     print("Congratulations! You calculated an optimized device operation for the selected neighborhood!")
-
-    A_2 = time.time()
-    print(str(A_1))
-    print(str(A_2))
 
     return data
 
