### 2024-12-16 472906a
update clustering with energy hub
