+++ classes/datahandler.py
@@ -573,28 +573,28 @@ class Datahandler:
             adjProfiles[id] = {}
             adjProfiles[id]["elec"] = self.district[id]["user"].elec[0:lenghtArray]
             adjProfiles[id]["dhw"] = self.district[id]["user"].dhw[0:lenghtArray]
-            #adjProfiles[id]["gains"] = self.district[id]["user"].gains[0:lenghtArray]
-            #adjProfiles[id]["occ"] = self.district[id]["user"].occ[0:lenghtArray]
             adjProfiles[id]["heat"] = self.district[id]["user"].heat[0:lenghtArray]
             adjProfiles[id]["cooling"] = self.district[id]["user"].cooling[0:lenghtArray]
-            if self.district[id]["buildingFeatures"]["EV"] != 0:
-                adjProfiles[id]["car"] = self.district[id]["user"].car[0:lenghtArray]
-            else:
-                # no EV exists; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["car"] = \
-                    self.district[id]["clusteringData"]["potentialEV"][0:lenghtArray] * sys.float_info.epsilon
-            if self.district[id]["buildingFeatures"]["PV"] != 0:
-                adjProfiles[id]["generationPV"] = self.district[id]["generationPV"][0:lenghtArray]
-            else:
-                # no PV module installed; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["generationPV"] = \
-                    self.district[id]["clusteringData"]["potentialPV"][0:lenghtArray] * sys.float_info.epsilon
-            if self.district[id]["buildingFeatures"]["STC"] != 0:
-                adjProfiles[id]["generationSTC"] = self.district[id]["generationSTC"][0:lenghtArray]
-            else:
-                # no STC installed; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["generationSTC"] = \
-                    self.district[id]["clusteringData"]["potentialSTC"][0:lenghtArray] * sys.float_info.epsilon
+        if centralEnergySupply == False:
+            for id in self.scenario["id"]:
+                if self.district[id]["buildingFeatures"]["EV"] != 0:
+                    adjProfiles[id]["car"] = self.district[id]["user"].car[0:lenghtArray]
+                else:
+                    # no EV exists; but array with just zeros leads to problem while clustering
+                    adjProfiles[id]["car"] = \
+                        self.district[id]["clusteringData"]["potentialEV"][0:lenghtArray] * sys.float_info.epsilon
+                if self.district[id]["buildingFeatures"]["PV"] != 0:
+                    adjProfiles[id]["generationPV"] = self.district[id]["generationPV"][0:lenghtArray]
+                else:
+                    # no PV module installed; but array with just zeros leads to problem while clustering
+                    adjProfiles[id]["generationPV"] = \
+                        self.district[id]["clusteringData"]["potentialPV"][0:lenghtArray] * sys.float_info.epsilon
+                if self.district[id]["buildingFeatures"]["STC"] != 0:
+                    adjProfiles[id]["generationSTC"] = self.district[id]["generationSTC"][0:lenghtArray]
+                else:
+                    # no STC installed; but array with just zeros leads to problem while clustering
+                    adjProfiles[id]["generationSTC"] = \
+                        self.district[id]["clusteringData"]["potentialSTC"][0:lenghtArray] * sys.float_info.epsilon
         # solar radiation on surfaces with different orientation
         adjProfiles["Sun"] = {}
         for drct in range(len(self.SunRad)):
@@ -636,13 +636,14 @@ class Datahandler:
         for id in self.scenario["id"]:
             inputsClustering.append(adjProfiles[id]["elec"])
             inputsClustering.append(adjProfiles[id]["dhw"])
[…]
