### 2024-12-16 9a50236
update 2 clustering with energy hub
