+++ functions/heating_profile_5R1C.py
@@ -10,12 +10,12 @@ import numpy.linalg as linalg
 def _solve(A, b):
     return linalg.solve(A,b)
 
-def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
+def _calculateNoHeat(zoneParameters, T_e, t_m_previous, dt,timestep):
     """
     Calculate the temperatures (T_op, T_m, T_air, T_s) if neither heating nor
-    cooling devices are activated. 
+    cooling devices are activated.
     This is necessary to enable a deadband between cooling and heating mode.
-    
+
     Parameters
     ----------
     zoneParameters : ZoneParameters
@@ -26,7 +26,7 @@ def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
         Initial temperature of the thermal mass in degree Celsius.
     timestep : integer, optional
         Define which index is relevant (zoneInputs, H_ve)
-        
+
     Returns
     -------
     T_op : float
@@ -38,116 +38,59 @@ def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
     T_s : float
         .
     """
-    
+
     # Note: If not stated differently, all equations, pages and sections
-    # refer to DIN EN ISO 13790:2008 (the official German version of 
+    # refer to DIN EN ISO 13790:2008 (the official German version of
     # ISO 13790:2008).
-    
+
     # Extract parameters
-    A_m     = zoneParameters.A_m            # in m2
-    #A_t     = zoneParameters.A_t            # in m2
     H_tr_is = zoneParameters.H_tr_is        # in W/K
     H_tr_ms = zoneParameters.H_tr_ms        # in W/K
     H_tr_w  = zoneParameters.H_tr_w         # in W/K
     H_ve    = zoneParameters.H_ve           # in W/K
     C_m     = zoneParameters.C_m            # in J/K
-    #if np.size(zoneParameters.H_tr_em) > 1:
-    #    H_tr_em = zoneParameters.H_tr_em [timestep] # in W/K
-    #else:
-    #    H_tr_em = zoneParameters.H_tr_em            # in W/K
     H_tr_em = zoneParameters.H_tr_em[0]  # in W/K
-
-    #dt      = zoneParameters.samplingRate # in s
-    
-    #Phi_int = zoneInputs.Phi_int[timestep]
-    #Phi_sol = zoneInputs.Phi_sol[timestep]
-
-    T_sup   = T_e
-    T_m_init = 20  # [°C]
-    
-    # Compute internal and solar heat sources
-    # Equations C1-C3, section C2, page 110
-    #Phi_ia = 0.5 * Phi_int
-    #Phi_m  = A_m / A_t * (0.5 * Phi_int + Phi_sol)
-    #Phi_st = (1 - A_m / A_t - H_tr_w / (9.1 * A_t)) * (0.5 * Phi_int + Phi_sol)
 
     Phi_ia = zoneParameters.phi_ia
     Phi_m  = zoneParameters.phi_m
     Phi_st = zoneParameters.phi_st
 
-    numberTimesteps = len(T_e)
-    # Initialize results
-    T_i  = np.zeros(numberTimesteps)
-    T_s  = np.zeros(numberTimesteps)
-    T_m  = np.zeros(numberTimesteps)
-    
+
     # Initialize A*x = b
     # x: T_m, T_s, T_air (T_i), Q_HC
     A = np.zeros((3,3))
     b = np.zeros(3)
-    
-    # Row wise entering
-    #A[0,0] = H_tr_em + H_tr_ms + C_m / dt
-    #A[0,1] = - H_tr_ms
-    #A[1,0] = - H_tr_ms
-    #A[1,1] = H_tr_ms + H_tr_is + H_tr_w
-    #A[1,2] = - H_tr_is
[…]
