### 2024-12-18 f4961f2
fix: merged some changes and added webtool settings for amortization period
