+++ classes/datahandler.py
@@ -545,8 +545,7 @@ class Datahandler:
             building["buildingFeatures"]["gamma"] = input_webtool[building["buildingFeatures"]["id"]]["gamma"]
             building["buildingFeatures"]["beta"] = input_webtool[building["buildingFeatures"]["id"]]["beta"]
 
-            building["buildingFeatures"]["EV"] = input_webtool[building["buildingFeatures"]["id"]][
-                "ev_input"]  # 0, small, medium or large
+            building["buildingFeatures"]["EV"] = input_webtool[building["buildingFeatures"]["id"]]["ev_input"]  # 0, small, medium or large
             building["buildingFeatures"]["ev_charging"] = input_webtool[building["buildingFeatures"]["id"]][
                 "ev_charging"]
 
@@ -701,69 +700,63 @@ class Datahandler:
         # loop over buildings
         for id in self.scenario["id"]:
             adjProfiles[id] = {}
-            # TODO: Ist es sinnvoll stochastische Zeitreihen zu clustern?
             adjProfiles[id]["elec"] = self.district[id]["user"].elec[0:lenghtArray]
             adjProfiles[id]["dhw"] = self.district[id]["user"].dhw[0:lenghtArray]
-            # adjProfiles[id]["gains"] = self.district[id]["user"].gains[0:lenghtArray]
-            # adjProfiles[id]["occ"] = self.district[id]["user"].occ[0:lenghtArray]
             adjProfiles[id]["heat"] = self.district[id]["user"].heat[0:lenghtArray]
             adjProfiles[id]["cooling"] = self.district[id]["user"].cooling[0:lenghtArray]
-            if self.district[id]["buildingFeatures"]["EV"] != 0:
-                adjProfiles[id]["car"] = self.district[id]["user"].car[0:lenghtArray]
-            else:
-                # no EV exists; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["car"] = \
-                    self.district[id]["clusteringData"]["potentialEV"][0:lenghtArray] * sys.float_info.epsilon
-            if self.district[id]["buildingFeatures"]["PV"] != 0:
-                adjProfiles[id]["generationPV"] = self.district[id]["generationPV"][0:lenghtArray]
-            else:
-                # no PV module installed; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["generationPV"] = \
-                    self.district[id]["clusteringData"]["potentialPV"][0:lenghtArray] * sys.float_info.epsilon
-            if self.district[id]["buildingFeatures"]["STC"] != 0:
-                adjProfiles[id]["generationSTC"] = self.district[id]["generationSTC"][0:lenghtArray]
-            else:
-                # no STC installed; but array with just zeros leads to problem while clustering
-                adjProfiles[id]["generationSTC"] = \
-                    self.district[id]["clusteringData"]["potentialSTC"][0:lenghtArray] * sys.float_info.epsilon
-
+        if centralEnergySupply == False:
+            for id in self.scenario["id"]:
+                if self.district[id]["buildingFeatures"]["EV"] != 0:
+                    adjProfiles[id]["car"] = self.district[id]["user"].car[0:lenghtArray]
+                else:
+                    # no EV exists; but array with just zeros leads to problem while clustering
+                    adjProfiles[id]["car"] = \
+                        self.district[id]["clusteringData"]["potentialEV"][0:lenghtArray] * sys.float_info.epsilon
+                if self.district[id]["buildingFeatures"]["PV"] != 0:
+                    adjProfiles[id]["generationPV"] = self.district[id]["generationPV"][0:lenghtArray]
+                else:
+                    # no PV module installed; but array with just zeros leads to problem while clustering
+                    adjProfiles[id]["generationPV"] = \
[…]
