+++ functions/opti_central.py
@@ -8,7 +8,7 @@ import os, json
 import gurobipy as gp
 import time
 
-def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {}):
+def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath, optiData = {}):
 
     if optiData == {}:
         optiData["webtool"] = False
@@ -19,15 +19,44 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
     # number of buildings in neighborhood
     nb = len(buildingData)
 
+
+    ecoData = {}
+    timeData = {}
+    for data in ("eco", "time"):
+        with open(srcPath + "/data/" + data + "_data.json") as json_file:
+            jsonData = json.load(json_file)
+            for subData in jsonData:
+                if data == "eco":
+                    ecoData[subData["name"]] = subData["value"]
+                else:
+                    timeData[subData["name"]] = subData["value"]
+    time_steps = range(int(timeData["clusterLength"] / timeData["timeResolution"]))
+    dt = timeData["timeResolution"] / timeData["dataResolution"]
+    last_time_step = len(time_steps) - 1
+
     # load parameters of decentral energy conversion devices
-    path = srcPath
     param_dec_devs = {}
-    with open(path + "/data/" + "decentral_device_data.json") as json_file:
+    with open(srcPath + "/data/" + "decentral_device_data.json") as json_file:
         jsonData = json.load(json_file)
         for subData in jsonData:
             param_dec_devs[subData["abbreviation"]] = {}
             for subsubData in subData["specifications"]:
                 param_dec_devs[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
+
+    # load parameters of central energy conversion devices
+    model_param_eh = {}
+    with open(srcPath + "/data/" + "model_parameters_EHDO.json") as json_file:
+        jsonData = json.load(json_file)
+        for subData in jsonData:
+            model_param_eh[subData["name"]] = subData["value"]
+    technical_param_eh = {}
+    with open(srcPath + "/data/" + "technical_parameters_EHDO.json") as json_file:
+        jsonData = json.load(json_file)
+        for subData in jsonData:
+            technical_param_eh[subData["abbreviation"]] = {}
+            for subsubData in subData["specifications"]:
+                technical_param_eh[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
+
 
     T_e = site["T_e_cluster"][cluster]  # ambient temperature [°C]
     Q_DHW = {}      # DHW (domestic hot water) demand [W]
@@ -40,37 +69,44 @@ def run_opti_central(model, buildingData, site, cluster, srcPath, optiData = {})
     for n in range(nb):
         Q_DHW[n] = buildingData[n]["user"].dhw_cluster[cluster]
         Q_heating[n] = buildingData[n]["user"].heat_cluster[cluster]
-        PV_gen[n] = buildingData[n]["generationPV_cluster"][cluster]
-        STC_heat[n] = buildingData[n]["generationSTC_cluster"][cluster]
-        EV_dem[n] = buildingData[n]["user"].car_cluster[cluster]
         elec_dem[n] = buildingData[n]["user"].elec_cluster[cluster]
[…]
