### 2025-01-06 1e71db7
Correction of the initial temperature in the year
