+++ classes/envelope.py
@@ -397,7 +397,6 @@ class Envelope:
             self.U["opaque"]["wall"] = prj.parameters["u_aw"]
             self.U["opaque"]["roof"] = prj.parameters["u_d_opak"]
             self.U["opaque"]["floor"] = prj.parameters["u_ug"]
-            self.U["opaque"]["window"] =prj.parameters["u_fen"]
             self.U["window"]  = prj.parameters["u_fen"]
             self.g_gl["window"] = prj.parameters["g_gl_fen"]
 
@@ -581,26 +580,27 @@ class Envelope:
                             * self.A["opaque"][x]) for x in self.opaque)
         elif isinstance(prj, NonResidential):
             # Assumption of thermal heat capacity
-            # According to EN ISO 13790:2008-09, S. 81 / DIBS
+            # According to EN ISO 13790:2008-09, S. 81
             # -----------------------------------------------------------------------------------------
             # |   Building Mass   |           Cm [in J/K]                      |
             # -----------------------------------------------------------------------------------------
-            # |    Very Light     |         80000 * Af                         |
-            # |      Light        |         110000 * Af                        |
-            # |      Medium       |         165000 * Af                        |
-            # |      Heavy        |         260000 * Af                        |
-            # |   Very Heavy      |         370000 * Af                        |
+            # |    Very Light     |         80000 * Am                         |
+            # |      Light        |         110000 * Am                        |
+            # |      Medium       |         165000 * Am                        |
+            # |      Heavy        |         260000 * Am                        |
+            # |   Very Heavy      |         370000 * Am                        |
             # -----------------------------------------------------------------------------------------
-            # Af are conditioned areas
+            # Am is the effective area responsible for the heat capacity
+            A_m = sum(self.A["opaque"][x] for x in self.opaque)
 
             if prj.construction_type == "Tabula":
-                self.C_m = 162000 * sum(self.A["opaque"][x] for x in self.opaque)
+                self.C_m = 162000 * A_m
             elif prj.construction_type == "Light":
-                self.C_m = 110000 * sum(self.A["opaque"][x] for x in self.opaque)
+                self.C_m = 110000 * A_m
             elif prj.construction_type == "Medium":
-                self.C_m = 165000 * sum(self.A["opaque"][x] for x in self.opaque)
+                self.C_m = 165000 * A_m
             elif prj.construction_type == "Heavy":
-                self.C_m = 260000 * sum(self.A["opaque"][x] for x in self.opaque)
+                self.C_m = 260000 * A_m
             else:
                 raise ValueError(
                     f"{prj.construction_type} currently not implemented for calculateHeatCapacity for Non Residential Buildings")
@@ -629,24 +629,6 @@ class Envelope:
         if SunRad is None:
             SunRad = []
         C_m = self.calculateHeatCapacity(self.prj)
-        temp = C_m / self.A["f"]
-
-        # 5 possible building classes DIN EN ISO 13790
-        x = np.zeros(5)
-        # upper and lower bounds for classes
-        low = [0.0, 95000.0, 137500.0, 212500.0, 315000.0]
-        up = [95000.0, 137500.0, 212500.0, 315000.0, 10000000.0]
-
-        for i in range(len(low)):
-            if low[i] <= temp <= up[i]:
-                x[i] = 1
-            else:
-                x[i] = 0
-
-        # Constants for calculation of A_m, dependent of building class
[…]
