+++ examples/e1.4_generate first_district.py
@@ -19,7 +19,7 @@ def example1_4_generate_first_district():
     data = Datahandler()
 
     # Generate Environment for the District
-    data.generateEnvironment(plz=35444)
+    data.generateEnvironment(plz="52064")
 
     # Initialize Buildings to the District
     data.initializeBuildings(scenario_name="example")
