+++ examples/e7_optiEnergyCentral_EHDO.py
@@ -25,14 +25,13 @@ def example7_optiEnergyCentral_EHDO():
     # the district generator. Further information about the technologies to be considered for the dimensioning of
     # the energy central and economic parameters are read in from further .csv and . json. data sources
 
-    data.designCentralDevices()
+    data.designCentralDevices(saveGenerationProfiles=True)
     # Within data the results of EHDO are given. For each device the annual generated amount of
     # electricity or heat as well as the nominal power or storage capacity are calculate.
     # Furthermore, ecological and economic indicatoers are calculated.
 
     print("Congratulations! You generated your energy central for the selected neighborhood!")
 
-    a=1
     return data
 
 if __name__ == '__main__':
