+++ EHDO/clustering_medoid.py
@@ -63,7 +63,7 @@ def cluster(inputs, number_clusters=12, norm=2, time_limit=300, mip_gap=0.0,
     Parameters
     ----------
     inputs : 2-dimensional array
-        First dimension: Number of different input types.
+        First dimension: Number of different input types "heat", "cool", "power", "hydrogen", "T_air", "GHI", "DHI", "wind_speed".
         Second dimension: Values for each time step of interes.
     number_clusters : integer, optional
         How many clusters shall be computed?
