### 2025-01-10 1ee024a
Change in the method for determining the number of residents in residential buildings
