### 2025-01-10 dca29cc
Change in the method for determining the number of residents in residential buildings (#23)
