### 2025-01-10 dfa9323
fix: merged some calculation for decentral devices
