+++ functions/opti_central.py
@@ -12,6 +12,9 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
 
     if optiData == {}:
         optiData["webtool"] = False
+    else:
+        optiData["webtool"] = True
+
 
     now = time.time()
     # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -624,9 +627,9 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
                                 name="End_" + str(device) + "_storage_" + str(n) + "_" + str(t))
 
             model.addConstr(ch_dom[device][n][t] == heat_dom["CHP"][n][t] + heat_dom["HP"][n][t] + heat_dom["BOI"][n][t]
-                            + heat_dom["EH"][n][t] + dhw_dom["EH"][n][t] + heat_dom["STC"][n][t] + heat_dom["heat_grid"][n][t],
+                            + heat_dom["EH"][n][t] + dhw_dom["EH"][n][t] + heat_dom["STC"][n][t],
                             name="Heat_charging_" + str(n) + "_" + str(t))
-            model.addConstr(dch_dom[device][n][t] == Q_DHW[n][t] + Q_heating[n][t],
+            model.addConstr(dch_dom[device][n][t] + heat_dom["heat_grid"][n][t] == Q_DHW[n][t] + Q_heating[n][t],
                             name="Heat_discharging_" + str(n) + "_" + str(t))
 
             device = "BAT"
