### 2025-01-10 c3601f7
#24 Correcting the 5R1C Method and the norm heat load calculation
