+++ classes/envelope.py
@@ -477,7 +477,7 @@ class Envelope:
 
     def calcHeatLoad(self, site, method="design"):
         """
-        Calculate heat load.
+        Calculate design (nominal) heat load at norm outside temperature
 
         Parameters
         ----------
@@ -505,28 +505,28 @@ class Envelope:
         U_TB = 0.05  # [W/m²K] Thermal bridge surcharge
         f_g1 = 1.45  # Correction factor for annual fluctuation of the outdoor temperature
         # Reduction factor
-        f_g2 = (self.T_set_min - T_me) / (self.T_set_min - T_ne)
+        f_g2 = (self.T_set_min - site["T_me"]) / (self.T_set_min - site["T_ne"])
         G_w = 1.0  # influence of groundwater neglected
 
         if method == "design":
             Q_nHC = (self.A["opaque"]["wall"] * (self.U["opaque"]["wall"] + U_TB) +
-                     self.A["window"]["sum"] * self.U["window"] +
-                     self.A["opaque"]["roof"] * self.U["opaque"]["roof"] +
-                     self.A["opaque"]["floor"] * self.U["opaque"]["floor"] * f_g1 * f_g2 * G_w
-                     + self.ventilationRate * self.c_p_air * self.rho_air * self.V / 3600) * (self.T_set_min - T_ne)
+                            self.A["window"]["sum"] * (self.U["window"] + U_TB) +
+                            self.A["opaque"]["roof"] * (self.U["opaque"]["roof"] + U_TB) +
+                            self.A["opaque"]["floor"] * self.U["opaque"]["floor"] * f_g1 * f_g2 * G_w +
+                            self.ventilationRate * self.c_p_air * self.rho_air * self.V / 3600) * (self.T_set_min - site["T_ne"])
 
         if method == "bivalent":
             Q_nHC = (self.A["opaque"]["wall"] * (self.U["opaque"]["wall"] + U_TB) +
-                     self.A["window"]["sum"] * self.U["window"] +
-                     self.A["opaque"]["roof"] * self.U["opaque"]["roof"] +
+                     self.A["window"]["sum"] * (self.U["window"] + U_TB) +
+                     self.A["opaque"]["roof"] * (self.U["opaque"]["roof"] + U_TB) +
                      self.A["opaque"]["floor"] * self.U["opaque"]["floor"] * f_g1 * f_g2 * G_w
                      + self.ventilationRate * self.c_p_air * self.rho_air * self.V / 3600) \
                        * (self.T_set_min - self.T_bivalent)
 
         if method == "heatlimit":
             Q_nHC = (self.A["opaque"]["wall"] * (self.U["opaque"]["wall"] + U_TB) +
-                     self.A["window"]["sum"] * self.U["window"] +
-                     self.A["opaque"]["roof"] * self.U["opaque"]["roof"] +
+                     self.A["window"]["sum"] * (self.U["window"] + U_TB) +
+                     self.A["opaque"]["roof"] * (self.U["opaque"]["roof"] + U_TB) +
                      self.A["opaque"]["floor"] * self.U["opaque"]["floor"] * f_g1 * f_g2 * G_w
                      + self.ventilationRate * self.c_p_air * self.rho_air * self.V / 3600) \
                        * (self.T_set_min - self.T_heatlimit)
