+++ classes/users.py
@@ -348,7 +348,7 @@ class Users:
         #self.elec = np.loadtxt(path + '/elec_' + unique_name + '.csv', delimiter=',')
         #self.gains = np.loadtxt(path + '/gains_' + unique_name + '.csv', delimiter=',')
 
-    def calcHeatingProfile(self, site, envelope, time_resolution):
+    def calcHeatingProfile(self, site, envelope, night_setback, holidays, time_resolution):
         """
         Calculate heat demand for each building.
 
@@ -369,21 +369,18 @@ class Users:
         None.
         """
 
-        dt = time_resolution/(60*60)
+        dt = time_resolution / (60 * 60)
         # calculate the temperatures (Q_HC, T_op, T_m, T_air, T_s)
-        (Q_HC, T_i, T_s, T_m, T_op) = heating.calculate(envelope, envelope.T_set_min, site["T_e"], dt)
-        #(Q_H, Q_C, T_op, T_m, T_i, T_s) = heating.calc(envelope, site["T_e"], dt)
-        # heating  load for the current time step in Watt
-        self.heat = np.zeros(len(Q_HC))
-        for t in range(len(Q_HC)):
-            self.heat[t] = max(0, Q_HC[t])
-        self.annual_heat_demand = np.sum(self.heat)
-
-        (Q_HC, T_i, T_s, T_m, T_op) = heating.calculate(envelope, envelope.T_set_max, site["T_e"], dt)
-        self.cooling = np.zeros(len(Q_HC))
-        for t in range(len(Q_HC)):
-            self.cooling[t] = min(0, Q_HC[t]) * (-1)
-        self.annual_cooling_demand = np.sum(self.cooling)
+        if night_setback == 1:
+            (Q_H, Q_C, T_op, T_m, T_i, T_s) = heating.calc_night_setback(envelope, site["T_e"], holidays, dt,
+                                                                         self.building)
+        elif night_setback == 0:
+            (Q_H, Q_C, T_op, T_m, T_i, T_s) = heating.calc(envelope, site["T_e"], holidays, dt, self.building)
+        # heating and cooling loads for the current time step in Watt
+        self.heat = Q_H
+        self.cooling = Q_C
+        self.annual_heat_demand = np.sum(Q_H)
+        self.annual_cooling_demand = np.sum(Q_C)
 
     def saveProfiles(self, unique_name, path):
         """
