### 2025-01-10 fff1a09
#24 Correcting the 5R1C Method and the norm heat load calculation (#25)
