+++ functions/heating_profile_5R1C.py
@@ -7,15 +7,17 @@ from __future__ import division
 import numpy as np
 import numpy.linalg as linalg
 
-def _solve(A, b):
-    return linalg.solve(A,b)
 
-def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
+def _solve(A, b):
+    return linalg.solve(A, b)
+
+
+def _calculateNoHeat(zoneParameters, T_e, t_m_previous, dt, timestep):
     """
     Calculate the temperatures (T_op, T_m, T_air, T_s) if neither heating nor
-    cooling devices are activated. 
+    cooling devices are activated.
     This is necessary to enable a deadband between cooling and heating mode.
-    
+
     Parameters
     ----------
     zoneParameters : ZoneParameters
@@ -26,7 +28,7 @@ def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
         Initial temperature of the thermal mass in degree Celsius.
     timestep : integer, optional
         Define which index is relevant (zoneInputs, H_ve)
-        
+
     Returns
     -------
     T_op : float
@@ -38,73 +40,440 @@ def _calculateNoHeat(zoneParameters, T_e, t_previous, dt, timestep):
     T_s : float
         .
     """
-    
+
     # Note: If not stated differently, all equations, pages and sections
-    # refer to DIN EN ISO 13790:2008 (the official German version of 
+    # refer to DIN EN ISO 13790:2008 (the official German version of
     # ISO 13790:2008).
-    
+
     # Extract parameters
-    A_m     = zoneParameters.A_m            # in m2
-    #A_t     = zoneParameters.A_t            # in m2
-    H_tr_is = zoneParameters.H_tr_is        # in W/K
-    H_tr_ms = zoneParameters.H_tr_ms        # in W/K
-    H_tr_w  = zoneParameters.H_tr_w         # in W/K
-    H_ve    = zoneParameters.H_ve           # in W/K
-    C_m     = zoneParameters.C_m            # in J/K
-    #if np.size(zoneParameters.H_tr_em) > 1:
-    #    H_tr_em = zoneParameters.H_tr_em [timestep] # in W/K
-    #else:
-    #    H_tr_em = zoneParameters.H_tr_em            # in W/K
+    H_tr_is = zoneParameters.H_tr_is  # in W/K
+    H_tr_ms = zoneParameters.H_tr_ms  # in W/K
+    H_tr_w = zoneParameters.H_tr_w  # in W/K
+    H_ve = zoneParameters.H_ve  # in W/K
+    C_m = zoneParameters.C_m  # in J/K
     H_tr_em = zoneParameters.H_tr_em[0]  # in W/K
 
-    #dt      = zoneParameters.samplingRate # in s
-    
-    #Phi_int = zoneInputs.Phi_int[timestep]
-    #Phi_sol = zoneInputs.Phi_sol[timestep]
-
-    T_sup   = T_e
-    T_m_init = 20  # [°C]
-    
-    # Compute internal and solar heat sources
-    # Equations C1-C3, section C2, page 110
-    #Phi_ia = 0.5 * Phi_int
-    #Phi_m  = A_m / A_t * (0.5 * Phi_int + Phi_sol)
-    #Phi_st = (1 - A_m / A_t - H_tr_w / (9.1 * A_t)) * (0.5 * Phi_int + Phi_sol)
-
     Phi_ia = zoneParameters.phi_ia
-    Phi_m  = zoneParameters.phi_m
+    Phi_m = zoneParameters.phi_m
[…]
