### 2025-01-10 104f59f
#28 Accounting for seasonal fluctuations to reflect changes in the desired domestic hot water temperature and the cold water temperature throughout the year. (#29)
