+++ classes/profils.py
@@ -3,14 +3,13 @@
 import os
 import random as rd
 import json
-
+import math
 import numpy as np
 import pylightxl as xl
-import richardsonpy.classes.occupancy as occ
+import richardsonpy.classes.occupancy as occ_residential
 import richardsonpy.functions.change_resolution as cr
 import functions.OpenDHW as OpenDHW
-import functions.dhw_stochastical as dhw_profil
-
+import functions.change_resolution as chres
 
 class Profiles:
     """
@@ -21,6 +20,8 @@ class Profiles:
     ----------
     number_occupants : integer
         Number of occupants who live in the house or flat.
+    number_occupants_building : integer
+        Number of occupants who live in the building.
     initial_day : integer
         Day of the week with which the generation starts.
         1-7 for monday-sunday.
@@ -34,14 +35,14 @@ class Profiles:
     activity_profile : array-like
         Numpy-array with active occupants 10-minutes-wise.
     occ_profile : array-like
-        Stochastic occupancy profiles for a district.
+        Stochastic occupancy profile.
     app_load : array-like
         Electric load profile of appliances in W.
     light_load : array-like
         Electric load profile of lighting in W.
     """
 
-    def __init__(self, number_occupants, initial_day, nb_days, time_resolution):
+    def __init__(self, number_occupants, number_occupants_building, initial_day, nb_days, time_resolution, building):
         """
         Constructor of Profiles class.
 
@@ -51,19 +52,23 @@ class Profiles:
         """
 
         self.number_occupants = number_occupants
+        self.number_occupants_building = number_occupants_building
         self.initial_day = initial_day
         self.nb_days = nb_days
         self.time_resolution = time_resolution
 
+        self.building = building
         self.activity_profile = []
         self.occ_profile = []
+        self.occ_profile_building = []
+        self.building_profiles = {}
+        self.temperature_difference = []
         self.light_load = []
         self.app_load = []
 
-        self.generate_activity_profile()
-        self.loadProbabilitiesDhw()
+        self.generate_activity_profile_residential()
 
-    def generate_activity_profile(self):
+    def generate_activity_profile_residential(self):
         """
         Generate a stochastic activity profile
         (on base of ridchardsonpy).
@@ -83,10 +88,11 @@ class Profiles:
         None.
         """
 
-        activity = occ.Occupancy(self.number_occupants, self.initial_day, self.nb_days)
-        self.activity_profile = activity.occupancy
+        if self.building in {"SFH", "TH", "MFH", "AB"}:
+            activity = occ_residential.Occupancy(self.number_occupants, self.initial_day, self.nb_days)
+            self.activity_profile = activity.occupancy
 
-    def generate_occupancy_profiles(self):
+    def generate_occupancy_profiles_residential(self):
         """
         Generate stochastic occupancy profiles for a district for calculating internal gains.
         Change time resolution of 10 min profiles to required resolution.
@@ -168,38 +174,58 @@ class Profiles:
         # Load profiles
         self.prob_profiles_dhw = profiles
 
-    def generate_dhw_profile(self, building):
+    def generate_dhw_profile(self, building, holidays):
         """
         Generate a stochastic dhw profile
         (on base of DHWclac).
-        https://www.uni-kassel.de/maschinenbau/institute/thermische-energietechnik/fachgebiete/solar-und-anlagentechnik/downloads
-        https://github.com/RWTH-EBC/OpenDHW
 
-        Parameters
-        ----------
-        s_step : integer
-            Resolution of time steps of output array in seconds.
-        Categories: 1 or 4
[…]
