### 2025-01-10 1e57cb8
#30 Improve Residential Electricity Consumption with a Probabilistic Approach. Correcting text, adding comments and sources, and fixing bugs.
