### 2025-01-10 088017e
#30 Improve Residential Electricity Consumption with a Probabilistic Approach. (#31)
Correcting text, adding comments and sources, and fixing bugs.
