### 2025-01-12 #7657 @editorialbot
[REVIEW]: DistrictGenerator: Generating Building-Specific Load Profiles for Residential Districts
**Submitting author:** @JoelSchoelzel (<a href="http://orcid.org/0009-0000-5053-3161">Joel Schoelzel</a>)
**Repository:** https://github.com/RWTH-EBC/districtgenerator
**Branch with paper.md** (empty if default branch): JOSS_submission
**Version:** v0.1.1
