+++ EHDO/solar_modeling.py
@@ -1,914 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-
-EHDO - ENERGY HUB DESIGN OPTIMIZATION Tool
-
-Developed by:   E.ON Energy Research Center, 
-                Institute for Energy Efficient Buildings and Indoor Climate, 
-                RWTH Aachen University, 
-                Germany
-               
-Contact:        Marco Wirtz 
-                marco.wirtz@eonerc.rwth-aachen.de
-                
-Solar modeling script developed by Rafal Broda (E.ON ERC EBC, RWTH Aachen University, 2021).                
-
-"""
-
-import os
-import math
-import numpy as np
-import pandas as pd
-import csv
-
-def load_epw(weather_file):
-    # Import energy plus weather file
-    header = {}
-    with open(os.path.join(weather_file), newline="", errors="ignore") as csvfile:
-        csvreader = csv.reader(csvfile, delimiter=",", quotechar='"')
-        for row in csvreader:
-            if row[0].isdigit():
-                break
-            else:
-                header[row[0]]=row[1:]
-
-    timezone = float(header["LOCATION"][7])
-    altitude = float(header["LOCATION"][8])
-
-    param = {}
-
-    file = open(os.path.join(weather_file), "rb")
-    T_air, GHI, DHI, wind_speed = np.loadtxt(file, delimiter=",", skiprows=8, usecols=[6,13,15,21], unpack=True)
-
-    param["T_air"] = T_air
-    param["wind_speed"] = wind_speed
-    param["direct_horiz_irrad"] = GHI - DHI
-    param["diffuse_horiz_irrad"] = DHI
-    param["global_irrad"] = GHI
-
-    data = pd.DataFrame(param)
-
-    return timezone, altitude, data
-
-def load_dwd(weather_file):
-    # Import dwd weather file
-    param = {}
-
-    T_air = []
-    wind_speed = []
-    direct_horizontal_irrad = []
-    diffuse_horizontal_irrad = []
-
-    with open(os.path.join(weather_file), "r") as weather_file:
-        file = weather_file.readlines()
-        for line in range(34,len(file)):
-            cleaned_list = [val for val in file[line].split(" ") if val != ""]
-            T_air.append(float(cleaned_list[5]))
-            wind_speed.append(float(cleaned_list[8]))
-            direct_horizontal_irrad.append(float(cleaned_list[12]))
-            diffuse_horizontal_irrad.append(float(cleaned_list[13]))
-
-    param["T_air"] = np.array(T_air)
-    param["wind_speed"] = np.array(wind_speed)
-    param["direct_horiz_irrad"] = np.array(direct_horizontal_irrad)
-    param["diffuse_horiz_irrad"] = np.array(diffuse_horizontal_irrad)
-    param["global_irrad"] = param["direct_horiz_irrad"] + param["diffuse_horiz_irrad"]
-
-    data = pd.DataFrame(param)
-
-    return data
-
-def getGeometry(initialTime, timeDiscretization, timesteps, timeZone=1,
-                location=(50.76, 6.07), altitude=0):
-    """
-    This function computes hour angle, declination, zenith angle of the sun
-    and solar azimuth angle for a given location and time.
-
-    The implemented equations can be found in:
-    Duffie, Beckman - Solar Engineering of Thermal Processes, 2013 (4th ed.)
-
-    Function copied from sun.py script, developed by author Thomas Schuetz
-    https://github.com/RWTH-EBC/BuildingOPT/blob/master/sun.py
-
-    Parameters
-    ----------
-    initialTime : integer
-        Time passed since January 1st, 00:00:00 in seconds
-    timeDiscretization : integer
[…]
