+++ examples/e8_scenario_evaluation.py
@@ -5,7 +5,7 @@ We reached the final step, to generate our first district: Generate demand profi
 """
 
 # Import classes of the districtgenerator to be able to use the district generator.
-from classes import *
+from districtgenerator.classes import *
 
 
 
@@ -14,24 +14,18 @@ def example8_scenario_evaluation():
     # Initialize District
     data = Datahandler()
 
-    # We directly generate a complete district.
-    #data.generateDistrictComplete(scenario_name='Quartier_3', calcUserProfiles=False, saveUserProfiles=False,
-    #                              fileName_centralSystems="BF_Strategie_central_devices", saveGenProfiles=False)
     data.initializeBuildings(scenario_name='example')
     data.generateEnvironment(plz="52064")
     data.generateBuildings()
-    data.generateDemands(calcUserProfiles=True, saveUserProfiles=False)
-
+    data.generateDemands(calcUserProfiles=False, saveUserProfiles=False)
 
     centralEnergySupply = True
     # Sizing of the selected devices
     if centralEnergySupply == True:
-        # data.designDecentralDevices(saveGenerationProfiles=True)
-        data.designCentralDevices(saveGenerationProfiles=True)
+        data.designCentralDevices(saveGenerationProfiles=False)
     else:
         data.designDecentralDevices(saveGenerationProfiles=True)
         data.centralDevices = {}
-        # data.designDecentralDevices(saveGenerationProfiles=True, input_webtool)
 
     # Within a clustered time series, data points are aggregated across different time periods
     # based on the k-medoids method
