+++ districtgenerator/functions/opti_central.py
@@ -101,7 +101,7 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
                 ]
 
     try:
-        print(buildingData[0]["capacities"])
+        a = buildingData[0]["capacities"]
     except KeyError:
         for n in range(nb):
             buildingData[n]["capacities"] = {}
@@ -242,11 +242,6 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
     for device in eh_devs:
         eh_cap[device] = model.addVar(vtype="C", name="eh_nominal_capacity_" + str(device))
 
-    # Area used for PV and solar thermal collector installation
-    eh_area = {}
-    for device in ["PV", "STC"]:
-        eh_area[device] = model.addVar(vtype = "C", name="eh_roof_area_" + str(device))
-
     # Gas flow to/from devices
     eh_gas = {}
     for device in ["CHP", "BOI", "GHP", "SAB", "from_grid", "to_grid"]:
@@ -299,14 +294,19 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
             eh_waste[device][t] = model.addVar(vtype="C", name="waste_" + device + "_t" + str(t))
 
     # Storage variables
-    eh_ch = {}  # Energy flow to charge storage device
+    eh_ch = {}# Energy flow to charge storage device
+    eh_dch = {}# Energy flow to discharge storage device
     eh_soc = {} # State of charge
+    eh_soc_init = {} # Initial state of charge
     for device in ["TES", "CTES", "BAT", "H2S", "GS"]:
         eh_ch[device] = {}
         eh_soc[device] = {}
+        eh_dch[device] = {}
+        # Initial state of charge
+        eh_soc_init[device] = eh_cap[device] * 0.5  # Wh
         for t in time_steps:
-            # For charge variable: ch is positive if storage is charged, and negative if storage is discharged
             eh_ch[device][t] = model.addVar(vtype="C", lb=-gp.GRB.INFINITY, name="eh_ch_" + device + "_t" + str(t))
+            eh_dch[device][t] = model.addVar(vtype="C", lb=-gp.GRB.INFINITY, name="eh_dch_" + device + "_t" + str(t))
             eh_soc[device][t] = model.addVar(vtype="C", name="soc_" + device + "_t" + str(t))
 
 
@@ -369,13 +369,19 @@ def run_opti_central(model, buildingData, energyHubData, site, cluster, srcPath,
             if energyHubData == {}:
                 model.addConstr(eh_heat[device][t] == 0)
             else:
-                model.addConstr(eh_heat[device][t] <= energyHubData["capacities"]["heat_kW"][device] * 1000) # W
-        for device in ["PV", "WT", "WAT", "CHP", "BCHP", "WCHP", "ELYZ", "FC"]:
+                model.addConstr(eh_heat[device][t] <= energyHubData["capacities"]["heat_kW"][device] * 10000) # W
+        model.addConstr(eh_heat["STC"][t] <= energyHubData["generation"]["STC_cluster"][cluster][t],
+                        name="STC_generation_energyHub_" + str(t))
+        for device in ["PV", "WT", "WAT", "CHP", "BCHP", "WCHP", "ELYZ", "FC", "from_grid", "to_grid"]:
             # todo: webtool
             if energyHubData == {}:
                 model.addConstr(eh_power[device][t] == 0)
             else:
                 model.addConstr(eh_power[device][t] <= energyHubData["capacities"]["power_kW"][device] * 1000) # W
[…]
