+++ districtgenerator/classes/datahandler.py
@@ -11,15 +11,15 @@ import openpyxl
 import pandas as pd
 from itertools import count
 from teaser.project import Project
-from classes.envelope import Envelope
-from classes.solar import Sun
-from classes.users import Users
-from classes.system import BES
-from classes.system import CES
-from classes.plots import DemandPlots
-from classes.optimizer import Optimizer
-from classes.KPIs import KPIs
-import functions.clustering_medoid as cm
+from .envelope import Envelope
+from .solar import Sun
+from .users import Users
+from .system import BES
+from .system import CES
+from .plots import DemandPlots
+from .optimizer import Optimizer
+from .KPIs import KPIs
+import districtgenerator.functions.clustering_medoid as cm
 
 class Datahandler:
     """
@@ -383,7 +383,7 @@ class Datahandler:
             }
             building["user"].car *= building["buildingFeatures"]["EV"]
 
-            building["envelope"].calcNormativeProperties(self.SunRad, building["user"].gains)
+            building["envelope"].calcNormativeProperties(self.site["SunRad"], building["user"].gains)
 
             night_setback = building["buildingFeatures"]["night_setback"]
 
@@ -403,8 +403,7 @@ class Datahandler:
         print("Finished generating demands!")
 
     def generateDistrictComplete(self, scenario_name='example', calcUserProfiles=True, saveUserProfiles=True, plz="52064",
-                                 fileName_centralSystems="central_devices_example", saveGenProfiles=True,
-                                 designDevs=False, clustering=False, optimization=False):
+                                 saveGenProfiles=True, designDevs=False, clustering=False, optimization=False):
         """
         All in one solution for district and demand generation.
 
@@ -443,7 +442,7 @@ class Datahandler:
         self.generateBuildings()
         self.generateDemands(calcUserProfiles, saveUserProfiles)
         if designDevs:
-            self.designDevicesComplete(fileName_centralSystems, saveGenProfiles)
+            self.designDevicesComplete(saveGenProfiles)
         if clustering:
             if designDevs:
                 self.clusterProfiles()
@@ -511,11 +510,11 @@ class Datahandler:
 
             # optionally save generation profiles
             if saveGenerationProfiles == True:
-                np.savetxt(os.path.join(self.resultPath, 'renewableGeneration')
+                np.savetxt(os.path.join(self.resultPath, 'generation')
                            + '/decentralPV_' + building["unique_name"] + '.csv',
                            building["generationPV"],
                            delimiter=',')
-                np.savetxt(os.path.join(self.resultPath, 'renewableGeneration')
+                np.savetxt(os.path.join(self.resultPath, 'generation')
                            + '/decentralSTC_' + building["unique_name"] + '.csv',
                            building["generationSTC"],
                            delimiter=',')
@@ -547,23 +546,23 @@ class Datahandler:
 
         # calculate theoretical PV, STC and Wind generation
         self.centralDevices["generation"] = {}
-        self.centralDevices["generation"]["PV"] = self.centralDevices["ces_obj"].generation(self.filePath, self.time, self.site)[0]
-        self.centralDevices["generation"]["STC"] = self.centralDevices["ces_obj"].generation(self.filePath, self.time, self.site)[1]
-        self.centralDevices["generation"]["Wind"] = self.centralDevices["ces_obj"].generation(self.filePath, self.time, self.site)[2]
+        (self.centralDevices["generation"]["PV"], self.centralDevices["generation"]["STC"],
+         self.centralDevices["generation"]["Wind"]) = self.centralDevices["ces_obj"].generation(self)
+
 
         # optionally save generation profiles
         if saveGenerationProfiles == True:
[…]
