### 2025-01-16 59bbdc4
#33 move first imports into new function load_all_data
