### 2025-01-16 2a964c7
Add: Function to export building model
