+++ classes/datahandler.py
@@ -9,6 +9,7 @@ import datetime
 import numpy as np
 import openpyxl
 import pandas as pd
+import matplotlib.pyplot as plt
 from itertools import count
 from teaser.project import Project
 from classes.envelope import Envelope
@@ -103,7 +104,7 @@ class Datahandler:
 
         return weatherdatafile_location
 
-    def generateEnvironment(self, plz):
+    def generateEnvironment(self, plz, TRY: int = None, weather_conditions: str = None):
         """
         Load physical district environment - site and weather.
 
@@ -119,8 +120,16 @@ class Datahandler:
             for subData in jsonData:
                 self.site[subData["name"]] = subData["value"]
 
+        # %% load weather data for site
+        # extract irradiation and ambient temperature
+        if TRY is not None:
+            self.site["TRYYear"] = "TRY" + str(TRY)
+
+        if weather_conditions is not None:
+            self.site["TRYType"] = weather_conditions
 
         # %% load first day of the year
+
         if self.site["TRYYear"] == "TRY2015":
             first_row = 35
         elif self.site["TRYYear"] == "TRY2045":
@@ -199,7 +208,6 @@ class Datahandler:
         self.site["T_me"] = filtered_data.iloc[0]['T_me'] # mean annual temperature for calculating the design heat load
 
         # Calculate solar irradiance per surface direction - S, W, N, E, Roof represented by angles gamma and beta
-        global sun
         sun = Sun(filePath=self.filePath)
         self.SunRad = sun.getSolarGains(initialTime=0,
                                         timeDiscretization=self.time["timeResolution"],
@@ -213,7 +221,8 @@ class Datahandler:
                                         diffuseRadiation=self.site["SunDiffuse"],
                                         albedo=self.site["albedo"])
 
-    def initializeBuildings(self, scenario_name='example'):
+    # def initializeBuildings(self, scenario_name='example'):
+    def initializeBuildings(self, list_all: list = None, scenario_name='example', path=None):
         """
         Fill district with buildings from scenario file.
 
@@ -230,29 +239,70 @@ class Datahandler:
         num_sfh = 0
         num_mfh = 0
         self.scenario_name = scenario_name
-        self.scenario = pd.read_csv(os.path.join(self.filePath, 'scenarios')
-                                    + "/"
-                                    + self.scenario_name + ".csv",
-                                    header=0, delimiter=";")
-
-        # initialize buildings for scenario
-        # loop over all buildings
-        for id in self.scenario["id"]:
-            # create empty dict for observed building
-            building = {}
-
-            # store features of the observed building
-            building["buildingFeatures"] = self.scenario.loc[id]
-
-            # append building to district
-            self.district.append(building)
-
-            # Count number of builidings to predict the approximate calculation time
-            if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"]["building"] == 'TH': num_sfh +=1
-            elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"]["building"] == 'AB': num_mfh +=1
+        if list_all is not None and len(list_all) != 0:
+            self.scenario = pd.DataFrame(list_all, columns=['id', 'type', 'year', 'condition', 'area'])
+        else:
+            if path is None:
+                self.scenario = pd.read_csv(os.path.join(self.filePath, 'scenarios')
+                                            + "/"
+                                            + self.scenario_name + ".csv",
+                                            header=0, delimiter=";")
+            else:
+                self.scenario = pd.read_csv(os.path.join(path)
+                                            + "/"
+                                            + self.scenario_name + ".csv",
[…]
