### 2025-01-16 7a3e375
add further data imports to functions #31
