### 2025-01-17 ed2d4d5
debug data loading and example 1_3, #33
