+++ districtgenerator/functions/load_params_central_devices.py
@@ -83,11 +83,8 @@ def load_params(data):
             dhw += data.district[b]["user"].dhw / 1000 # kW
             electricityAppliances += data.district[b]["user"].elec / 1000 # kW
 
-
-
     heat_total = heating + dhw
     electricity_total = electricityAppliances
-
     dem_uncl["heat"] = heat_total
     dem_uncl["cool"] = cooling
     dem_uncl["power"] = electricity_total
@@ -95,26 +92,35 @@ def load_params(data):
         param["peak_"+k] = np.max(dem_uncl[k])
     param["peak_hydrogen"] = 0
 
-
     ################################################################
     # DESIGN CLUSTERING
-    param["n_clusters"] = 12 #data.time["clusterNumber"]  # Number of time cluster
+
+    # calculate cluster time horizon
+    clusterHorizon = int((data.time["clusterLength"] / data.time["timeResolution"]))
+    adjustedHorizon = clusterHorizon
+    while adjustedHorizon <= len(data.site["T_e"]):
+        adjustedHorizon += clusterHorizon
+    adjustedHorizon -= clusterHorizon
+    adjustedHorizon = int(adjustedHorizon)
 
     # Collect the time series to be clustered
-    time_series = [dem_uncl["heat"], dem_uncl["cool"], dem_uncl["power"], param_uncl["T_air"], param_uncl["GHI"], param_uncl["DHI"], param_uncl["wind_speed"]]
+    time_series = [dem_uncl["heat"][0:adjustedHorizon], dem_uncl["cool"][0:adjustedHorizon], dem_uncl["power"][0:adjustedHorizon],
+                   param_uncl["T_air"][0:adjustedHorizon], param_uncl["GHI"][0:adjustedHorizon], param_uncl["DHI"][0:adjustedHorizon],
+                   param_uncl["wind_speed"][0:adjustedHorizon]]
+
     # Only building demands and weather data are clustered using k-medoids algorithm; secondary time series are clustered manually according to k-medoids result
     inputs = np.array(time_series)
     # Execute k-medoids algorithm
     print("Cluster design days...")
     start = time.time()
     (clustered_series, nc, y, z, inputsTransformed) = clustering.cluster(inputs,
-                                     param["n_clusters"],
-                                     len_cluster=24,#int(inputs.shape[1] / 365),
+                                     data.time["clusterNumber"],
+                                     len_cluster=int(clusterHorizon),
                                      norm = 2,
                                      mip_gap = 0.02,
                                      )
 
-    print("Design day clustering finished. (" + str(time.time()-start) + ")\n")
+    print("Design clustering finished. (" + str(time.time()-start) + ")\n")
 
     dem = {}
     dem["heat"] = clustered_series[0]
@@ -131,14 +137,14 @@ def load_params(data):
 
     # Get sigma-function: for each day of the year, find the corresponding design day
     # Get list of days which are used as design days
-    typedays = np.zeros(param["n_clusters"], dtype = np.int32)
+    typedays = np.zeros(data.time["clusterNumber"], dtype = np.int32)
     n = 0
-    for d in range(365):
+    for d in range(52):
         if any(z[d]):
             typedays[n] = d
             n += 1
     # Assign each day of the year to its design day
-    sigma = np.zeros(365, dtype = np.int32)
+    sigma = np.zeros(52, dtype = np.int32)
     for day in range(len(sigma)):
         d = np.where(z[:,day] == 1 )[0][0]
         sigma[day] = np.where(typedays == d)[0][0]
[…]
