### 2025-01-17 2031495
Add Load all data function
