### 2025-01-17 44ba510
Add possibility to change weather data
