+++ districtgenerator/non_residential_users.py
@@ -1,622 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Written by Felix Rehmann, 2024
-rehmann@tu-berlin.de
-"""
-
-import os, math
-import json 
-import random as rd
-import numpy as np
-from districtgenerator.profils import Profiles, NonResidentialProfiles
-from districtgenerator.solar import SunIlluminance
-import functions.heating_profile_5R1C as heating
-import functions.schedule_reader as schedules
-import functions.light_demand as light_demand
-from typing import Dict, List, Any, Optional
-import pandas as pd
-
-current_file_path = os.path.abspath(__file__)
-base_path = os.path.dirname(os.path.dirname(current_file_path))
-
-class NonResidentialUsers():
-    """
-    A class for modeling non-residential building usage profiles, including occupancy, 
-    electricity, and water heating demands based on the building's usage type and area. 
-    This class leverages TEASER project data structures and is designed to support 
-    energy analysis tasks.
-
-    Parameters
-    ----------
-    building_usage : str
-        The usage type of the building, which should match keys in occupancy and 
-        electricity data files.
-    area : float
-        Total floor area of the building in square meters.
-    file : str
-        Path to the building envelope data file.
-    envelope : object
-        An object representing the physical envelope of the building including 
-        attributes like windows, walls, and thermal properties.
-    site : dict
-        A dictionary containing site-specific data such as timezone, location, and 
-        local climate data.
-    time : dict
-        A dictionary with time resolution and time steps for simulation.
-    nb_of_days : int
-        Number of days for which the simulation is to be run.
-
-    Attributes
-    ----------
-    annual_appliance_demand : float
-        The calculated annual demand for appliances in kWh.
-    annual_lightning_demand : float
-        The calculated annual lighting demand in kWh.
-    nb_occ : list
-        List of number of occupants based on building usage type.
-    occ : np.array
-        Occupancy profile for the building.
-    dhw : np.array
-        Daily hot water usage profile for the building.
-    elec : np.array
-        Electrical demand profile for the building.
-    gains : np.array
-        Internal gains from occupants, appliances, and lighting.
-    heat : np.array
-        Heating demand profile for the building.
-    occupancy_schedule : np.array
-        Scheduled occupancy profile based on typical usage patterns.
-    appliance_schedule : np.array
-        Scheduled appliance usage profile.
-    lighntning_schedule : np.array
-        Scheduled lighting usage profile.
-    occupancy_data : dict
-        Loaded occupancy data from JSON.
-    electricity_data : dict
-        Loaded electricity data from JSON.
-
-    Methods
-    -------
-    load_json_data(json_path)
-        Loads data from a JSON file located at the specified path.
-    generate_number_occupants()
-        Calculates the number of occupants based on building usage and area.
-    generate_schedules()
-        Generates daily usage schedules for occupancy, appliances, and lighting.
-    generate_annual_el_consumption_equipment()
-        Calculates annual electricity consumption for equipment based on building type.
-    generate_annual_el_consumption_lightning()
-        Calculates annual electricity consumption for lighting.
-    generate_dhw_profile()
-        Generates a daily hot water usage profile based on occupancy and building type.
-    create_el_wrapper()
-        Creates an electricity usage profile wrapper for lighting and appliances.
-    calcProfiles(site, time_resolution, time_horizon, initital_day=1)
[…]
