+++ tests/test_mix_used.py
@@ -1,0 +1,122 @@
+import unittest
+import random as rd
+from random import sample
+import os
+import pandas as pd
+from districtgenerator.classes import datahandler
+from districtgenerator.functions import path_checks
+# Shoud create simulations for non-residential buildings
+# Write a test, that all functions run through without errors
+# Check if the results exist and are not zero
+
+
+
+PARENT_FOLDER_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+
+class TextMixUsed(unittest.TestCase):
+    def setUp(self):
+        # Create test scenario
+        self.test_data = { 
+            'id': [0, 1, 2, 3],
+            'building': [
+                "IWU Office, Administrative or Government Buildings", 
+                "IWU Research and University Teaching",
+                "SFH", "TH" ],
+            
+            'year': [rd.randint(1900, 2024) for _ in range(4)],
+            'retrofit': [0] * 4,
+            'area': [rd.randint(0, 100) for _ in range(4)],
+            'construction_type': [""] * 4,
+            'night_setback': [rd.randint(0, 1) for _ in range(4)],
+            'heater': ["HP"] * 4,
+            'PV': [rd.randint(0, 1) for _ in range(4)],
+            'STC': [rd.randint(0, 1) for _ in range(4)],
+            'EV': [rd.randint(0, 1) for _ in range(4)],
+            'BAT': [rd.randint(0, 1) for _ in range(4)],
+            'f_TES': [rd.randint(0, 1) for _ in range(4)],
+            'f_BAT': [rd.randint(0, 1) for _ in range(4)],
+            'f_EV': ["M"] * 4,
+            'f_PV': [rd.randint(0, 1) for _ in range(4)],
+            'f_STC': [rd.randint(0, 1) for _ in range(4)],
+            'gamma_PV': [rd.randint(0, 1) for _ in range(4)],  
+            'ev_charging': ["on_demand"] * 4
+        }
+
+        df = pd.DataFrame(self.test_data, columns=self.test_data.keys())
+        self.scenario_path = os.path.join(PARENT_FOLDER_PATH, 'districtgenerator', 'data', 'scenarios', 'test_scenario.csv')
+        df.to_csv(self.scenario_path, index=False, sep=';')
+        df.to_csv(self.scenario_path, index=False, sep=';')
+        print(df.head())
+        self.results_path = os.path.join(PARENT_FOLDER_PATH,  'test_scenario', 'demands')
+
+    def test_mix_used(self):
+        # Test all functions in the non_residential module
+        data = datahandler.Datahandler()
+        data.setWeatherFile(os.path.join(PARENT_FOLDER_PATH, 'districtgenerator', 'data', 'weather', 'EPW', 'DEU_BE_Berlin-Schonefeld.AP.103850_TMYx.2004-2018.epw'))
+        data.generateEnvironment(plz="52070")
+
+        # Use the created scenario for testing
+        data.initializeBuildings('test_scenario')
+        data.setResultPath('test_scenario')
+        data.generateBuildings()
+
+        data.generateDemands()
+
+        # Add assertions to check if the results exist and are not zero
+        self.assertTrue(os.path.exists(self.scenario_path), f"Scenario file not found: {self.scenario_path}")
+        self.assertGreater(os.path.getsize(self.scenario_path), 0, f"Scenario file is empty: {self.scenario_path}")
+        scenario_data = pd.read_csv(self.scenario_path, sep=';')
+        for file in os.listdir(self.results_path):
+            if file.endswith('.csv'):
+                file_path = os.path.join(self.results_path, file)
+                df = pd.read_csv(file_path, sep=',')
+                # Check if the required columns are present
[…]
