### 2025-01-21 7a9fa1b
Extend Load all data function
