### 2025-01-21 80b45a3
update_requirements_and_datahandler (#39)
* added missing packages

* improved __init__ function
