### 2025-01-27 RWTH-EBC/districtgenerator#42 @jtock
JOSS review - Paper
Issue connected to https://github.com/openjournals/joss-reviews/issues/7657
Referring to the reviewer checklist I would like to raise some issues about your paper.

_A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?_

You gave a good general introduction.
- [ ] Please also name the targeted / potential user groups and the specific problem they want to solve with the tool.

_State of the field: Do the authors describe how this software compares to other commonly-used packages?_

Please add some information e.g. in the statement of need chapter.
- [ ] What makes the software unique? What does it offer, what other tools can not do? E.g. compared to SimStadt.

_Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?_

I found some minor editing problems:
- [ ] Line 34-36 "directive" is twice in the sentence
- [ ] Line 57 -59 the caption is broken
- [ ] Line 60 the reference to the figure is broken
- [ ] Line 89-94 the acronyms PV, EV, GUI are not introduced
- [ ] Line 92 there is no url connected to EHDO (as done for the other tools)

_References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?_

- [ ] Line 29-33 you can probably find a shot title for that reference
- [ ] Line 37 - 39 I suggest giving a reference for those two sentences and maybe also change the formulation a bit that it "...can lead to..." / "... its expected to have..."
- [ ] Line 53 has a broken reference
- [ ] Line 94 has an unclear reference placement
- [ ] The dates of the DIN references are missing
