### 2025-01-27 RWTH-EBC/districtgenerator#43 @jtock
JOSS review - Documentation not findable
Issue connected to https://github.com/openjournals/joss-reviews/issues/7657

Since JOSS has different requirements for the documentation, I wanted to have a look at it before the installation. Unfortunately, I could not find a link to the documentation, but I could find various files.

Can you provide a link and add it to the README file?

Additional for master branch:
I am reviewing the JOSS-branch, but was also checking the master branch. In the README file of the master branch the documentation batch is leading to another tool (FiLiP). Also some links regarding the examples do not work as well as the image.
