+++ examples/e4_generate_buildings.py
@@ -16,7 +16,7 @@ def example4_generate_buildings():
     data = Datahandler()
 
     # Generate Environment for the District
-    data.generateEnvironment()
+    data.generateEnvironment(plz="52062")
 
     # Initialize Buildings to the District
     data.initializeBuildings(scenario_name="example")
