### 2025-01-29 RWTH-EBC/districtgenerator#45 @jtock
JOSS review - Running and testing examples
Issue connected to https://github.com/openjournals/joss-reviews/issues/7657

**Please describe the bug**
There are three problems:

There is no output when running the examples. I followed the installation guide and installed the JOSS_submisson branch via
`git clone -b JOSS_submission --single-branch https://github.com/RWTH-EBC/districtgenerator `
There is no data output when running e1. The following examples do not run either based on that. There is no error message.
(Same problem on Mac (Python 3.9.10 and 3.13.1) and JupyterHub (Python 3.11.10).)

Testing the examples:
JupyterHub: The tests are passing for e1 to e3. e4 and e5 fail.
Mac: All tests fail.

**What are the steps to reproduce the bug?**
1. Install JOSS_submission branch
2. Run on machines as defined below

**What behavior did you expect?**
1. Data output from e1 in the result folder (districtgenerator/districtgenerator/results)
2. Passing test.

**Can you add screenshots?**
Error message when running the test:
On Mac (for every test):
```
from examples import e1_initialize_datahandler as e1
ModuleNotFoundError: No module named 'examples'
```
On JupyterHub (same for e4 and e5):
```
`...EE
ERROR: test_e4_generate_buildings (__main__.TestExamples.test_e4_generate_buildings)
Tests the executability of example 4
Traceback (most recent call last):
  File "/home/jovyan/districtgenerator/tests/test_examples.py", line 43, in test_e4_generate_buildings
    data_e4 = e4.example4_generate_buildings()
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/jovyan/districtgenerator/examples/e4_generate_buildings.py", line 29, in example4_generate_buildings
    data.generateBuildings()
  File "/home/jovyan/districtgenerator/districtgenerator/classes/datahandler.py", line 186, in generateBuildings
    prj = Project(load_data=True)
          ^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/conda/lib/python3.11/site-packages/teaser/project.py", line 92, in __init__
    raise ValueError("This option was deprecated")
ValueError: This option was deprecated`
```

**What environment are you using?**
Mac:
Apple M2 chip
Ventura 13.5.1
Running in VS Code and global terminal
Python 3.9.10 and Python 3.13.1

JupyterHub:
Python 3.11.10 via JupyterLab version 4.2.5
