### 2025-01-29 RWTH-EBC/districtgenerator#43 @JoelSchoelzel
Dear @jtock, you will find the documentation in folder path ...\districtgenerator\docs\_build\html of the branch JOSS_. Use index.html to open the documentation in your browser.
For the submission to JOSS, we have only added the final documentation to this branch for now.
Nevertheless, thank you for your comment on the master branch. We have already added many new functions to the district generator there. We probably made a mistake when merging. We will check the error!
