### 2025-01-29 RWTH-EBC/districtgenerator#45 @JoelSchoelzel
Dear @jtock, in the first four examples, a neighbourhood property is created and gradually expanded. If you execute the examples individually, you should see this object displayed in the output console.
So far we have decided against saving the object because we see pickle as the only possibility. However, there is the problem that the pickle file can only be loaded on the computer on which it was saved.

Which teaser version are you using?
