### 2025-01-29 RWTH-EBC/districtgenerator#45 @jtock
In both cases there is also no console / terminal output at all. Neither an error message nor a result.

Teaser 0.7.7 is installed.
