+++ districtgenerator/classes/KPIs.py
@@ -5,6 +5,14 @@ import numpy as np
 import os
 import json
 import math
+import reportlab
+from reportlab.pdfgen import canvas
+from reportlab.lib import colors
+from reportlab.graphics.shapes import *
+from reportlab.graphics.charts.piecharts import Pie
+from datetime import datetime
+from reportlab.lib.styles import ParagraphStyle
+from reportlab.platypus import Paragraph
 
 class KPIs:
 
@@ -39,6 +47,9 @@ class KPIs:
         self.dcf_year = None
         self.scf_year = None
         self.annual_investment_total = None
+        self.totalarea = None
+        self.totalnumberflats = None
+        self.totalnumberocc = None
 
         # initialize input data for calculation of KPIs
         inputData = {}
@@ -461,6 +472,25 @@ class KPIs:
         #    self.EnergyAutonomy_year += EnergyAutonomy[c] * (self.inputData["clusterWeights"][self.inputData["clusters"][c]]
         #                                                  / sum_ClusterWeights)
 
+    def calcareas(self, data):
+        """
+        Returns
+        -------
+        None.
+        """
+        total_net_leased_area = 0
+        total_number_flats = 0
+        total_number_occ = 0
+
+        for building in data.district:
+            total_net_leased_area += building["buildingFeatures"]["area"]
+            total_number_flats += building["user"].nb_flats
+            total_number_occ += building["user"].nb_occ
+
+        self.totalarea = total_net_leased_area
+        self.totalnumberflats = total_number_flats
+        self.totalnumberocc = total_number_occ
+
     def calculateAllKPIs(self, data):
         """
         Calculate all KPIs.
@@ -481,3 +511,647 @@ class KPIs:
         self.calculateCO2emissions(data)
         self.calculateAutonomy()
         self.calc_annual_cost_total(data.scenario, data.decentral_device_data, data.district, data.physics)
+
+    def create_certificate(self, data, result_path):
+        """
+        Generate a certificate as PDF file with a list of KPIs and a list with building information.
+
+        Parameters:
+        - filename: The name of the PDF file to create.
+        - title: The title of the document.
+        - kpis: A list of strings, where each string is a KPI to be written in the document.
+        """
+
+        # preprocessing buildinglist
+        template_dict = {"Anzahl": 0,
+               "Gesamtfläche": 0,
+               "vor 1968": 0,
+               "1968-1978": 0,
+               "1979-1983": 0,
+               "1984-1994": 0,
+               "1995-2001": 0,
+               "2002-2009": 0,
+               "2010-2015": 0,
+               "ab 2016": 0,
+               }
+        SFH = dict(template_dict)
+        TH = dict(template_dict)
+        MFH = dict(template_dict)
+        AB = dict(template_dict)
+
+        for building in data.district:
+            if building["buildingFeatures"]["building"] == 'SFH':
+                SFH["Anzahl"] += 1
+                SFH["Gesamtfläche"] += building["buildingFeatures"]["area"]
+                if building["buildingFeatures"]["year"] < 1968:
+                    SFH["vor 1968"] += building["buildingFeatures"]["area"]
+                elif building["buildingFeatures"]["year"] >= 1968 and building["buildingFeatures"]["year"] <= 1978 :
+                    SFH["1968-1978"] += building["buildingFeatures"]["area"]
+                elif building["buildingFeatures"]["year"] >= 1979 and building["buildingFeatures"]["year"] <= 1983 :
+                    SFH["1979-1983"] += building["buildingFeatures"]["area"]
+                elif building["buildingFeatures"]["year"] >= 1984 and building["buildingFeatures"]["year"] <= 1994 :
[…]
