### 2025-01-29 29ce259
debug setup.py and correct gitignore, solving #45
