### 2025-01-29 RWTH-EBC/districtgenerator#45 @TBeckhoelter
Hi @jtock , I could reproduce the bug. Unfortunately, the package requirements were only adjusted in the requirements.txt file and not in setup.py. As a result, an incorrect version of teaser was installed, which caused the error. Can you please pull the repo and install it again and test if it works now? The tests and examples ran successfully on my system.
