+++ districtgenerator/classes/datahandler.py
@@ -155,7 +155,7 @@ class Datahandler:
         with open(os.path.join(self.filePath, 'central_device_data.json')) as json_file:
             self.central_device_data = json.load(json_file)
 
-    def select_plz_data(self, plz):
+    def select_plz_data(self):
         """
         Select the closest TRY weather station for the location of the postal code.
 
@@ -176,9 +176,9 @@ class Datahandler:
             sheet = workbook.active
 
             for row in sheet.iter_rows(values_only=True):
-                if plz == str(row[0]):
+                if self.site["zip"] == str(row[0]):
                     weatherdatafile = row[3]
-                    weatherdatafile_location = weatherdatafile[8:-9]
+                    self.site["Location"] = weatherdatafile[8:-9]
                     break
             else:
                 # If postal code cannot be found: Message and select weather data file from Aachen
@@ -188,7 +188,8 @@ class Datahandler:
         except Exception as e:
             # If postal code cannot be found: Message and select weathter data file from Aachen
             print("Postal code cannot be found, location changed to Aachen")
-            weatherdatafile_location = 507755060854
+            self.site["zip"] = "52064"
+            self.site["Location"] = 507755060854
             """  
                 Add new weatherdatafile_location, if you want an individual location: 
                 Files can be found here: https://www.dwd.de/DE/leistungen/testreferenzjahre/testreferenzjahre.html 
@@ -197,8 +198,6 @@ class Datahandler:
                 Uncomment the following line  
             """
             # weatherdatafile_location = 507755060854
-
-        return weatherdatafile_location
 
     def generateEnvironment(self, plz):
         """
@@ -214,12 +213,15 @@ class Datahandler:
         elif self.site["TRYYear"] == "TRY2045":
             first_row = 37
 
+
+        self.site["zip"] = plz
+        self.select_plz_data()
         # load weather data
         # select the correct file depending on the TRY weather station location
         weatherData = np.loadtxt(os.path.join(self.filePath, "weather", "TRY_" + self.site["TRYYear"][-4:] + "_" + self.site["TRYType"] + "er")
             + "\\"
             + self.site["TRYYear"] + "_"
-            + str(self.select_plz_data(plz)) + "_" + str(self.site["TRYType"])
+            + str(self.site["Location"]) + "_" + str(self.site["TRYType"])
             + ".dat",
             skiprows=first_row - 1)
 
@@ -278,7 +280,7 @@ class Datahandler:
         site_data = pd.read_csv(filePath, delimiter='\t', dtype={'Zip': str})
 
         # Filter data for the specific zip code
-        filtered_data = site_data[site_data['Zip'] == plz]
+        filtered_data = site_data[site_data['Zip'] == self.site["zip"]]
 
         # extract the needed values
         self.site["altitude"] = filtered_data.iloc[0]['Altitude']
