### 2025-02-04 67d27c9
Enter zip code in site_data.json; fixes #47
