+++ examples/e5_generate_demands.py
@@ -19,7 +19,7 @@ def example5_generate_demands():
     data = Datahandler()
 
     # Generate Environment for the District
-    data.generateEnvironment(plz="52062")
+    data.generateEnvironment()
 
     # Initialize Buildings to the District
     data.initializeBuildings(scenario_name="example")
