### 2025-02-05 9e82866
47 use changed zip code for environment generation (#50)
* Save plz in self.site dictionary and change zip for environment

* Enter zip code in site_data.json; fixes #47

---------

Co-authored-by: Carla Wüller <carla.wueller@eonerc.rwth-aachen.de>
