### 2025-02-05 7d55bae
added scenario_path attribute to datahandler (#49)
