### 2025-02-05 92bf1fb
bugfix installation setup.py (#54)
