### 2025-02-05 1de41ef
added possibility to save plots in different path
