### 2025-02-05 8071b1c
added possibility to save plots in different path
