### 2025-02-05 ddcf19e
added attribute pass through to DemandPlots initialization
