+++ districtgenerator/classes/datahandler.py
@@ -216,7 +216,7 @@ class Datahandler:
         # load weather data
         # select the correct file depending on the TRY weather station location
         weatherData = np.loadtxt(os.path.join(self.filePath, "weather", "TRY_" + self.site["TRYYear"][-4:] + "_" + self.site["TRYType"] + "er")
-            + "\\"
+            + "/"
             + self.site["TRYYear"] + "_"
             + str(self.site["Location"]) + "_" + str(self.site["TRYType"])
             + ".dat",
@@ -841,7 +841,7 @@ class Datahandler:
         """
 
         # initialize plots and prepare data for plotting
-        demandPlots = DemandPlots()
+        demandPlots = DemandPlots(resultPath=self.resultPath)
         demandPlots.preparePlots(self)
 
         # check which resolution for plots is used
