### 2025-02-05 f839205
added attribute pass through to DemandPlots initialization
