### 2025-02-05 636d497
51 change plot result path (#55)
* added possibility to save plots in different path

* added possibility to save plots in different path

* added attribute pass through to DemandPlots initialization

* changed pathing problem

* added attribute pass through to DemandPlots initialization
