+++ districtgenerator/functions/load_params_central_devices.py
@@ -19,7 +19,6 @@ import math
 import districtgenerator.functions.clustering_medoid as clustering
 import time
 import os
-import json
 import sys
 import copy
 
@@ -29,6 +28,7 @@ def load_params(data):
     # import model parameters
     central_device_data = copy.deepcopy(data.central_device_data)
     param = copy.deepcopy(data.params_ehdo_model)
+    ecoData = copy.deepcopy(data.ecoData)
     param_uncl = {}  # unclustered time series for weather data
 
     ################################################################
@@ -181,13 +181,13 @@ def load_params(data):
 
     # Photovoltaics
     devs["PV"] = {
-        "feasible": all_models["Photovoltaic"]["enabled"],
-        "eta": all_models["Photovoltaic"]["eta"] / 100,
-        "life_time": all_models["Photovoltaic"]["life_time"],
-        "inv_var": all_models["Photovoltaic"]["inv_var"],
-        "cost_om": all_models["Photovoltaic"]["cost_om"] / 100,
-        "max_area": all_models["Photovoltaic"]["max_area"],
-        "min_area": all_models["Photovoltaic"]["min_area"],
+        "feasible": all_models["PV"]["enabled"],
+        "eta": all_models["PV"]["eta"] / 100,
+        "life_time": all_models["PV"]["life_time"],
+        "inv_var": all_models["PV"]["inv_var"],
+        "cost_om": all_models["PV"]["cost_om"] / 100,
+        "max_area": all_models["PV"]["max_area"],
+        "min_area": all_models["PV"]["min_area"],
         # For correlation between area and peak power:
         "G_stc": 1,  # kW/m^2,  solar radiation under standard test conditions (STC)
     }
@@ -200,38 +200,38 @@ def load_params(data):
 
     # Wind turbine
     devs["WT"] = {
-        "feasible": all_models["WindTurbine"]["enabled"],
-        "inv_var": all_models["WindTurbine"]["inv_var"],
-        "life_time": all_models["WindTurbine"]["life_time"],
-        "cost_om": all_models["WindTurbine"]["cost_om"] / 100,
-        "min_cap": all_models["WindTurbine"]["min_cap"],
-        "max_cap": all_models["WindTurbine"]["max_cap"],
-        "h_coeff": all_models["WindTurbine"]["h_coeff"],  # hellmann_coeff
-        "hub_h": all_models["WindTurbine"]["hub_h"],
-        "ref_h": all_models["WindTurbine"]["ref_h"],
+        "feasible": all_models["WT"]["enabled"],
+        "inv_var": all_models["WT"]["inv_var"],
+        "life_time": all_models["WT"]["life_time"],
+        "cost_om": all_models["WT"]["cost_om"] / 100,
+        "min_cap": all_models["WT"]["min_cap"],
+        "max_cap": all_models["WT"]["max_cap"],
+        "h_coeff": all_models["WT"]["h_coeff"],  # hellmann_coeff
+        "hub_h": all_models["WT"]["hub_h"],
+        "ref_h": all_models["WT"]["ref_h"],
     }
     devs["WT"]["norm_power"] = calc_WT_power(devs, param, data.time["clusterNumber"])  # relative power between 0 and 1
 
     # Hydropower
     devs["WAT"] = {
-        "feasible": all_models["Hydropower"]["enabled"],
-        "inv_var": all_models["Hydropower"]["inv_var"],
-        "life_time": all_models["Hydropower"]["life_time"],
-        "cost_om": all_models["Hydropower"]["cost_om"] / 100,
[…]
