+++ districtgenerator/functions/opti_central.py
@@ -16,7 +16,7 @@ def run_opti_central(model, data, cluster):
     siteData = data.site
     param_dec_devs = data.decentral_device_data
     model_param_eh = data.params_ehdo_model
-    technical_param_eh = data.params_ehdo_technical
+    central_device_data = data.central_device_data
     buildingData = data.district
     energyHubData = data.centralDevices
 
@@ -76,8 +76,12 @@ def run_opti_central(model, data, cluster):
     except KeyError:
         for n in range(nb):
             buildingData[n]["capacities"] = {}
-            for dev in ["HP", "EH", "CHP", "FC", "BOI", "STC", "BAT", "TES", "EV"]:
+            buildingData[n]["capacities"]["PV"] = {}
+            buildingData[n]["capacities"]["STC"] = {}
+            for dev in ["HP", "EH", "CHP", "FC", "BOI", "BAT", "TES", "EV"]:
                 buildingData[n]["capacities"][dev] = 0
+            buildingData[n]["capacities"]["PV"]["area"] = 0
+            buildingData[n]["capacities"]["STC"]["area"] = 0
 
     # %% TECHNICAL PARAMETERS
     soc_nom = {}
@@ -369,38 +373,38 @@ def run_opti_central(model, data, cluster):
     for t in time_steps:
         # Electric heat pump
         # todo: variable COP
-        model.addConstr(eh_heat["HP"][t] == eh_power["HP"][t] * technical_param_eh["HP"]["COP"])
+        model.addConstr(eh_heat["HP"][t] == eh_power["HP"][t] * central_device_data["HP"]["COP_const"])
         # Electric boiler
-        model.addConstr(eh_heat["EB"][t] == eh_power["EB"][t] * technical_param_eh["EB"]["eta_th"])
+        model.addConstr(eh_heat["EB"][t] == eh_power["EB"][t] * central_device_data["EB"]["eta_th"])
         # Compression chiller
-        model.addConstr(eh_cool["CC"][t] == eh_power["CC"][t] * technical_param_eh["CC"]["COP"])
+        model.addConstr(eh_cool["CC"][t] == eh_power["CC"][t] * central_device_data["CC"]["COP"])
         # Absorption chiller
-        model.addConstr(eh_cool["AC"][t] == eh_heat["AC"][t] * technical_param_eh["AC"]["eta_th"])
+        model.addConstr(eh_cool["AC"][t] == eh_heat["AC"][t] * central_device_data["AC"]["eta_th"])
         # Gas CHP
-        model.addConstr(eh_power["CHP"][t] == eh_gas["CHP"][t] * technical_param_eh["CHP"]["eta_el"])
-        model.addConstr(eh_heat["CHP"][t] == eh_gas["CHP"][t] * technical_param_eh["CHP"]["eta_th"])
+        model.addConstr(eh_power["CHP"][t] == eh_gas["CHP"][t] * central_device_data["CHP"]["eta_el"])
+        model.addConstr(eh_heat["CHP"][t] == eh_gas["CHP"][t] * central_device_data["CHP"]["eta_th"])
         # Gas boiler
-        model.addConstr(eh_heat["BOI"][t] == eh_gas["BOI"][t] * technical_param_eh["BOI"]["eta_th"])
+        model.addConstr(eh_heat["BOI"][t] == eh_gas["BOI"][t] * central_device_data["BOI"]["eta_th"])
         # Gas heat pump
-        model.addConstr(eh_heat["GHP"][t] == eh_gas["GHP"][t] * technical_param_eh["GHP"]["COP"])
[…]
