### 2025-02-11 942ea46
#56 Fix occupant assignment edge case by adding fallback condition
