### 2025-02-12 RWTH-EBC/districtgenerator#58 @snjsomnath
JOSS review - Tests failing
This is related to the on going JOSS review [here](https://github.com/openjournals/joss-reviews/issues/7657#issuecomment-2602726379)

**Please describe the bug**
The tests provided in the repository fail.

**What are the steps to reproduce the bug?**
1. git clone -b JOSS_submission https://github.com/RWTH-EBC/districtgenerator
2. created a .venv using requirements.txt
3. run test_examples.py in terminal
4. See error

**What behavior did you expect?**
Expect to see 5 tests run and return OK

**Can you add screenshots?**

```
ERROR: test_e5_generate_demands (__main__.TestExamples)
Tests the executability of example 5
Traceback (most recent call last):
  File "d:\GitHub\districtgenerator\tests\test_examples.py", line 48, in test_e5_generate_demands
    from examples import e5_generate_demands as e5
ModuleNotFoundError: No module named 'examples'

Ran 5 tests in 0.006s

FAILED (errors=5)

```

**What environment are you using?**
Python 3.10.9
venv built on requirements.txt

**Additional context**
This works fine running in Jupyter notebooks

Ran 5 tests in 72.487s

OK
<unittest.main.TestProgram at 0x16a40d1eb00>
