### 2025-02-12 RWTH-EBC/districtgenerator#59 @snjsomnath
JOSS review - Examples
This is related to the ongoing JOSS review [here](https://github.com/openjournals/joss-reviews/issues/7657)

While the authors have provided a minimum working example demonstrating the calculation of annual demands using six examples, example-six could be better formulated. Currently, it requires users to manually modify input files to change parameters—otherwise, it remains identical to Example 5.

To enhance usability and clarity, I suggest one of the following improvements:

- The "site_data.json" and "design_building_data.json" files were consistently referred to throughout all six examples.
or
- Have another scenario for example-6 that the user could run directly, just as in the previous examples.

JOSS guidelines state that examples should "ideally solve real-world analysis problems." Making Example 6 more explicit would better align with this requirement, and incorporating additional real-world examples would further strengthen the submission.
