### 2025-02-12 RWTH-EBC/districtgenerator#60 @snjsomnath
JOSS review - Documentation
This is related to the ongoing JOSS review [here](https://github.com/openjournals/joss-reviews/issues/7657)

While the documentation provided at `\docs\_build\html\index.html` is thorough, it must be more intuitive to find and accessible to the user. Please update the instructions to access the documentation directly in the README until a long-term solution is applied.
