### 2025-02-17 RWTH-EBC/districtgenerator#43 @jtock
Thank you!

One of the missions of JOSS is to make sure the software can easily be used as Open Source. Therefore it is necessary that the core information about the software is quick and easy accessible for potential users.
Your README is a good start for that.
The documentation should provide more detailed information. Therefore I strongly suggest to publish the documentation online  (e.g. via readthedocs) and refer to it in your README.

Since you already have a Sphinx documentation the effort to publish it should be relatively low and worth it.
