### 2025-02-17 RWTH-EBC/districtgenerator#62 @jtock
JOSS Review - Improve documentation
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

Next the to the accessibility of the documentation there are some areas that can be improved.

- [x] The workflow of the tool should be added to the documentation in addition to the existing structure tree. This would help a lot to understand the software and the examples.
- [x] The input files and their input variables should be explained. If there are discrete input values like the building type (sfh, mfh, ...) the user should have a lot of possible and allowed inputs. It would be ok to refer to other documentation (like Tabular), but should be done more detailed like in _Minimum required input data_ in the README.
- [x] Also the climate zones information given in design_weather_data.json would be better placed in the documentation.
- [x] A short description of the result files and their directory should also be added.
