### 2025-02-17 RWTH-EBC/districtgenerator#63 @jtock
JOSS review - Example design
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

The examples help to understand the functionality of the software, but they can be further improved since it is not completely intuitive.

- [x] When I run the examples I get no visible output. It seems that the functions are not working. You can improve the description of the examples or add an exemplary output.
- [x] Example e2 refers to "open data.site". This can be improved by printing data.site and referring to the file in which the information is provided by the user (in this case: data/site_data.json). The same improvement can be made in some other examples. Please check for that.
- [x] In e6, the user is supposed to open design_weather_data.json which is placed in another directory as given in the example.

A more general improvements:
- [ ] A general description of the workflow of districtgeneration would greatly improve understanding. The functionality remains unclear until you finish e6. See #62.

I agree to #59.
