### 2025-02-18 RWTH-EBC/districtgenerator#64 @jtock
JOSS review - Testing
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

The tests are covering the examples e1 to e5. They mainly check whether example scripts execute, but it does not validate output correctness.

Testing should be used, especially when implementing several sub-packages and databases, to ensure the following steps:
- testing dependency compatibility
- compare expected vs. actual results
- ideally edge cases to cover unexpected behavior like wrong input formats

It is up to you how you implement the tests, e.g. by testing units / functions or the output. In any case the test need to ensure that the main functionality works as expected.

I recommend using an automated test suite e.g. in GitHub to validate changes in code and changing dependencies.
