### 2025-02-18 RWTH-EBC/districtgenerator#65 @jtock
JOSS review - Contribution and authorship
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

The list of authors in the README and documentation, the authors in the JOSS paper and the list of contributors are differing. The references / papers listed in the README also have different authors.

- [ ] Please update the list of authors in README and documentation.
- [ ] The authorships in the JOSS paper are dated Juni 2023 (see: [paper.md](https://github.com/RWTH-EBC/districtgenerator/blob/JOSS_submission/joss/paper.md?plain=1)). Please check if the list is complete.

Note: Somehow the list of contributors in the GitHub is broken. Perhaps you can fix it to further improve the acknowledgement of the contributor and transparency.
