### 2025-02-18 RWTH-EBC/districtgenerator#66 @jtock
JOSS review - Code documentation (docstrings and comments)
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

Some function currently lack proper docstrings and inline comments, making the code harder to understand and maintain.

- [x] Ensure every function and class has a proper docstring explaining:
  - Purpose (What does it do?)
  - Parameters (What inputs does it take? What is behind this parameter?)
  - Returns (What output does it provide?)
- [x] Use a standardized docstring format
- [x] Make sure all comments are in english
- [x] Improve inline comments within functions where logic may be unclear

Tip: AI can help you create docstrings and comments. It can dramatically reduce the time needed.
