### 2025-02-18 RWTH-EBC/districtgenerator#67 @jtock
JOSS review - Assumption handling and documentation
Issue related to https://github.com/openjournals/joss-reviews/issues/7657

The code contains hard-coded assumptions that may limit transferability between different buildings or districts. Some of these assumptions are not documented or configurable, making it difficult for users to adapt them to different contexts.

Here are some examples where the building / district layout is heavily influenced by fixed assumptions.
Fixed flat area without any reference for this assumption:
```
    def generate_number_flats(self,area):
        '''
        Generate number of flats for different of building types.

        Parameters
        ----------
        area : integer
            Floor area of different building types

        '''
        if self.building == "SFH":
            self.nb_flats = 1
        elif self.building == "TH":
            self.nb_flats = 1
        elif self.building == "MFH":
            if area <= 4*100:
                self.nb_flats = 4
            elif area > 4 * 100:
                self.nb_flats = math.floor(area/100)
        elif self.building == "AB":
            if area <= 10*100:
                self.nb_flats = 10
            elif area > 10*100:
                self.nb_flats = math.floor(area/100)
```
Number of floors and floor height:
```
            # add buildings to TEASER project
            prj.add_residential(
                method='tabula_de',
                usage=building_type,
                name="ResidentialBuildingTabula",
                year_of_construction=building["buildingFeatures"]["year"],
                number_of_floors=3,
                height_of_floors=3.125,
                net_leased_area=building["buildingFeatures"]["area"],
                construction_type=retrofit_level)
```

Other assumptions may have a better transferability, but are not documented properly.
Another examples is the annual electricity consumption:
```
        # source: https://www.stromspiegel.de/stromverbrauch-verstehen/stromverbrauch-im-haushalt/#c120951
        # method: https://www.stromspiegel.de/ueber-uns-partner/methodik-des-stromspiegels/
        standard_consumption = {"SFH" : {1 : 2300,
                                         2 : 3000,
                                         3 : 3500,
                                         4 : 4000,
                                         5 : 5000},
                                "MFH" : {1 : 1300,
                                         2 : 2000,
                                         3 : 2500,
                                         4 : 2600,
                                         5 : 3000}}
```

- [x] Identify and list all hardcoded values in the districtgenertor (e.g., building properties, climate conditions, material constants).
- [x] Determine which values should be configurable and which are reasonable defaults. The transferability between different districts should be the main evaluation criterion.
- [x] Please also specify under which conditions the assumptions hold (e.g., climate zone, building age, renovation state, geographical area).
- [x] Give (scientific) references for each assumption, if possible.
- [x] Document assumptions in a structured and standardized way. I recommend adding a chapter to the documentation. It needs to be clear for a user of the software that there are assumptions, in which functionalities they are and where to find them.
