### 2025-02-19 415987a
Bugfix opti_central bat
