### 2025-02-20 #7657 @JoelSchoelzel
Dear @jtock , thank you very much for the hints in the issues. These help us to further improve the use and documentation! We will definitely implement these points, but in the best case scenario we would wait for both final reviews in order to divide up the tasks within the team and complete them as efficiently as possible. Can you continue and finalize the review or are there issues that need to be dealt before?

@AdamRJensen Is this procedure in line with the JOSS review process or is an ongoing implementation of the review hints desired?
