### 2025-02-20 #7657 @AdamRJensen
> @AdamRJensen Is this procedure in line with the JOSS review process or is an ongoing implementation of the review hints desired?

The review process is preferably iterative.
