### 2025-02-20 fcfca92
Add new mixed building
