+++ classes/datahandler.py
@@ -9,6 +9,7 @@ import datetime
 import numpy as np
 import openpyxl
 import pandas as pd
+import random as rd
 from itertools import count
 from teaser.project import Project
 from classes.envelope import Envelope
@@ -237,7 +238,7 @@ class Datahandler:
         num_OB = 0
         num_schools = 0
         num_grocery_store = 0
-
+        num_mfh_grocery_store = 0
 
         self.scenario_name = scenario_name
         self.scenario = pd.read_csv(os.path.join(self.filePath, 'scenarios')
@@ -245,33 +246,111 @@ class Datahandler:
                                     + self.scenario_name + ".csv",
                                     header=0, delimiter=";")
 
-        # initialize buildings for scenario
-        # loop over all buildings
-        for id in self.scenario["id"]:
-            # create empty dict for observed building
-            building = {}
+        # Prepare a new list to collect updated scenario rows
+        new_scenario = []
+        next_id = 0
 
-            # store features of the observed building
-            building["buildingFeatures"] = self.scenario.loc[id]
+        # Loop over each row in the scenario DataFrame.
+        for idx, row in self.scenario.iterrows():
+            # Make a copy of the row to work on.
+            building_features = row.copy()
 
-            # append building to district
-            self.district.append(building)
+            # Check if the building type is a combined type (MFH+Grocery_store or AB+Grocery_store)
+            if row["building"] in ['MFH+Grocery_store', 'AB+Grocery_store']:
+                # Make a copy of the building features for the MFH part
+                building_mfh = building_features.copy()
+                # Assign a new unique id using the original id with a .1 suffix.
+                building_mfh["id"] = next_id
+                # Set the building type to MFH.
+                building_mfh["building"] = "MFH"
+                # Reset the construction type so that it matches with teaser.
+                building_mfh["construction_type"] = ""
+
+                # Determine the area of the MFH portion.
+                # Assumption: The grocery store occupies the area of 3 flats,
+                # and a typical flat has an area of about 80 m^2.
+                total_area = building_mfh["area"]
+                # Determine the number of flats.
+                if total_area <= 4 * 100:
+                    # Minimum case: At least 3 flats if the area is small
+                    nb_flats = 3
+                elif total_area > 4 * 100:
+                    # For larger areas, pick a random number around total_area/80,
+                    # then subtract 3 (the grocery store's assumed share).
+                    nb_flats = rd.randint((total_area // 80) - 1, (total_area // 80) + 1) - 3
+                # Calculate the area per flat (total area divided by the total number of flats including the grocery store).
+                area_flat = total_area / (nb_flats + 3)
+                # Set the MFH building area to be the area occupied by its flats.
+                building_mfh["area"] = int(nb_flats * area_flat)
+
+
+                # --- Create the Grocery Store building entry ---
+                building_grocery = building_features.copy()
+                # Assign a new unique id using the original id with a .2 suffix.
+                building_grocery["id"] = next_id + 1  # assign the next id
+                # Set the building type to Grocery_store
+                building_grocery["building"] = "Grocery_store"
+                # Set the area for the grocery store as the remaining area.
+                building_grocery["area"] = int(total_area - building_mfh["area"])
+
+                # Set various technical attributes for the Grocery_store to "0"
+                # since the roof is assumed to be exclusively used by the MFH
+                building_grocery["PV"] = 0
+                building_grocery["STC"] = 0
+                building_grocery["BAT"] = 0
+                building_grocery["f_BAT"] = 0
[…]
