### 2025-02-20 ec563a4
#68 Integration of OpenDHW as an installable package
