+++ districtgenerator/functions/dhw_stochastical.py
@@ -1,233 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Script to generate domestic hot water demands.
-This script is a copy of dhw_stochastical.py from pyCity.
-https://github.com/RWTH-EBC/pyCity
-"""
-
-from __future__ import division
-
-import os
-import numpy as np
-import math
-import random
-import districtgenerator.functions.change_resolution as chres
-import pylightxl as xl
-
-def load_profiles(filename):
-    """
-    """
-    # Initialization
-    profiles = {"we": {}, "wd": {}}
-    #book = xlrd.open_workbook(filename)
-    book = xl.readxl(fn=filename)
-    sheetnames = book.ws_names
-
-    
-    # Iterate over all sheets    
-    for sheetname in sheetnames:
-        #sheet = xl.readxl(fn=filename, ws=sheetname)
-        
-        # Read values
-        values = [book.ws(ws = sheetname).index(row=i, col=1) for i in range(1, 1441)] #[sheet.cell_value(i,0) for i in range(1440)]
-
-        # Store values in dictionary
-        if sheetname in ("wd_mw", "we_mw"):
-            profiles[sheetname] = np.array(values)
-        elif sheetname[1] == "e":
-            profiles["we"][int(sheetname[2])] = np.array(values)
-        else:
-            profiles["wd"][int(sheetname[2])] = np.array(values)
-    
-    # Return results
-    return profiles
-
-
-def compute_daily_demand(probability_profiles, average_profile, occupancy,
-                         current_day, temperature_difference=35):
-    """
-    Parameters
-    ----------
-    probability_profiles : array-like
-        Minute-wise sampled probability distribution.
-        "Haushaltewd" and "Haushaltewe" in Lion's thesis.
-        This input should also be equivalent to "pwd" and "pwe", because only 
-        one household is taken into account.
-    average_profile : array-like
-        Minute-wise sampled average tap water profiles (in liters per hour).
-        "mwwd" and "mwwe" in Lion's thesis.
-    occupancy : array-like
-        10-Minute-wise sampled occupancy of the considered building/apartment.
-    current_day : integer
-        Current day of the year (January 1st -> 0, February 1st -> 31, ...)
-    temperature_difference : float
-        How much does the tap water has to be heated up? Either enter a float
-        or an array with the same dimension as probability_profiles.
-    
-    Returns
-    -------
-    water : array-like
-        Tap water volume flow in liters per hour.
-    heat : array-like
-        Resulting minute-wise sampled heat demand in Watt.
-        The heat capacity of water is assumed to be 4180 J/(kg.K) and the
-        density is assumed to be 980 kg/m3
-    """
-    # Initialization
-    water = []
-    timesteps = 1440
-    time = np.arange(timesteps)
-    
-    # Compute seasonal factor
-    # Introduce abbreviation to stay below 80 characters per line
-    arg = math.pi * (2 / 365 * (current_day + time / timesteps) - 1 / 4)
-    probability_season = 1 + 0.1 * np.cos(arg)
-    
-    # Iterate over all time steps
-    for t in time:
-        # Compute the product of occupancy and probability_profiles
-        current_occupancy = occupancy[int(t/10)]
-        if current_occupancy > 0:
-            probability_profile = probability_profiles[current_occupancy][t]
-        else:
-            probability_profile = 0
-    
-        # Compute probability for tap water demand at time t
-        probability = probability_profile * probability_season[t]
-
-        # Check if tap water demand occurs at time t
-        if random.random() < probability:
-            # Compute amount of tap water consumption. This consumption has 
-            # to be positive!
[…]
