+++ districtgenerator/functions/OpenDHW.py
@@ -1,1609 +1,0 @@
-# -*- coding: utf-8 -*-
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
-from pathlib import Path
-import math
-import statistics
-import random
-import scipy
-from scipy.stats import beta
-import matplotlib.dates as mdates
-
-"""
-This is the script that stores all function of the DHWcalc package.
-It is not meant to be executed on its own, but rather a toolbox for building
-small scripts. Examples are given in OpenDHW/Examples.
-
-OpenDHW is mostly built on Pandas, a good introduction is given here:
-https://pandas.pydata.org/pandas-docs/stable/user_guide/10min.html
-https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html
-
-OpenDHW_Utilities stores a few other functions that do not generate DHW 
-Timeseries directly, like the StorageLoad Function.
-"""
-
-# RWTH colours
-rwth_blue = "#00549F"
-rwth_red = "#CC071E"
-# sns.set_style("white")
-sns.set_context("paper")
-
-# --- Constants ---
-rho = 980 / 1000  # kg/L for Water (at 60°C? at 10°C its = 1)
-cp = 4180  # J/kgK
-
-
-def import_from_dhwcalc(s_step, daylight_saving, categories,
-                        mean_drawoff_vol_per_day=200, max_flowrate=1200):
-    """
-    DHWcalc yields Volume Flow TimeSeries (in Liters per hour).
-
-    :param  s_step:                     int:    resolution of file in seconds
-    :param  categories:                 int:    either '1' or '4'
-    :param  mean_drawoff_vol_per_day:   int:    daily water demand in Liters
-    :param  daylight_saving:            Bool:   apply daylight saving or not
-    :param  max_flowrate:               int:    maximum water flowrate in L/h
-
-    :return timeseries_df:              df:     dataframe that holds the data
-    """
-
-    if daylight_saving:
-        ds_string = 'ds'
-    else:
-        ds_string = 'nods'
-
-    # --- DHWcalc result files, saved in the OpenDHW Package
-    dhw_file = "{vol}L_{s_step}min_{cats}cat_sf_{ds}_max{max_flow}.txt".format(
-        vol=mean_drawoff_vol_per_day,
-        s_step=int(s_step / 60),
-        cats=categories,
-        ds=ds_string,
-        max_flow=max_flowrate,
-    )
-
-    dhw_profile = Path.cwd().parent / "DHWcalc_Files" / dhw_file
-
-    assert dhw_profile.exists(), 'No DHWcalc File for the selected ' \
-                                 'parameters: {}'.format(dhw_file)
-
-    # Flowrate in Liter per Hour in each Step
-    water_LperH = [int(word.strip('\n')) for word in
-                   open(dhw_profile).readlines()]  # L/h each step
-
-    date_range = pd.date_range(start='2019-01-01', end='2020-01-01',
-                               freq=str(s_step) + 'S')
-    date_range = date_range[:-1]
-
-    # make dataframe
-    timeseries_df = pd.DataFrame(water_LperH, index=date_range, columns=[
-        'Water_LperH'])
-
-    timeseries_df['Water_L'] = timeseries_df['Water_LperH'] / 3600 * s_step
-    timeseries_df['method'] = 'DHWcalc'
-    timeseries_df['mean_drawoff_vol_per_day'] = mean_drawoff_vol_per_day
-    timeseries_df['categories'] = categories
-    timeseries_df['initial_day'] = 0
-    timeseries_df['weekend_weekday_factor'] = 1.2
[…]
