### 2025-02-21 RWTH-EBC/districtgenerator#62 @jtock
Edit:

- [ ] The result files should also be explained. Please add a short explanation of what the different result files show and which units they have.
