### 2025-02-21 #7657 @jtock
Dear @JoelSchoelzel,
I came across these issues while following my checklist (above) and the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). These issues need to be solved in a way, that I am able to approve the missing items / topics in the checklist.
This will require the rest of the review process to be iterative. I have just added another comment to https://github.com/RWTH-EBC/districtgenerator/issues/62 after doing a functional test.
Waiting for the second reviewer is fine with me, of course!
