+++ districtgenerator/functions/heating_profile_5R1C.py
@@ -1,0 +1,441 @@
+# -*- coding: utf-8 -*-
+"""
+Script to generate heating and cooling demands according to DIN EN ISO 13790.
+"""
+
+from __future__ import division
+import numpy as np
+import numpy.linalg as linalg
+
+
+def _solve(A, b):
+    return linalg.solve(A, b)
+
+
+def _calculateNoHeat(zoneParameters, T_e, t_m_previous, dt, timestep):
+    """
+    Calculate the temperatures (T_op, T_m, T_air, T_s) if neither heating nor
+    cooling devices are activated.
+    This is necessary to enable a deadband between cooling and heating mode.
+
+    Parameters
+    ----------
+    zoneParameters : ZoneParameters
+        Resistances and capacity
+    zoneInputs : ZoneInputs
+        External inputs (solar, internal gains, set temperatures)
+    T_m_init : float
+        Initial temperature of the thermal mass in degree Celsius.
+    timestep : integer, optional
+        Define which index is relevant (zoneInputs, H_ve)
+
+    Returns
+    -------
+    T_op : float
+        .
+    T_m : float
+        .
+    T_air : float
+        .
+    T_s : float
+        .
+    """
+
+    # Note: If not stated differently, all equations, pages and sections
+    # refer to DIN EN ISO 13790:2008 (the official German version of
+    # ISO 13790:2008).
+
+    # Extract parameters
+    H_tr_is = zoneParameters.H_tr_is        # in W/K
+    H_tr_ms = zoneParameters.H_tr_ms        # in W/K
+    H_tr_w  = zoneParameters.H_tr_w         # in W/K
+    H_ve    = zoneParameters.H_ve           # in W/K
+    C_m     = zoneParameters.C_m            # in J/K
+    H_tr_em = zoneParameters.H_tr_em[0]  # in W/K
+
+    Phi_ia = zoneParameters.phi_ia
+    Phi_m  = zoneParameters.phi_m
+    Phi_st = zoneParameters.phi_st
+
+
+    # Initialize A*x = b
+    # x: T_m, T_s, T_air (T_i), Q_HC
+    A = np.zeros((3,3))
+    b = np.zeros(3)
+
+    # Row wise entering
+    A[0,0] = H_tr_em + H_tr_ms + C_m / (3600 * dt)
+    A[0,1] = - H_tr_ms
+    A[1,0] = - H_tr_ms
+    A[1,1] = H_tr_ms + H_tr_is + H_tr_w
+    A[1,2] = - H_tr_is
+    A[2,1] = - H_tr_is
+    A[2, 2] = H_ve + H_tr_is
+
+    b[0] = Phi_m[timestep] + H_tr_em * T_e[timestep] + C_m * t_m_previous / (3600 * dt)
+    b[1] = Phi_st[timestep] + H_tr_w * T_e[timestep]
+    b[2] = Phi_ia[timestep] + H_ve * T_e[timestep]
+
+    # Solve for "x"
+    x = _solve(A, b)
+
+    T_i = x[2]
+    T_s = x[1]
+    T_m = x[0]
+
+    weight = 0.3
+    T_op = weight * T_i + (1 - weight) * T_s
+    return (T_op, T_m, T_i, T_s)
+
+
+def _calculateHeat(zoneParameters, T_e, T_set,T_m_init, dt, timestep):
+    """
+    Calculate the temperatures (Q_HC, T_op, T_m, T_air, T_s) that result when
+    reaching a given set temperature T_set.
+
+    Parameters
+    ----------
+    zoneParameters : ZoneParameters
+        Resistances and capacity
+    zoneInputs : ZoneInputs
+        External inputs (solar, internal gains, set temperatures)
+    T_m_init : float
+        Initial temperature of the thermal mass in degree Celsius.
+    T_set : float
[…]
