+++ classes/__init__.py
@@ -1,12 +1,0 @@
-# import of the main classes
-from districtgenerator.classes.datahandler import Datahandler
-from districtgenerator.classes.optimizer import Optimizer
-# import of subclasses
-from districtgenerator.classes.envelope import Envelope
-from districtgenerator.classes.solar import Sun
-from districtgenerator.classes.users import Users
-from districtgenerator.classes.profils import Profiles
-from districtgenerator.classes.system import BES
-from districtgenerator.classes.plots import DemandPlots
-# import of participants as houses and the aggregator
-from .participants import *
