+++ classes/participants/__init__.py
@@ -1,3 +1,0 @@
-from .house import House
-from .aggregator import Aggregator
-from .centralEnergyUnit import CentralEnergyUnit
