+++ districtgenerator/classes/users.py
@@ -94,7 +94,7 @@ class Users:
 
         # Initialize SIA class and read data
         self.SIA2024 = SIA.read_SIA_data()
-        if self.building in {"OB", "School", "Grocery_store"}:
+        if self.building in {"OB", "SC", "GS", "RE"}:
             self.building_zones = self.SIA2024[self.building]
 
         self.generate_number_flats_and_rooms(area)
@@ -108,42 +108,53 @@ class Users:
 
     def generate_number_flats_and_rooms(self, area):
         """
-        Generate number of flats for different of building types.
+        Generate number of flats and main rooms for different building types.
         Possible building types are:
             - single family house (SFH)
             - terraced house (TH)
             - multifamily house (MFH)
             - apartment block (AP)
+            - office building (OB)
+            - school (SC)
+            - grocery store (GS)
+            - restaurant (RE)
 
         Parameters
         ----------
         area : integer
             Floor area of different building types.
-        total_main_area : float
-            The main area consists only of the workspaces, which here are the offices
+
         Returns
         -------
-        None.
+        nb_flats : int
+            The estimated number of flats.
+        total_main_area : float
+            The main area consists only of the main zones in the building
         """
 
-        if self.building == "SFH":
+        # Residential buildings
+
+        if self.building in ["SFH", "TH"]:
             self.nb_flats = 1
-        elif self.building == "TH":
-            self.nb_flats = 1
-        elif self.building == "MFH":
-            if area <= 2 * 100:
-                self.nb_flats = 3
-            elif area <= 4 * 100:
-                self.nb_flats = 5
-            elif area > 4 * 100:
-                self.nb_flats = rd.randint((area//80)-1, (area//80)+1)
-        elif self.building == "AB":
-            if area <= 2 * 100:
-                self.nb_flats = 3
-            elif area <= 4 * 100:
-                self.nb_flats = 5
-            elif area > 4 * 100:
-                self.nb_flats = rd.randint((area//80)-1, (area//80)+1)
+
+        elif self.building in ["MFH", "AB"]:
+            # Source: Ergebnisse des Zensus 2022 Gebäude- und Wohnungszählung
+            area_categories = [(20, 40), (40, 59), (60, 79), (80, 99), (100, 119), (120, 139), (140, 159), (160, 179), (180, 199)]
+            probabilities = [0.05896, 0.181403, 0.23811, 0.17094, 0.118865, 0.107155, 0.067465, 0.0350566, 0.02204]
+            while True:
+                # Choose one consistent flat size for the entire building
+                chosen_area_range = rd.choices(area_categories, weights=probabilities, k=1)[0]
+                chosen_area = (chosen_area_range[0] + chosen_area_range[1]) // 2  # Use the mean area of the selected range
+
+                # Calculate the number of flats using rounding to the nearest integer
+                self.nb_flats = round(area / chosen_area)
+
+                # Ensure at least 2 flats
+                if self.nb_flats > 1:
+                    break
+
+        # Non-residential buildings
+
         else:
             if self.building == "OB":
                 proportion_office = 0
@@ -161,7 +172,7 @@ class Users:
                 self.total_main_area = area * proportion_office
                 self.nb_main_rooms = round(self.total_main_area / rd.gauss(mean_area_per_main_room, mean_area_per_main_room * 0.1))   # A random number of offices following a Gaussian distribution with a mean of mean_area_per_main_room (36 m²) and a standard deviation of 10%. The number of offices is calculated in order to later determine the number of occupants of the building depending on it
 
-            if self.building == "School":
+            if self.building == "SC":
                 proportion_classrooms = 0
                 nb_rooms = 0
[…]
