+++ districtgenerator/classes/datahandler.py
@@ -10,6 +10,7 @@ import numpy as np
 import openpyxl
 import pandas as pd
 import random as rd
+import holidays as hol
 from teaser.project import Project
 from .envelope import Envelope
 from .solar import Sun
@@ -58,6 +59,7 @@ class Datahandler:
 
         self.site = {}
         self.time = {}
+        self.initial_day = None
         self.district = []
         self.scenario_name = scenario_name
         self.scenario = None
@@ -193,6 +195,29 @@ class Datahandler:
             """
             # weatherdatafile_location = 507755060854
 
+    def get_holidays(self, country_code: str, year: int, state: str = None):
+        """
+        Get the Julian day (day of the year) for holidays in a specific country, year, and state.
+
+        Args:
+            country_code (str): The country's ISO 3166-1 alpha-2 code (e.g., 'DE' for Germany).
+            year (int): The year for which to retrieve holidays.
+            state (str): The state or region subdivision code (e.g., 'NW' for North Rhine-Westphalia in Germany).
+
+        Returns:
+            list: A list of tuples containing the Julian day of the holiday.
+        """
+        try:
+            # Initialize the holidays object for the given country, year, and state
+            holidays = hol.CountryHoliday(country_code, years=year, subdiv=state)
+
+            # Get the Julian day for each holiday
+            julian_holidays = [holiday_date.timetuple().tm_yday for holiday_date in holidays.keys()]
+
+            return julian_holidays
+        except KeyError:
+            return f"Invalid country or state code '{country_code}', '{state}'. Please provide valid codes."
+
     def generateEnvironment(self):
         """
         Load physical district environment - site and weather.
@@ -204,8 +229,11 @@ class Datahandler:
         # %% load first day of the year
         if self.site["TRYYear"] == "TRY2015":
             first_row = 35
+            self.initial_day = 3 # Thursday
         elif self.site["TRYYear"] == "TRY2045":
             first_row = 37
+            self.initial_day = 6 # Sunday
+
 
         self.select_plz_data()
         # load weather data
@@ -239,9 +267,9 @@ class Datahandler:
 
         # load the holidays
         if self.site["TRYYear"] == "TRY2015":
-            self.time["holidays"] = self.time["holidays2015"]
+            self.time["holidays"] = self.get_holidays(country_code="DE", year=2015)
         elif self.site["TRYYear"] == "TRY2045":
-            self.time["holidays"] = self.time["holidays2045"]
+            self.time["holidays"] = self.get_holidays(country_code="DE", year=2045)
 
         # interpolate input data to achieve required data resolution
         # transformation from values for points in time to values for time intervals
@@ -302,13 +330,15 @@ class Datahandler:
         -------
         None.
         """
-        duration = datetime.timedelta(minutes=1)
         num_sfh = 0
         num_mfh = 0
         num_OB = 0
-        num_schools = 0
+        num_school = 0
         num_grocery_store = 0
+        num_restaurant = 0
+        num_mfh_office_building = 0
         num_mfh_grocery_store = 0
+        num_mfh_restaurant = 0
 
         # Prepare a new list to collect updated scenario rows
         new_scenario = []
@@ -320,65 +350,74 @@ class Datahandler:
             # Make a copy of the row to work on.
             building_features = row.copy()
 
-            # Check if the building type is a combined type (MFH+Grocery_store or AB+Grocery_store)
-            if row["building"] in ['MFH+Grocery_store', 'AB+Grocery_store']:
+            # Check if the building type is a combined type
+            if row["building"] in ['MFH+OB', 'AB+OB', 'MFH+GS', 'AB+GS', 'MFH+RE', 'AB+RE']:
                 # Make a copy of the building features for the MFH part
                 building_mfh = building_features.copy()
-                # Assign a new unique id using the original id with a .1 suffix.
[…]
