### 2025-03-07 dc87447
final edits for certificate and restructure save and load functions #44
