+++ districtgenerator/classes/KPIs.py
@@ -10,9 +10,11 @@ from reportlab.pdfgen import canvas
 from reportlab.lib import colors
 from reportlab.graphics.shapes import *
 from reportlab.graphics.charts.piecharts import Pie
+from reportlab.graphics.charts.legends import Legend
 from datetime import datetime
 from reportlab.lib.styles import ParagraphStyle
 from reportlab.platypus import Paragraph
+from itertools import zip_longest
 
 class KPIs:
 
@@ -48,8 +50,14 @@ class KPIs:
         self.scf_year = None
         self.annual_investment_total = None
         self.totalarea = None
+        self.totalheatload = None
+        self.totalcoolingload = None
         self.totalnumberflats = None
         self.totalnumberocc = None
+        self.total_heating_demand = None
+        self.total_cooling_demand = None
+        self.total_electricity_demand = None
+        self.total_dhw_demand = None
 
         # initialize input data for calculation of KPIs
         inputData = {}
@@ -130,7 +138,6 @@ class KPIs:
 
         ### for central energy unit
 
-
     def calculateResidualLoad(self, data):
         """
         Calculate residual load at grid connection point (GCP) in [kW].
@@ -166,7 +173,6 @@ class KPIs:
 
         # maximal injection [kW]
         self.peakInjection = round(np.min(self.residualLoad) *(-1), 3)
-
 
     def calculatePeakToValley(self):
         """
@@ -223,7 +229,6 @@ class KPIs:
             self.W_inj_buildings_year += sum(self.sum_res_inj[c, :] * data.time["timeResolution"] / 3600 / 1000) \
                                          * self.inputData["clusterWeights"][self.inputData["clusters"][c]]
 
-
     def calculateCoverFactors(self, data):
         """
         Calculate the ratio between the self-consumed electricity and the total electricity demand.
@@ -232,9 +237,6 @@ class KPIs:
         -------
         None.
         """
-
-        # How much of the electricity demand is covered by the self-generated electricity? [-]
-
 
         self.supplyCoverFactor = np.zeros(len(self.inputData["clusters"]))
         self.demandCoverFactor = np.zeros(len(self.inputData["clusters"]))
@@ -320,6 +322,7 @@ class KPIs:
                                         + calc_annual_investment["EV"] * heater_counts["EV"]
                                         + calc_annual_investment["BAT"] * heater_counts["BAT"]
                                         + calc_annual_investment["BAT"] * heater_counts["BAT"])
+
     def calc_annual_cost_device(self, dev, param, cap):
         """
         Calculation of total investment costs including replacements (based on VDI 2067-1, pages 16-17).
@@ -339,9 +342,15 @@ class KPIs:
         # Projektlaufzeit (observation_time)
         # Zinssatz (für Kapitalkosten) (interest_rate)
         # Anschaffungskosten inkl. Installation, etc. [€/kW]
-        # Lebensdauer [Jahre]
-        # Betriebs- und Wartungskosten [% of Invest]
-        # Defaultwerte aus Technikkatalog KKW (Moritz, Nico)
+        # Lebensdauer (life_time) [years]
+        # Betriebs- und Wartungskosten (operation and maintenance costs, c_om) [% of Invest]
+
+        # Values from Technikkatalog (Langreder et al. 2024):
+        # Langreder, Nora; Lettow, Frederik; Sahnoun, Malek; Kreidelmeyer, Sven; Wünsch, Aurel; Lengning, Saskia et al.
+        # (2024): Technikkatalog Wärmeplanung. Hg. v. ifeu – Institut für Energie- und Umweltforschung Heidelberg,
+        # Öko-Institut e.V., IER Stuttgart, adelphi consult GmbH, Becker Büttner Held PartGmbB, Prognos AG, et al.
+        # Online available at:
[…]
