+++ districtgenerator/classes/users.py
@@ -31,7 +31,7 @@ class Users:
         Building type according to TABULA database.
     nb_flats : integer
         Number of flats in building.
-    annual_el_demand : array-like
+    annual_el_demand_per_flat : array-like
         Annual electricity consumption in dependency of the building type and the number of occupants.
     lighting_index : integer
         This index defines the lighting configuration of the household.
@@ -62,6 +62,7 @@ class Users:
 
         self.building = building
         self.nb_flats = None
+        self.annual_el_demand_per_flat = None
         self.annual_el_demand = None
         self.annual_heat_demand = None
         self.annual_dhw_demand = None
@@ -214,14 +215,15 @@ class Users:
 
             probabilities = [0.143, 0.143, 0.143, 0.142, 0.143, 0.143, 0.143]
 
-            self.annual_el_demand = np.zeros(self.nb_flats)
+            self.annual_el_demand_per_flat = np.zeros(self.nb_flats)
+            self.annual_el_demand = 0
             for j in range(self.nb_flats):
                 if self.building == "SFH":
                     random_nb = rd.random()  # picking random number in [0,1) to decide between which 2 values of consumption_range the annual electricity consumption lies
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -230,7 +232,7 @@ class Users:
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1],consumption_range["SFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -239,7 +241,7 @@ class Users:
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -248,10 +250,12 @@ class Users:
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
[…]

[…]
