### 2025-03-10 77b1c94
44 add certificate (#72)
* add first code from certificate development, #44

* final edits for certificate and restructure save and load functions #44
