+++ districtgenerator/classes/datahandler.py
@@ -100,7 +100,7 @@ class Datahandler:
             for subData in jsonData:
                 self.site[subData["name"]] = subData["value"]
 
-                # %% load time information and requirements (used in generateEnvironment)
+        # %% load time information and requirements (used in generateEnvironment)
         # needed for data conversion into the right time format
         with open(os.path.join(self.filePath, 'time_data.json')) as json_file:
             jsonData = json.load(json_file)
@@ -303,22 +303,58 @@ class Datahandler:
         duration = datetime.timedelta(minutes=1)
         num_sfh = 0
         num_mfh = 0
+        name_pool = []
 
         # initialize buildings for scenario
         # loop over all buildings
         for id in self.scenario["id"]:
-            # create empty dict for observed building
-            building = {}
+            try:
+                # Check if ID is a number
+                try:
+                    # Try to convert id to float to check if it's numeric
+                    float(id)
+                except (ValueError, TypeError):
+                    raise ValueError(f"Building ID '{id}' is not a number")
 
-            # store features of the observed building
-            building["buildingFeatures"] = self.scenario.loc[id]
+                # Create empty dict for observed building
+                building = {}
 
-            # append building to district
-            self.district.append(building)
+                # Store features of the observed building
+                building["buildingFeatures"] = self.scenario.loc[id]
 
-            # Count number of builidings to predict the approximate calculation time
-            if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"]["building"] == 'TH': num_sfh +=1
-            elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"]["building"] == 'AB': num_mfh +=1
+                # %% Create unique building name
+                # needed for loading and storing data with unique name
+                # name is composed of building id, and building type
+                name = str(id) + "_" + building["buildingFeatures"]["building"]
+
+                # Check if the name is already in the district
+                if name in name_pool:
+                    raise ValueError(f"Building name '{name}' is not unique. ID collision detected.")
+
+                name_pool.append(name)
+
+                # Assign the unique name to the building
+                building["unique_name"] = name
+
+                # Append building to district
+                self.district.append(building)
+
+                # Count number of builidings to predict the approximate calculation time
+                if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"]["building"] == 'TH':
+                    num_sfh +=1
+                elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"]["building"] == 'AB':
+                    num_mfh +=1
+            except ValueError as e:
+                # Handle the case where we have a duplicate name
+                print(f"Error: {e}")
+                print(f"The building ID must be a unique number to ensure proper identification and data tracking.")
+                print(f"Building with ID {id} will be skipped and not added to the district")
+                continue
+            except Exception as e:
+                # Handle any other unexpected errors
+                print(f"Unexpected error processing building ID {id}: {e}")
+                print(f"Building with ID {id} will be skipped and not added to the district.")
+                continue
 
         # Calculate calculation time for the whole district generation
         duration += datetime.timedelta(seconds= 3 * num_sfh + 12 * num_mfh)
@@ -412,18 +448,7 @@ class Datahandler:
         None.
         """
 
-
-        set = []
         for building in self.district:
-            # %% create unique building name
-            # needed for loading and storing data with unique name
-            # name is composed of building type, number of flats, serial number of building of this properties
[…]
