+++ districtgenerator/classes/non_residential.py
@@ -3,39 +3,25 @@
 
 import os
 import json
-from teaser.logic.buildingobjects.thermalzone import ThermalZone
-from teaser.logic.buildingobjects.useconditions import UseConditions as UseCond
 
 
 class NonResidential(object):
     """Base class for each non-residential archetype.
 
-    This is the base class for all non-residential archetype buildings according to data NWG.
+    This is the base class for all non-residential archetype buildings.
     The setup is similiar to TEASER, considering that there are no information about layers and 
-    details about material available. The class provides the necessary information for the districtgenerator. 
-
-    These are:
-        - prj.data.material_bind -> No data available for non-residential Archetypes
-        - prj.data.element_bind - -> No data available for non-residential Archetypes
-        - prj.buildings[self.id].volume 
-        - prj.buildings[self.id].outer_area
-        - prj.buildings[self.id].window_area -> has to be calculated, as it is not present in data NWG.
+    details about material available. The class provides the necessary information for the districtgenerator.
 
     Teaser uses typical layers and material for the build environment according to MASEA ( https://www.masea-ensan.de/ ).
     Currently, the gathering and combinatory selection of this is not feasible. 
 
     Several parameters are obligatory: name, year_of_construction, net_leased_area, type.
-    
-    Other parameters are optional, and are implemented to provide further options for configuration. 
 
     In accordance with TEASER it is assumed, that each building has four orientations of outer walls and windows (north,
-    east, south and west), two orientations for rooftops (south and north), 
-    with tilt of 35 degree and one orientation for ground floors and one door (default orientation is west).
+    east, south and west)
 
     In TEASER it is assumed,  that the surface is the product of the given net_leased_area and specific estimation factors. These
     estimation factors where build by dividing the given 'surface area' by the 'reference floor area' in TABULA. The same approach is followed here. 
-    The factors are given in: 
-        - 
 
     Obliagotry Parameters
     ----------
@@ -45,82 +31,24 @@ class NonResidential(object):
     year_of_construction : int
         Year of first construction
     net_leased_area : float [m2]
-        Total net leased area of building. This is area is NOT the footprint
-        of a building
+            Total net leased area of the building, or of the building part if it is a mixed-use building. This is area is NOT the footprint of a building
+    total_building_area : float [m2]
+            Total net leased area of building. This is area is NOT the footprint of a building
     usage: str 
         Type of the building, according to Data NWG. Options are: "oag", "rnt", "hlc", "sdc", "clt", "spf", "hbr", "pwo", "trd", "tud", "trs", "gs1", "gs2"
-    
-        
-    Optional Parameters
-    ------------- 
-
-    street_name : string
-        Name of the street the building is located at. (optional)
-    city : string
-        Name of the city the building is located at. (optional)
-    longitude : float [degree]
-        Longitude of building location.
-    latitude : float [degree]
-        Latitude of building location.
 
     Attributes
     ----------
-    central_ahu : instance of BuildingAHU
-        Teaser Instance of BuildingAHU if a central AHU is embedded into the
-        building (currently mostly needed for AixLib simulation).
-    number_of_floors : int
-        number of floors above ground (default: None)
     height_of_floors : float [m]
         Average height of the floors (default: None)
-    internal_id : float
-        Random id for the distinction between different buildings.
-    year_of_retrofit : int
-        Year of last retrofit.
-    type_of_building : string
-        Type of a Building (e.g. Building (unspecified), Office etc.).
-    building_id : None
[…]
