### 2025-03-10 9fdb7c8
add requirements for sphinx documentation
