### 2025-03-10 2999e7e
Improve assumption handling
