### 2025-03-10 2eb2439
Improve assumption handling
