### 2025-03-10 RWTH-EBC/districtgenerator#67 @RawadHamze
Thank you for your feedback!

1. Regarding your first point: There is no hard coding for building properties and material constants, as these values are sourced from tabula via teaser. Similarly, climate conditions are not hardcoded but taken from the DWD dataset.
2. Our goal is to make the tool as user-friendly as possible. Therefore, every value in example.csv is fully configurable, while the remaining parameters are derived from sources.
3. Any hardcoded values in the tool are independent of climate zones and building specifications. These assumptions are general and apply to all cases.
4. For assumptions that previously lacked references, we have now added appropriate sources.
5. Regarding your suggestion to document assumptions in a structured and standardized way: We will include this in the documentation.

The relevant fixes have been implemented in commit [2eb2439]
