+++ districtgenerator/data_handling/data_generator.py
@@ -1,0 +1,218 @@
+import json
+import csv
+import pandas as pd
+from typing import Dict, List, Tuple
+from dataclasses import dataclass, asdict
+import itertools
+
+
+@dataclass
+class TechnologyParams:
+    PV: bool = False
+    STC: bool = False
+    EV: bool = False
+    BAT: bool = False
+
+    # Class method to get valid ranges for binary parameters
+    @classmethod
+    def get_valid_ranges(cls) -> Dict[str, Tuple[int, int]]:
+        return {
+            "PV": (0, 1),
+            "STC": (0, 1),
+            "EV": (0, 1),
+            "BAT": (0, 1)
+        }
+
+
+@dataclass
+class StorageParams:
+    f_TES: float = 35.0  # Default thermal energy storage factor
+    f_BAT: float = 1.0  # Default battery storage factor
+    f_EV: float = 6000.0  # Default EV storage factor
+    f_PV: float = 0.4  # Default PV sizing factor
+    f_STC: float = 0.04  # Default solar thermal collector factor
+
+    # Class method to get valid ranges for continuous parameters
+    @classmethod
+    def get_valid_ranges(cls) -> Dict[str, Tuple[float, float]]:
+        # TODO: check if valid
+        return {
+            "f_TES": (20.0, 50.0),  # Thermal storage 20-50 L/m²
+            "f_BAT": (0.5, 2.0),  # Battery sizing factor 0.5-2
+            "f_EV": (3000.0, 9000.0),  # EV storage 3-9 kWh
+            "f_PV": (0.2, 1.0),  # PV sizing factor 20-100%
+            "f_STC": (0.02, 0.08)  # Solar thermal collector sizing 2-8%
+        }
+
+@dataclass
+class ChargingModes:
+    modes: List[str] = None
+
+    def __post_init__(self):
+        if self.modes is None:
+            self.modes = ["on_demand"]
+
+    @classmethod
+    def get_valid_modes(cls) -> List[str]:
+        return ["on_demand", "night_charge", "pv_optimized"]  # Todo: Example future modes
+
+
+@dataclass
+class LocationConfig:
+    time_zone: float
+    albedo: float
+    try_year: str
+    try_type: str
+    zip_code: str
+
+
+@dataclass
+class BuildingConfig:
+    building_types: List[str]
+    construction_years: List[int]
+    retrofit_options: List[int]
+    area_ranges: Dict[str, List[float]]
+    heater_types: List[str]
+    tech_params: TechnologyParams = TechnologyParams()
+    storage_params: StorageParams = StorageParams()
+    charging_modes: ChargingModes = ChargingModes()
+
+
+class DataGenerator:
+    def __init__(self, location_config: LocationConfig, building_config: BuildingConfig):
+        self.location = location_config
+        self.building = building_config
+        self.validate_parameters()
+
+    def validate_parameters(self):
+        """Validate that all parameters are within their allowed ranges"""
+        # Validate technology parameters
+        tech_ranges = TechnologyParams.get_valid_ranges()
+        for param, (min_val, max_val) in tech_ranges.items():
+            value = getattr(self.building.tech_params, param)
+            if not isinstance(value, bool) and (value < min_val or value > max_val):
+                raise ValueError(f"{param} must be between {min_val} and {max_val}")
+
+        # Validate storage parameters
+        storage_ranges = StorageParams.get_valid_ranges()
+        for param, (min_val, max_val) in storage_ranges.items():
+            value = getattr(self.building.storage_params, param)
+            if value < min_val or value > max_val:
+                raise ValueError(f"{param} must be between {min_val} and {max_val}")
+
+        # Validate charging modes
+        valid_modes = ChargingModes.get_valid_modes()
[…]
