### 2025-03-12 d902900
Solving issues #66 # 65 #62 #61 #42
