+++ joss/paper.bib
@@ -39,6 +39,28 @@ author = {Jan Schiefelbein and Jana Rudnick and Anna Scholl and Peter Remmen and
 keywords = {City district, Urban energy models, Geographic information systems, Data enrichment, Archetypes, Workflow automation},
 }
 
+@article{bose_reinforcement_2021,
+	title = {Reinforcement learning in local energy markets},
+	volume = {4},
+	doi = {10.1186/s42162-021-00141-z},
+	language = {en},
+	number = {1},
+	urldate = {2022-08-03},
+	journal = {Energy Informatics},
+	author = {Bose, Samrat and Kremers, Enrique and Mengelkamp, Esther Marie and Eberbach, Jan and Weinhardt, Christof},
+	month = dec,
+	year = {2021},
+	pages = {7},
+}
+
+@article{schoelzel_districtgenerator_2024,
+	title = {Districtgenerator: a novel open-source webtool to generate building-specific load profiles and evaluate energy systems of residential districts},
+	volume = {4},
+	doi = {10.52202/077185-0106},
+	language = {en},
+	author = {Schölzel, Joel and Beckhölter, Tobias and Henn, Sarah and Wüller, Carla and Streblow, Rita and Müller, Dirk},
+	year = {2024},
+}
 
 @article{richardson_domestic_2010,
 	title = {Domestic electricity use: {A} high-resolution energy demand model},
@@ -80,6 +102,7 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	author = {{DIN EN 12831-1:2017-09}},
  	key = {{DIN EN 12831-1}},
 	doi = {10.31030/2571775},
+	year = {2017},
 	pages = {101},
 }
 
@@ -87,6 +110,7 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	title = {Energy performance of buildings - Calculation of energy use for space heating and cooling ({ISO} 13790:2008); {German} version {EN} {ISO} 13790:2008},
 	author = {{DIN EN ISO 13790:2008-09}},
     key = {{DIN EN ISO 13790}},
+    year = {2008},
 	pages = {185},
 }
 
@@ -95,11 +119,13 @@ keywords = {City district, Urban energy models, Geographic information systems, 
     key = {{DIN 4708-1}},
 	author = {{DIN 4708-1:1994-04}},
 	doi = {10.31030/2480131},
+	year = {1994},
 	pages = {8},
 }
 
 @misc{REDII_2018,
 	title = {Directive ({EU}) 2018/2001 of the {European} {Parliament} and of the {Council} of 11 {December} 2018 on the promotion of the use of energy from renewable sources},
+	author = {{European Parliament}},
 	volume = {328},
  	key = {REDII},
 	url = {http://data.europa.eu/eli/dir/2018/2001/oj/eng},
@@ -110,6 +136,7 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 
 @misc{IEMD_2019,
 	title = {Directive ({EU}) 2019/944 of the {European} {Parliament} and of the {Council} of 5 {June} 2019 on common rules for the internal market for electricity and amending {Directive} 2012/27/{EU} (recast) ({Text} with {EEA} relevance.)},
+	author = {{European Parliament}},
 	volume = {158},
  	key = {IEMD},
 	url = {http://data.europa.eu/eli/dir/2019/944/oj/eng},
