+++ joss/paper.md
@@ -9,20 +9,20 @@ authors:
   - name: Sarah Henn
     orcid: 0000-0002-3423-7511
     affiliation: 1
-  - name: Marius Brünjes
-    affiliation: 1
   - name: Joel David Schölzel
     orcid: 0009-0000-5053-3161
     affiliation: 1
   - name: Tobias Beckhölter
     affiliation: 1
+  - name: Carla Wüller
+    affiliation: 1 
   - name: Dirk Müller
     orcid: 0000-0002-6106-6607
     affiliation: 1
 affiliations:
  - name: Institute for Energy Efficient Buildings and Indoor Climate, RWTH Aachen University
    index: 1
-date: 01 June 2023
+date: 13 July 2024
 bibliography: paper.bib
 ---
 
@@ -47,14 +47,21 @@ Thus, detailed time-resolved profiles of energy demand and generation are requir
 * **Synergy** potentials between buildings of different usage structures can be exploited through joint concepts.
 While joint heating concepts, such as local heating networks, have been the focus of research for a long time, joint electricity concepts have only recently emerged.
 For example, the European Union proposed and introduced the concept of energy communities [@REDII_2018; @IEMD_2019].
-According to the @REDII_2018 directive, a district can also form an energy community where energy is shared or traded Peer-to-Peer.
-By this, district concepts lead to financial benefits for consumers and prosumers. 
-Moreover, self-consumption increases, peaks are evened out and the superordinate grid might be decongested. 
+According to the Renewable Energy Directive [@REDII_2018], a district can also form an energy community where energy is shared or traded Peer-to-Peer.
+By this, district concepts can lead to financial benefits for consumers and prosumers. Moreover, its expected that self-consumption increases, 
+* peaks are evened out and the superordinate grid might be decongested. [@bose_reinforcement_2021]
 
 * **Scalability** is a key issue for a resource-efficient transformation of the energy system, especially as numerous existing districts will have to be retrofitted and newly equipped in the next few years [@IEA_2021].
 The `Districtgenerator` enables its users to easily gain information about districts in order to identify promising concepts and conduct comparative evaluations.
 By using representative building types, data for representative districts is created. 
 Thus, the data applies to a wide range of residential districts and can be included in guidelines or representative examples.
+
+Also existing tools facilitate during the strategic and preliminary planning phases the identification of measures that 
+support the detailed planning and subsequent implementation of sustainable and sector-coupling energy systems in neighborhoods.
+However, the models and calculation methods of the time series are often not transparent. For stakeholders such as 
+investors, urban planner, property owners or municipalities with limited resources, the complexity of these tools can be a barrier. Their user-friendliness 
+is low, leading to potentially time-intensive applications. Furthermore, utilizing them may require specialized 
+knowledge in energy planning, modeling, and data analysis, thus restricting access for less experienced users. [@schoelzel_districtgenerator_2024]
 
 # Functional principle 
 
@@ -67,11 +74,10 @@ For every site, the user is able to select between warm, cold and normal TRYs.
 Moreover, TRYs are available for 2015 and 2045, reflecting either current or future weather conditions.
 
 \setkeys{Gin}{width=.9\linewidth}
-![Usage of external tools and data sources to generate occupancy and demand profiles with the `Districtgenerator`.\label{fig:flowchart}](Schema_QG.pdf "Usage of external tools and data sources to generate results with the `Districtgenerator`") \label{fig:flowchart}
-% ![[Usage of external tools and data sources to generate occupancy and demand profiles with the `Districtgenerator`.\label{fig:flowchart}](Schema_QG.pdf) { width=80% }
+![Figure 1: Usage of external tools and data sources to generate occupancy and demand profiles with the `Districtgenerator`.}](Schema_QG.pdf)
 
[…]
