+++ districtgenerator/data_handling/data_generator.py
@@ -2,10 +2,39 @@ import json
 import csv
 import pandas as pd
 from typing import Dict, List, Tuple
-from dataclasses import dataclass, asdict
+from dataclasses import dataclass, asdict, field
 import itertools
 
+# Data classes for configuration
+# Location
+@dataclass
+class ValidAlbedo:
+    min_val: float = 0.0
+    max_val: float = 1.0
 
+    @classmethod
+    def get_valid_range(cls) -> Tuple[float, float]:
+        return (cls.min_val, cls.max_val)
+
+
+@dataclass
+class TRYYears:
+    years: List[str] = field(default_factory=lambda: ["TRY2015", "TRY2045"])
+
+    @classmethod
+    def get_valid_years(cls) -> List[str]:
+        return ["TRY2015", "TRY2045"]
+
+
+@dataclass
+class TRYTypes:
+    types: List[str] = field(default_factory=lambda: ["Jahr", "Somm", "Wint"])
+
+    @classmethod
+    def get_valid_types(cls) -> List[str]:
+        return ["Jahr", "Somm", "Wint"]
+
+# Building
 @dataclass
 class TechnologyParams:
     PV: bool = False
@@ -13,7 +42,6 @@ class TechnologyParams:
     EV: bool = False
     BAT: bool = False
 
-    # Class method to get valid ranges for binary parameters
     @classmethod
     def get_valid_ranges(cls) -> Dict[str, Tuple[int, int]]:
         return {
@@ -23,40 +51,33 @@ class TechnologyParams:
             "BAT": (0, 1)
         }
 
-
 @dataclass
 class StorageParams:
-    f_TES: float = 35.0  # Default thermal energy storage factor
-    f_BAT: float = 1.0  # Default battery storage factor
-    f_EV: float = 6000.0  # Default EV storage factor
-    f_PV: float = 0.4  # Default PV sizing factor
-    f_STC: float = 0.04  # Default solar thermal collector factor
+    f_TES: float = 35.0
+    f_BAT: float = 1.0
+    f_EV: float = 6000.0
+    f_PV: float = 0.4
+    f_STC: float = 0.04
 
-    # Class method to get valid ranges for continuous parameters
     @classmethod
     def get_valid_ranges(cls) -> Dict[str, Tuple[float, float]]:
-        # TODO: check if valid
         return {
-            "f_TES": (20.0, 50.0),  # Thermal storage 20-50 L/m²
-            "f_BAT": (0.5, 2.0),  # Battery sizing factor 0.5-2
-            "f_EV": (3000.0, 9000.0),  # EV storage 3-9 kWh
-            "f_PV": (0.2, 1.0),  # PV sizing factor 20-100%
-            "f_STC": (0.02, 0.08)  # Solar thermal collector sizing 2-8%
+            "f_TES": (20.0, 50.0),
+            "f_BAT": (0.5, 2.0),
+            "f_EV": (3000.0, 9000.0),
+            "f_PV": (0.2, 1.0),
+            "f_STC": (0.02, 0.08)
         }
 
 @dataclass
 class ChargingModes:
-    modes: List[str] = None
-
-    def __post_init__(self):
-        if self.modes is None:
-            self.modes = ["on_demand"]
+    modes: List[str] = field(default_factory=lambda: ["on_demand"])
 
     @classmethod
     def get_valid_modes(cls) -> List[str]:
-        return ["on_demand", "night_charge", "pv_optimized"]  # Todo: Example future modes
+        return ["on_demand", "night_charge", "pv_optimized"]
 
-
+#Config
 @dataclass
 class LocationConfig:
     time_zone: float
@@ -65,17 +86,26 @@ class LocationConfig:
     try_type: str
     zip_code: str
 
+    def __post_init__(self):
[…]
