+++ joss/paper.bib
@@ -155,18 +155,16 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 
 @misc{DWD_2023,
 	title = {Klimaberatungsmodul},
-    key = {DWD},
+    url = {https://kunden.dwd.de/obt/},
 	author = {{Deutscher Wetterdienst}},
 	year = {2023},
-	url = {https://kunden.dwd.de/obt/},
 }
 
 @misc{BBSR_2020,
 	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
-    key = {BBSR},
+	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
 	author = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
 	year = {2020},
-	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
 }
 
 
