### 2025-03-13 ab0a17a
minor adjustment building generation
