### 2025-03-13 e46e041
minor adjustment building generation
