### 2025-03-13 6f99f0b
minor adjustment building generation
