### 2025-03-13 3b95de4
Solving #64 #63 #59
