### 2025-03-13 a09a9fd
Solving #64 #63 #59
