+++ joss/paper.bib
@@ -97,6 +97,20 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	year = {2019},
 }
 
+@misc{DWD_2023,
+	title = {Klimaberatungsmodul},
+    url = {https://kunden.dwd.de/obt/},
+	author = {{Deutscher Wetterdienst}},
+	year = {2023},
+}
+
+@misc{BBSR_2020,
+	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
+	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
+	author = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
+	year = {2020},
+}
+
 @misc{din_en_12831,
 	title = {Energy performance of buildings - {Method} for calculation of the design heat load - {Part} 1: {Space} heating load, {Module} {M3}-3; {German} version {EN} 12831-1:2017},
 	author = {{DIN EN 12831-1:2017-09}},
@@ -153,19 +167,7 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	doi = {10.1787/c8328405-en},
 }
 
-@misc{DWD_2023,
-	title = {Klimaberatungsmodul},
-    url = {https://kunden.dwd.de/obt/},
-	author = {{Deutscher Wetterdienst}},
-	year = {2023},
-}
 
-@misc{BBSR_2020,
-	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
-	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
-	author = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
-	year = {2020},
-}
 
 
 
