### 2025-03-13 RWTH-EBC/districtgenerator#66 @JoelSchoelzel
Dear @jtock, we have now revised all docstrings and also checked the inline comments.
With regard to issue #60, we are still working on this
