### 2025-03-13 RWTH-EBC/districtgenerator#65 @JoelSchoelzel
Dear @jtock, i have corrected the date and the authors in the paper. However, it is correct that Sarah Henn is listed in the documentation under Alumni because she no longer works at our institute, but was the first author when the paper was written.
