### 2025-03-13 RWTH-EBC/districtgenerator#63 @JoelSchoelzel
Dear @jtock, we have added concrete output to the examples and also revised the other points.
