### 2025-03-13 RWTH-EBC/districtgenerator#62 @JoelSchoelzel
Dear @jtock, we have added the points mentioned above to the documentation and hope that the documentation is now more intuitive to understand.
