### 2025-03-13 RWTH-EBC/districtgenerator#59 @JoelSchoelzel
Dear @snjsomnath, since it is not possible to change the data from the .json files in the code due to the workflow, we have added the referencing in the other examples
