+++ joss/paper.bib
@@ -97,14 +97,14 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	year = {2019},
 }
 
-@misc{DWD_2023,
+@online{DWD_2023,
 	title = {Klimaberatungsmodul},
     url = {https://kunden.dwd.de/obt/},
 	author = {{Deutscher Wetterdienst}},
 	year = {2023},
 }
 
-@misc{BBSR_2020,
+@online{BBSR_2020,
 	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
 	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
 	author = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
