+++ joss/paper.bib
@@ -100,14 +100,14 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 @online{DWD_2023,
 	title = {Klimaberatungsmodul},
     url = {https://kunden.dwd.de/obt/},
-	author = {{Deutscher Wetterdienst}},
+	organization  = {Deutscher Wetterdienst},
 	year = {2023},
 }
 
 @online{BBSR_2020,
 	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
 	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
-	author = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
+	organization  = {Bundesinstitut für Bau-, Stadt- und Raumforschung},
 	year = {2020},
 }
 
