### 2025-03-13 #7657 @JoelSchoelzel
@AdamRJensen I tried multiple time fixing the wrong references "weather data (?; ?)" but i dont really see the error in paper.bib or paper.md. Do you have any idea or hint why those two are not working?
