### 2025-03-13 168c296
minor adjustment building generation
