### 2025-03-17 #7657 @AdamRJensen
> [@AdamRJensen](https://github.com/AdamRJensen) I tried multiple time fixing the wrong references "weather data (?; ?)" but i dont really see the error in paper.bib or paper.md. Do you have any idea or hint why those two are not working?

@JoelSchoelzel could you try and make the citations capital (the keys in the bib text file are capital). I don't know if it is case sensitive, but it's worth a short.
