+++ districtgenerator/classes/KPIs.py
@@ -5,6 +5,16 @@ import numpy as np
 import os
 import json
 import math
+import reportlab
+from reportlab.pdfgen import canvas
+from reportlab.lib import colors
+from reportlab.graphics.shapes import *
+from reportlab.graphics.charts.piecharts import Pie
+from reportlab.graphics.charts.legends import Legend
+from datetime import datetime
+from reportlab.lib.styles import ParagraphStyle
+from reportlab.platypus import Paragraph
+from itertools import zip_longest
 
 class KPIs:
 
@@ -39,6 +49,16 @@ class KPIs:
         self.dcf_year = None
         self.scf_year = None
         self.annual_investment_total = None
+        self.totalarea_residential = None
+        self.totalarea_non_residential = None
+        self.totalheatload = None
+        self.totalcoolingload = None
+        self.totalnumberflats = None
+        self.totalnumberocc = None
+        self.total_heating_demand = None
+        self.total_cooling_demand = None
+        self.total_electricity_demand = None
+        self.total_dhw_demand = None
 
         # initialize input data for calculation of KPIs
         inputData = {}
@@ -119,7 +139,6 @@ class KPIs:
 
         ### for central energy unit
 
-
     def calculateResidualLoad(self, data):
         """
         Calculate residual load at grid connection point (GCP) in [kW].
@@ -155,7 +174,6 @@ class KPIs:
 
         # maximal injection [kW]
         self.peakInjection = round(np.min(self.residualLoad) *(-1), 3)
-
 
     def calculatePeakToValley(self):
         """
@@ -212,7 +230,6 @@ class KPIs:
             self.W_inj_buildings_year += sum(self.sum_res_inj[c, :] * data.time["timeResolution"] / 3600 / 1000) \
                                          * self.inputData["clusterWeights"][self.inputData["clusters"][c]]
 
-
     def calculateCoverFactors(self, data):
         """
         Calculate the ratio between the self-consumed electricity and the total electricity demand.
@@ -221,9 +238,6 @@ class KPIs:
         -------
         None.
         """
-
-        # How much of the electricity demand is covered by the self-generated electricity? [-]
-
 
         self.supplyCoverFactor = np.zeros(len(self.inputData["clusters"]))
         self.demandCoverFactor = np.zeros(len(self.inputData["clusters"]))
@@ -240,8 +254,12 @@ class KPIs:
                 b = 0
                 # sum of all buildings for each timestep
                 for id in data.scenario["id"]:
-                    a += self.inputData["resultsOptimization"][c][id]["res_load"][t]
-                    b += self.inputData["resultsOptimization"][c][id]["res_inj"][t]
+                    a += self.inputData["resultsOptimization"][c][id]["res_load"][t] # res_load is the residual load of the building, meaning the net electricity needed for the building after accounting for all the consumers and producers in the building, including its PV, CHP...
+                    b += self.inputData["resultsOptimization"][c][id]["res_inj"][t]  # res_inj is the residual feed from the building, meaning the net electricity fed to the electrical grid from the building after accounting for all internal electricity usage in the building
+                # At the same time step t, either res_load or res_inj should be 0.
+                # However, a and b could both be greater than 0 at the same time step t,
+                # since they represent the sums of all the buildings.
+                # If both a and b are greater than 0, it means electricity is being transported from one building to another.
                 # sum of all timesteps
                 nenner_dem[c, t] += a
                 nenner_sup[c, t] += b
@@ -309,6 +327,7 @@ class KPIs:
[…]
