+++ districtgenerator/classes/datahandler.py
@@ -105,7 +105,7 @@ class Datahandler:
             for subData in jsonData:
                 self.site[subData["name"]] = subData["value"]
 
-                # %% load time information and requirements (used in generateEnvironment)
+        # %% load time information and requirements (used in generateEnvironment)
         # needed for data conversion into the right time format
         with open(os.path.join(self.filePath, 'time_data.json')) as json_file:
             jsonData = json.load(json_file)
@@ -123,13 +123,13 @@ class Datahandler:
             for subData in jsonData:
                 self.design_building_data[subData["name"]] = subData["value"]
 
-                # load building physics data (used in envelope and system BES/CES)
+        # load building physics data (used in envelope and system BES/CES)
         with open(os.path.join(self.filePath, 'physics_data.json')) as json_file:
             jsonData = json.load(json_file)
             for subData in jsonData:
                 self.physics[subData["name"]] = subData["value"]
 
-                # Load list of possible devices (used in system BES)
+        # Load list of possible devices (used in system BES)
         with open(os.path.join(self.filePath, 'decentral_device_data.json')) as json_file:
             jsonData = json.load(json_file)
             for subData in jsonData:
@@ -331,124 +331,108 @@ class Datahandler:
         -------
         None.
         """
-        num_sfh = 0
-        num_mfh = 0
-        num_OB = 0
-        num_school = 0
-        num_grocery_store = 0
-        num_restaurant = 0
-        num_mfh_office_building = 0
-        num_mfh_grocery_store = 0
-        num_mfh_restaurant = 0
+        # Building type counters
+        building_counts = {
+            "SFH": 0, "TH": 0, "MFH": 0, "AB": 0, "OB": 0, "SC": 0,
+            "GS": 0, "RE": 0, "MFH+OB": 0, "AB+OB": 0, "MFH+GS": 0,
+            "AB+GS": 0, "MFH+RE": 0, "AB+RE": 0
+        }
 
+        name_pool = set()
         # Prepare a new list to collect updated scenario rows
         new_scenario = []
         next_id = 0
 
         # initialize buildings for scenario
-        # loop over all buildings
+        # Iterate through all buildings in scenario
         for idx, row in self.scenario.iterrows():
+
+            # Ensure ID is numeric
+            if not isinstance(idx, (int, float)):
+                try:
+                    idx = float(idx)
+                except (ValueError, TypeError):
+                    raise ValueError(f"Building ID '{idx}' is not a number")
+
             # Make a copy of the row to work on.
             building_features = row.copy()
+            building_type = row["building"]
 
             # Check if the building type is a combined type
-            if row["building"] in ['MFH+OB', 'AB+OB', 'MFH+GS', 'AB+GS', 'MFH+RE', 'AB+RE']:
-                # Make a copy of the building features for the MFH part
-                building_mfh = building_features.copy()
-                # Assign a new unique id using the original id
-                building_mfh["id"] = next_id
-                # Set the building type to MFH.
-                building_mfh["building"] = "MFH"
-                # Reset the construction type so that it matches with teaser.
-                building_mfh["construction_type"] = ""
-
-                self.total_building_area = building_features["area"]
-
-                # Determine secondary share and type:
-                # For office building: occupies 4 flats; for grocery store: occupies 4 flats; for restaurant: occupies 2 flats.
-                if row["building"] in ['MFH+OB', 'AB+OB']:
-                    secondary_share = 4 # Assumption
-                    secondary_type = "OB"
-                elif row["building"] in ['MFH+GS', 'AB+GS']:
-                    secondary_share = 4 # Assumption
-                    secondary_type = "GS"
[…]

[…]
