### 2025-03-21 dd81a9b
Edit central optimization to use in district_network model 2
