+++ examples/e7_optimal_dimensioning_energyHub.py
@@ -10,11 +10,11 @@ from districtgenerator.classes import *
 def example7_optiEnergyCentral_EHDO():
 
     # Initialize District
-    data = Datahandler(scenario_name = "example")
+    data = Datahandler(scenario_name = "1_district")
 
     # We directly generate a complete district.
-    data.generateDistrictComplete(calcUserProfiles=False, saveUserProfiles=False,
-                                  saveGenProfiles=False)
+    data.generateDistrictComplete(calcUserProfiles=False, saveUserProfiles=True,
+                                  saveGenProfiles=True)
 
     # As last step we use the EHDO tool to get an optimized energy central for neighborhoods. EHDO is a tool for
     # planning and designing complex energy systems. The central feature is coupling of different
