### 2025-03-21 27e4c60
Solving #60 #43
