+++ joss/paper.md
@@ -69,7 +69,7 @@ The minimum input data required is the number of buildings  and basic informatio
 The district size is freely choosable and the number of buildings to compute is not limited by the program. 
 The default file format for the building information is .csv because of its compatibility and ease of use.
 Optionally, the site of the district, the time resolution of the profiles and the test reference year (TRY) for weather data 
-[@dwd_2023; @bbsr_2020] can be modified. 
+[@DWD_2023; @BBSR_2020] can be modified. 
 There are 16 choosable sites, which refer to the representative locations of the German climate zones. 
 For every site, the user is able to select between warm, cold and normal TRYs. 
 Moreover, TRYs are available for 2015 and 2045, reflecting either current or future weather conditions.
