### 2025-03-21 RWTH-EBC/districtgenerator#60 @carlawueller
Thank you for your feedback. A reference to the documentation has been added to the ReadMe (see commit [69012de](https://github.com/RWTH-EBC/districtgenerator/commit/69012defee4387519af85724a75f3240999f09ec)). We are currently working on making the documentation available online and will add this in the next few days.
