+++ joss/paper.bib
@@ -104,10 +104,10 @@ keywords = {City district, Urban energy models, Geographic information systems, 
 	year = {2023},
 }
 
-@online{BBSR_2020,
+@misc{BBSR_2020,
 	title = {Ortsgenaue Testreferenzjahre (TRY) von Deutschland für mittlere und extreme Witterungsverhältnisse},
 	url = {https://www.bbsr.bund.de/BBSR/DE/forschung/programme/zb/Auftragsforschung/5EnergieKlimaBauen/2013/testreferenzjahre/01-start.htm},
-	organization  = {Bundesinstitut für Bau-, Stadt- und Raumforschung},
+	author  = {{Bundesinstitut für Bau-, Stadt- und Raumforschung}},
 	year = {2020},
 }
 
