### 2025-03-21 de6ca2b
chore: implement first instance of parallel demand calculation
