+++ districtgenerator/classes/datahandler.py
@@ -3,9 +3,13 @@
 import json
 import pickle
 import os
+import queue
 import sys
 import copy
 import datetime
+from threading import Thread
+from time import sleep
+
 import numpy as np
 import openpyxl
 import pandas as pd
@@ -447,78 +451,120 @@ class Datahandler:
         -------
         None.
         """
+        thread_queue = queue.Queue()
+        # Anzahl der maximalen worker Threads
+        max_worker = 8
+
 
         for building in self.district:
 
-            # calculate or load user profiles
-            if calcUserProfiles:
-                building["user"].calcProfiles(site=self.site,
-                                              holidays=self.time["holidays"],
-                                              time_resolution=self.time["timeResolution"],
-                                              time_horizon=self.time["dataLength"],
-                                              building=building,
-                                              path=os.path.join(self.resultPath, 'demands'))
+            t = Thread(target=self.generate_demands_worker, kwargs={"building":building, "calcUserProfiles":calcUserProfiles, "saveUserProfiles":saveUserProfiles})
+            thread_queue.put(t)
 
-                if saveUserProfiles:
-                    self.saveProfiles(name=building["unique_name"],
-                                      elec=building["user"].elec,
-                                      dhw= building["user"].dhw,
-                                      occ= building["user"].occ,
-                                      gains= building["user"].gains,
-                                      car= building["user"].car,
-                                      nb_flats= building["user"].nb_flats,
-                                      nb_occ= building["user"].nb_occ,
-                                      heatload= building["envelope"].heatload,
-                                      bivalent= building["envelope"].bivalent,
-                                      heatlimit= building["envelope"].heatlimit,
-                                      path=os.path.join(self.resultPath, 'demands'))
-                    #building["user"].saveProfiles(building["unique_name"], building["envelope"], os.path.join(self.resultPath, 'demands'))
 
-                print("Calculate demands of building " + building["unique_name"])
+        # Laufende Threads verfolgen
+        active_threads = []
 
-            else:
-                (building["user"].elec, building["user"].dhw,
-                 building["user"].occ, building["user"].gains,
-                 building["user"].car, building["user"].nb_flats,
-                 building["user"].nb_occ, building["envelope"].heatload,
-                 building["envelope"].bivalent,
-                 building["envelope"].heatlimit) = self.loadProfiles(building["unique_name"],
-                                                                     os.path.join(self.resultPath, 'demands'))
-                #building["user"].loadProfiles(building["unique_name"], os.path.join(self.resultPath, 'demands'))
-                print("Load demands of building " + building["unique_name"])
+        # Maximal 10 Threads gleichzeitig starten
 
-            # check if EV exist
-            building["clusteringData"] = {
-                "potentialEV": copy.deepcopy(building["user"].car)
-            }
-            building["user"].car *= building["buildingFeatures"]["EV"]
+        # Starte die ersten 10 Threads
+        for _ in range(min(thread_queue.qsize(), max_worker)):
+            t = thread_queue.get()
+            t.start()
+            sleep(1)
+            active_threads.append(t)
 
-            building["envelope"].calcNormativeProperties(self.site["SunRad"], building["user"].gains)
+        # Überwachung der Threads
+        while active_threads:
+            for t in active_threads:
+                if not t.is_alive():
+                    active_threads.remove(t)
+                    if not thread_queue.empty():
+                        new_thread = thread_queue.get()
+                        new_thread.start()
+                        active_threads.append(new_thread)
+            sleep(1)
 
-            night_setback = building["buildingFeatures"]["night_setback"]
 
-            # calculate or load heating profiles
-            if calcUserProfiles:
-                building["user"].calcHeatingProfile(site=self.site,
-                                                    envelope=building["envelope"],
-                                                    night_setback=night_setback,
[…]
