### 2025-03-21 fe0a636
chore: implement first instance of parallel demand calculation
