### 2025-03-24 RWTH-EBC/districtgenerator#58 @snjsomnath
The problem was in the way I was trying to access the library in the notebook. Yes, this issue is now resolved
