### 2025-03-25 6faaeb6
add sphinxdoc in gitlab-ci.yml
