+++ .gitlab-ci.yml
@@ -35,6 +35,8 @@ include:
     file: 'python/tests/coverage.gitlab-ci.yml'
   - project: 'EBC/EBC_all/gitlab_ci/templates'
     file: 'python/pypi-release/release.gitlab-ci.yml'
+  - project: 'EBC/EBC_all/gitlab_ci/templates'
+    file: 'sphinxdoc/sphinxdoc.gitlab-ci.yml'
   #- project: 'EBC/EBC_all/gitlab_ci/templates'
   #  file: 'python/examples/examples.gitlab-ci.yml'
   - template: Dependency-Scanning.gitlab-ci.yml
