### 2025-03-25 e968dc6
remove unused docs config
