### 2025-03-25 c767435
manual add source files for docs
