+++ docs/master/_sources/index.rst.txt
@@ -1,0 +1,17 @@
+.. districtgenerator documentation documentation master file, created by
+   sphinx-quickstart on Fri Oct  4 14:00:50 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to the documentation of the DistrictGenerator!
+===========================================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   README
+   modules
+   EXAMPLES
+   about/index
+
