+++ docs/master/_sources/districtgenerator.classes.rst.txt
@@ -1,0 +1,42 @@
+Classes
+=============================
+
+districtgenerator.classes.datahandler
+--------------------------------------------
+
+.. automodule:: districtgenerator.classes.datahandler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+districtgenerator.classes.envelope
+-----------------------------------------
+
+.. automodule:: districtgenerator.classes.envelope
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+districtgenerator.classes.profils
+----------------------------------------
+
+.. automodule:: districtgenerator.classes.profils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+districtgenerator.classes.solar
+--------------------------------------
+
+.. automodule:: districtgenerator.classes.solar
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+districtgenerator.classes.users
+--------------------------------------
+
+.. automodule:: districtgenerator.classes.users
+   :members:
+   :undoc-members:
+   :show-inheritance:
