+++ docs/master/_sources/about/CHANGELOG.md.txt
@@ -1,0 +1,11 @@
+# Version History
+
+## v0.1.1
+- Reworked the structure of the library
+- Added documentation  
+- Added test file for functional testing
+
+## v0.1.0
+- Configuration via environment variables, json or local
+- Moved to github.com/RWTH-EBC
+- Bugfix
