+++ docs/master/_sources/about/index.rst.txt
@@ -1,0 +1,13 @@
+.. _about:
+
+=======================
+About DistrictGenerator
+=======================
+
+This sections is all about the non-technical stuff.
+
+.. toctree::
+   :maxdepth: 1
+
+   CHANGELOG
+   LICENSE
