+++ docs/master/_sources/modules.rst.txt
@@ -1,0 +1,8 @@
+Classes and internal functions
+==============================
+
+.. toctree::
+   :maxdepth: 2
+
+   districtgenerator.classes
+   districtgenerator.functions
