### 2025-03-25 8a011e6
Update paper.md - add Rawad Hamze as author
