+++ README.md
@@ -38,6 +38,7 @@ of the DistrictGenerator as follows:
   selection and penetrations. We thereby create a platform for early-stage comparison of neighborhood concepts 
   with the flexibility of selecting different variants, given the tool's rapid recalculations.
 
+
 ## Getting started
 
 ### Install the DistrictGenerator
@@ -51,10 +52,10 @@ and secondly run:
 pip install -e districtgenerator
 ```
 
-Once you have installed the DistrictGenerator, you can check the [examples](EXAMPLES.md) 
+Once you have installed the DistrictGenerator, you can check the [examples](./examples/) 
 to learn how to use the different components.
 
-### Minimum required input data
+### Minimum manual required input data
 
 To generate your district, you need to know some information about its buildings. 
 The minimal input data set was defined following the [TABULA archetype approach](https://webtool.building-typology.eu/#bm):
@@ -65,29 +66,76 @@ The minimal input data set was defined following the [TABULA archetype approach]
 - retrofit: retrofit state according to TABULA (0: existing state, 1: usual refurbishment, 2: advanced refurbishment)
 - area: reference floor area (given in square meters)
 
-The example.csv file can be used as [template](../districtgenerator/data/scenarios/example.csv).
+The example.csv file can be used as [template](./districtgenerator/data/scenarios/example.csv).
 
-### What you get
+### Additional input data
 
-After executing district generation you can find building-specific and time-dependent profiles in 
-the .csv format in folder results/demands. The results contain: 
+In the folder [data](./districtgenerator/data) 
+further data can be found. Default values are already stored there.
+Optionally, the following data can be changed to:
+- design_building_data.json: Maximum and minimum indoor temperature and the room ventilation rate
+- site_data.json: Test referece year and its conditions, 
+- time_data.json: The time resolution of the profiles 
+
+The data in the following files must not be changed because they are fixed parameters and external data:
+- design_weather_data.json: Contains the 16 German climate zones of the DWD. Climate zones are large areas 
+in which the main characteristics of the climate are the same. Each zone is represented by a city.
+- physics_data.json
+- dhw_stochstical.xlsx
+
+The weather data can be found in [weather](./districtgenerator/data/weather).
+
+## Structure of the DistrictGenerator
+
+![Library Structure](./docs/img/Struktur_Quartiersgenerator.png)
+
+## Workflow of the DistrictGenerator
+
+The district generator integrates multiple open-source tools and databases. 
+The figure below visualizes the dependencies of external tools and data with internal 
+functions. The user input for the parameterization of a neighborhood consists 
+of a minimum of data. First, the user enters the number of buildings and basic 
+information about each building, namely the building type, year of construction, 
+retrofit level, and net floor area. The number of buildings to be calculated is 
+not limited by the program. Optionally, the site of the district, the time 
+resolution of the profiles and the test reference year (TRY) for weather data 
+can be modified.
+
+![Library Structure](./docs/img/Workflow_DistrictGenerator.png)
+
+To obtain a fully parameterized building model, the TEASER tool performs a data 
+enrichment with data from the TABULA WebTool and uses statistical and normative 
+information about the building stock. Finally, the TEASER python package determines 
+the geometry and material properties of the buildings. As the TABULA WebTool 
+defines archetypal building properties for type, age class and retrofit level, the 
+generated districts are composed of representative buildings, making them ideal 
+for representative analyses or scalability studies. The richardsonpy tool generates 
+occupancy profiles that fluctuates pseudo-randomly throughout the day, mirroring the 
[…]
