### 2025-03-26 RWTH-EBC/districtgenerator#43 @carlawueller
Dear @jtock, the documentation is now available online (https://rwth-ebc.github.io/districtgenerator/master/docs/README.html). It is also linked in the ReadMe. I´ve also checked and edited all the links in the ReadMe on the master branch.
