### 2025-03-31 2bf88e5
data_handling reimplementation
