+++ build/lib/districtgenerator/classes/profils.py
@@ -1,0 +1,433 @@
+# -*- coding: utf-8 -*-
+
+import os
+import random as rd
+import json
+import math
+import numpy as np
+import pylightxl as xl
+import richardsonpy.classes.occupancy as occ_residential
+import richardsonpy.functions.change_resolution as cr
+import OpenDHW
+import districtgenerator.functions.change_resolution as chres
+
+class Profiles:
+    """
+    Profile class.
+    Calculating user related profiles of a building or flat.
+
+    Parameters
+    ----------
+    number_occupants : integer
+        Number of occupants who live in the house or flat.
+    number_occupants_building : integer
+        Number of occupants who live in the building.
+    initial_day : integer
+        Day of the week with which the generation starts.
+        1-7 for monday-sunday.
+    nb_days : integer
+        Number of days for which a stochastic profile is generated.
+    time_resolution : integer
+        resolution of time steps of output array in seconds.
+
+    Attributes
+    ----------
+    activity_profile : array-like
+        Numpy-array with active occupants 10-minutes-wise.
+    occ_profile : array-like
+        Stochastic occupancy profile.
+    app_load : array-like
+        Electric load profile of appliances in W.
+    light_load : array-like
+        Electric load profile of lighting in W.
+    """
+
+    def __init__(self, number_occupants, number_occupants_building, initial_day, nb_days, time_resolution, building):
+        """
+        Constructor of Profiles class.
+
+        Returns
+        -------
+        None.
+        """
+
+        self.number_occupants = number_occupants
+        self.number_occupants_building = number_occupants_building
+        self.initial_day = initial_day
+        self.nb_days = nb_days
+        self.time_resolution = time_resolution
+
+        self.building = building
+        self.activity_profile = []
+        self.occ_profile = []
+        self.occ_profile_building = []
+        self.building_profiles = {}
+        self.temperature_difference = []
+        self.light_load = []
+        self.app_load = []
+
+        self.generate_activity_profile_residential()
+
+    def generate_activity_profile_residential(self):
+        """
+        Generate a stochastic activity profile
+        (on base of ridchardsonpy).
+
+        Parameters
+        ----------
+        number_occupants : integer
+            Number of occupants who live in the house or flat.
+        initial_day : integer
+            Day of the week with which the generation starts
+            1-7 for monday-sunday.
+        nb_days : integer
+            Number of days for which a stochastic profile is generated.
+
+        Returns
+        -------
+        None.
+        """
+
+        if self.building in {"SFH", "TH", "MFH", "AB"}:
+            activity = occ_residential.Occupancy(self.number_occupants, self.initial_day, self.nb_days)
+            self.activity_profile = activity.occupancy
+
+    def generate_occupancy_profiles_residential(self):
+        """
+        Generate stochastic occupancy profiles for a district for calculating internal gains.
+        Change time resolution of 10 min profiles to required resolution.
+
+        Parameters
+        ----------
+        time_resolution : integer
+            Resolution of time steps of output array in seconds.
+        activity_profile : array-like
+            Numpy-arry with active occupants 10-minutes-wise.
+
+        Returns
+        -------
+        self.occ_profile : array-like
+            Number of present occupants.
+        """
+
+        tr_min = int(self.time_resolution/60)
+        sia_profile_daily_min = np.concatenate((np.ones(60*8),
+                                                np.zeros(60*13),
+                                                np.ones(60*3)),
+                                                axis=None)
+
+        # generate array for minutely profile
[…]
