+++ build/lib/districtgenerator/functions/opti_dimensioning_central_devices.py
@@ -1,0 +1,780 @@
+# -*- coding: utf-8 -*-
+
+"""
+
+EHDO - ENERGY HUB DESIGN OPTIMIZATION Tool
+
+Developed by:   E.ON Energy Research Center,
+                Institute for Energy Efficient Buildings and Indoor Climate,
+                RWTH Aachen University,
+                Germany
+
+Contact:        Marco Wirtz
+                marco.wirtz@eonerc.rwth-aachen.de
+
+"""
+
+import gurobipy as gp
+import numpy as np
+import time
+import os
+#from optim_app.help_functions import create_excel_file
+
+
+def run_optim(data, devs, param, dem, result_dict):
+
+    #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    # Load model parameters
+    start_time = time.time()
+
+    days = range(data.time["clusterNumber"])
+    # calculate cluster time horizon
+    clusterHorizon = int((data.time["clusterLength"] / data.time["timeResolution"]))
+    time_steps = range(clusterHorizon)
+    year = range(52)
+
+    # Get sigma function which assigns every day of the year to a design day
+    sigma = param["sigma"]
+
+    # Create set for devices
+    all_devs = ["PV", "WT", "STC", "WAT",
+                "HP", "EB", "CC", "AC",
+                "CHP", "BOI", "GHP",
+                "BCHP", "BBOI", "WCHP", "WBOI",
+                "ELYZ", "FC", "H2S", "SAB",
+                "TES", "CTES", "BAT", "GS",
+                ]
+
+    #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+    # Set up model and create variables
+
+    # Create a new model
+    model = gp.Model("Energy_hub_model")
+
+    # Device's capacity (i.e. rated power)
+    cap = {}
+    for device in all_devs:
+        cap[device] = model.addVar(vtype="C", name="nominal_capacity_" + str(device))
+
+    # Roof area used for PV and solar thermal collector installation
+    area = {}
+    for device in ["PV", "STC"]:
+        area[device] = model.addVar(vtype = "C", name="roof_area_" + str(device))
+
+    # Gas flow to/from devices
+    gas = {}
+    for device in ["CHP", "BOI", "GHP", "SAB", "from_grid", "to_grid"]:
+        gas[device] = {}
+        for d in days:
+            gas[device][d] = {}
+            for t in time_steps:
+                gas[device][d][t] = model.addVar(vtype="C", name="gas_" + device + "_d" + str(d) + "_t" + str(t))
+
+    # Electric power to/from devices
+    power = {}
+    for device in ["PV", "WT", "WAT", "HP", "EB", "CC", "CHP", "BCHP", "WCHP", "ELYZ", "FC", "from_grid", "to_grid"]:
+        power[device] = {}
+        for d in days:
+            power[device][d] = {}
+            for t in time_steps:
+                power[device][d][t] = model.addVar(vtype="C", name="power_" + device + "_d" + str(d) + "_t" + str(t))
+
+    # Heat to/from devices
+    heat = {}
+    for device in ["STC", "HP", "EB", "AC", "CHP", "BOI", "GHP", "BCHP", "BBOI", "WCHP", "WBOI", "FC"]:
+        heat[device] = {}
+        for d in days:
+            heat[device][d] = {}
+            for t in time_steps:
[…]
