### 2025-03-31 da89e50
data_handling reimplementation
