+++ districtgenerator/classes/users.py
@@ -31,7 +31,7 @@ class Users:
         Building type according to TABULA database.
     nb_flats : integer
         Number of flats in building.
-    annual_el_demand : array-like
+    annual_el_demand_per_flat : array-like
         Annual electricity consumption in dependency of the building type and the number of occupants.
     lighting_index : integer
         This index defines the lighting configuration of the household.
@@ -51,7 +51,7 @@ class Users:
         Heat demand for each building.
     """
 
-    def __init__(self, building, area, elec):
+    def __init__(self, building, area):
         """
         Constructor of Users class.
 
@@ -62,6 +62,7 @@ class Users:
 
         self.building = building
         self.nb_flats = None
+        self.annual_el_demand_per_flat = None
         self.annual_el_demand = None
         self.annual_heat_demand = None
         self.annual_dhw_demand = None
@@ -71,7 +72,7 @@ class Users:
         self.nb_occ = []
         self.occ = None
         self.dhw = None
-        self.elec = elec
+        self.elec = None
         self.gains = None
         self.heat = None
         self.cooling = None
@@ -238,14 +239,15 @@ class Users:
 
             probabilities = [0.143, 0.143, 0.143, 0.142, 0.143, 0.143, 0.143]
 
-            self.annual_el_demand = np.zeros(self.nb_flats)
+            self.annual_el_demand_per_flat = np.zeros(self.nb_flats)
+            self.annual_el_demand = 0
             for j in range(self.nb_flats):
                 if self.building == "SFH":
                     random_nb = rd.random()  # picking random number in [0,1) to decide between which 2 values of consumption_range the annual electricity consumption lies
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -254,7 +256,7 @@ class Users:
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1],consumption_range["SFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["SFH"][self.nb_occ[j]][i - 1], consumption_range["SFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -263,7 +265,7 @@ class Users:
                     i = 1
                     while i <= 7:
                         if random_nb < sum(probabilities[:i]):
-                            self.annual_el_demand[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
+                            self.annual_el_demand_per_flat[j] = rd.randint(consumption_range["MFH"][self.nb_occ[j]][i - 1], consumption_range["MFH"][self.nb_occ[j]][i])
                             # A random integer is selected as the current demand, which must lie between the two values determined by the first random number
                             break
                         i += 1
@@ -272,10 +274,12 @@ class Users:
                     i = 1
                     while i <= 7:
[…]
