### 2025-04-01 af1a409
Added automatic generation for central and decentral devices
