+++ districtgenerator/data_handling/central_device_config.py
@@ -1,0 +1,232 @@
+from dataclasses import dataclass
+
+@dataclass
+class CentralDeviceConfig:
+    # PV parameters
+    PV_feasible: bool = True
+    PV_eta: float = 0.15
+    PV_life_time: int = 25
+    PV_inv_var: int = 1000
+    PV_cost_om: float = 0.02
+    PV_max_area: int = 10000
+    PV_min_area: int = 100
+    PV_G_stc: int = 1
+
+    # WT parameters
+    WT_feasible: bool = True
+    WT_inv_var: int = 1500
+    WT_life_time: int = 20
+    WT_cost_om: float = 0.015
+    WT_min_cap: int = 100
+    WT_max_cap: int = 3000
+    WT_h_coeff: float = 0.2
+    WT_hub_h: int = 100
+    WT_ref_h: int = 10
+    WT_norm_power: float = 0.85
+
+    # WAT parameters
+    WAT_feasible: bool = False
+    WAT_inv_var: int = 2000
+    WAT_life_time: int = 30
+    WAT_cost_om: float = 0.01
+    WAT_min_cap: int = 50
+    WAT_max_cap: int = 2000
+    WAT_potential: int = 50000
+
+    # STC parameters
+    STC_feasible: bool = False
+    STC_eta: float = 0.7
+    STC_inv_var: int = 800
+    STC_life_time: int = 20
+    STC_cost_om: float = 0.02
+    STC_max_area: int = 5000
+    STC_min_area: int = 10
+    STC_G_stc: int = 1
+
+    # CHP parameters
+    CHP_feasible: bool = False
+    CHP_inv_var: int = 1200
+    CHP_eta_el: float = 0.4
+    CHP_eta_th: float = 0.5
+    CHP_life_time: int = 20
+    CHP_cost_om: float = 0.03
+    CHP_min_cap: int = 50
+    CHP_max_cap: int = 1000
+
+    # BOI parameters
+    BOI_feasible: bool = False
+    BOI_inv_var: int = 500
+    BOI_eta_th: float = 0.9
+    BOI_life_time: int = 20
+    BOI_cost_om: float = 0.02
+    BOI_min_cap: int = 10
+    BOI_max_cap: int = 500
+
+    # GHP parameters
+    GHP_feasible: bool = False
+    GHP_inv_var: int = 1000
+    GHP_COP: float = 3.5
+    GHP_life_time: int = 20
+    GHP_cost_om: float = 0.02
+    GHP_min_cap: int = 10
+    GHP_max_cap: int = 500
+
+    # HP parameters
+    HP_feasible: bool = True
+    HP_CCOP_feasible: bool = False
+    HP_ASHP_feasible: bool = False
+    HP_CSV_feasible: bool = True
+    HP_inv_var: int = 800
+    HP_life_time: int = 20
+    HP_cost_om: float = 0.015
+    HP_min_cap: int = 10
+    HP_max_cap: int = 500
+    HP_ASHP_carnot_eff: float = 0.4
+    HP_ASHP_supply_temp: int = 60
+    HP_COP_const: int = 4
+
+    # EB parameters
+    EB_feasible: bool = True
+    EB_inv_var: int = 600
+    EB_eta_th: float = 0.99
+    EB_life_time: int = 20
+    EB_cost_om: float = 0.01
+    EB_min_cap: int = 10
+    EB_max_cap: int = 300
+
+    # CC parameters
+    CC_feasible: bool = True
+    CC_inv_var: int = 700
+    CC_COP: float = 3.5
+    CC_life_time: int = 20
+    CC_cost_om: float = 0.02
[…]
