+++ build/lib/districtgenerator/data_handling/decentral_device_data_generator.py
@@ -1,0 +1,383 @@
+from typing import Any, Dict, List
+from districtgenerator.data_handling.decentral_device_config import DecentralDeviceConfig
+
+class DecentralDeviceDataGenerator:
+    def __init__(self, tech_config: DecentralDeviceConfig):
+        self.config = tech_config
+
+        # Mapping of device abbreviations to their meta data.
+        self.devices_meta: Dict[str, Dict[str, Any]] = {
+            "HP": {
+                "device": "heat pump",
+                "abbreviation": "HP",
+                "description": "Air-to-water heat pump.",
+                "specifications": {
+                    "grade": {
+                        "unit": "-",
+                        "description": "Quality grade. Ratio of the achieved coefficient of performance to the Carnot coefficient of performance."
+                    },
+                    "life_time": {
+                        "unit": "years",
+                        "description": "Maximum life time."
+                    },
+                    "inv_var": {
+                        "unit": "€/liter",
+                        "description": ""
+                    },
+                    "cost_om": {
+                        "unit": "percentage",
+                        "description": ""
+                    }
+                }
+            },
+            "EH": {
+                "device": "electric heater",
+                "abbreviation": "EH",
+                "description": "Electric heating device for bivalent operation in combination with the heat pump.",
+                "specifications": {
+                    "eta_th": {
+                        "unit": "-",
+                        "description": "Thermal efficiency."
+                    }
+                }
+            },
+            "BOI": {
+                "device": "gas boiler",
+                "abbreviation": "BOI",
+                "description": "Gas boiler.",
+                "specifications": {
+                    "eta_th": {
+                        "unit": "-",
+                        "description": "Thermal efficiency."
+                    },
+                    "life_time": {
+                        "unit": "years",
+                        "description": "Maximum life time."
+                    },
+                    "inv_var": {
+                        "unit": "€/kW",
+                        "description": ""
+                    },
+                    "cost_om": {
+                        "unit": "percentage",
+                        "description": ""
+                    }
+                }
+            },
+            "CHP": {
+                "device": "combined heat and power plant",
+                "abbreviation": "CHP",
+                "description": "Gas based combined heat and power plant.",
+                "specifications": {
+                    "eta_th": {
+                        "unit": "-",
+                        "description": "Thermal efficiency."
+                    },
+                    "eta_el": {
+                        "unit": "-",
+                        "description": "Electrical efficiency."
+                    },
+                    "life_time": {
+                        "unit": "years",
+                        "description": "Maximum life time."
+                    },
+                    "inv_var": {
+                        "unit": "€/kW",
+                        "description": ""
+                    },
+                    "cost_om": {
+                        "unit": "percentage",
+                        "description": ""
+                    }
+                }
+            },
+            "FC": {
+                "device": "fuel cell",
+                "abbreviation": "FC",
+                "description": "Gas based fuel cell.",
+                "specifications": {
+                    "eta_th": {
+                        "unit": "-",
+                        "description": "Thermal efficiency."
+                    },
+                    "eta_el": {
+                        "unit": "-",
+                        "description": "Electrical efficiency."
+                    }
+                }
+            },
+            "PV": {
+                "device": "photovoltaics",
+                "abbreviation": "PV",
+                "description": "photovoltaics",
+                "specifications": {
+                    "area_real": {
+                        "unit": "squaremeters",
+                        "description": "Module area."
+                    },
+                    "eta_el_ref": {
+                        "unit": "-",
+                        "description": "Electrical efficiency under reference conditions."
+                    },
+                    "t_cell_ref": {
+                        "unit": "degree Celsius",
+                        "description": "Reference ambient air temperature."
+                    },
+                    "G_ref": {
+                        "unit": "Watt per squaremeter",
+                        "description": "Reference solar irradiance."
+                    },
[…]
