### 2025-04-01 fdd9627
fixed file path in main_data_generator.py and missing requirements.txt
