+++ build/lib/districtgenerator/data_handling/EHDO_generator.py
@@ -1,212 +1,0 @@
-from typing import Any, Dict, List
-from districtgenerator.data_handling.config import EHDOConfig  # Import the config dataclass
-
-class EHDOGenerator:
-    def __init__(self, EHDO_config: EHDOConfig):
-        self.config = EHDO_config
-
-    def generate_energy_data(self) -> List[Dict[str, Any]]:
-        return [
-            {
-                "name": "enable_supply_el",
-                "value": self.config.enable_supply_el,
-                "description": "Enable electricity supply."
-            },
-            {
-                "name": "enable_feed_in_el",
-                "value": self.config.enable_feed_in_el,
-                "description": "Enable electricity supply."
-            },
-            {
-                "name": "enable_price_cap_el",
-                "value": self.config.enable_price_cap_el,
-                "unit": self.config.unit_placeholder,
-                "description": "Enable electricity capacity price."
-            },
-            {
-                "name": "price_cap_el",
-                "value": self.config.price_cap_el,
-                "unit": "€/kW",
-                "description": "Electricity capacity price."
-            },
-            {
-                "name": "enable_cap_limit_el",
-                "value": self.config.enable_cap_limit_el,
-                "unit": self.config.unit_placeholder,
-                "description": "Consider capacity of grid connection."
-            },
-            {
-                "name": "cap_limit_el",
-                "value": self.config.cap_limit_el,
-                "unit": "kW",
-                "description": "Capacity of grid connection."
-            },
-            {
-                "name": "enable_supply_limit_el",
-                "value": self.config.enable_supply_limit_el,
-                "unit": self.config.unit_placeholder,
-                "description": "Enable restriction of electricity demand from grid."
-            },
-            {
-                "name": "supply_limit_el",
-                "value": self.config.supply_limit_el,
-                "unit": "MWh/year",
-                "description": "Restrict electricity demand from grid."
-            },
-            {
-                "name": "enable_supply_gas",
-                "value": self.config.enable_supply_gas,
-                "description": "Enable gas supply."
-            },
-            {
-                "name": "enable_price_cap_gas",
-                "value": self.config.enable_price_cap_gas,
-                "unit": self.config.unit_placeholder,
-                "description": "Enable gas capacity price."
-            },
-            {
-                "name": "price_cap_gas",
-                "value": self.config.price_cap_gas,
-                "unit": "€/kWh",
-                "description": "Gas capacity price."
-            },
-            {
-                "name": "enable_feed_in_gas",
-                "value": self.config.enable_feed_in_gas,
-                "unit": self.config.unit_dash,
-                "description": "Enable natural gas feed-in."
-            },
-            {
-                "name": "revenue_feed_in_gas",
-                "value": self.config.revenue_feed_in_gas,
-                "unit": "€/kWh",
-                "description": "Revenue for natural gas feed-in ."
-            },
-            {
-                "name": "enable_cap_limit_gas",
-                "value": self.config.enable_cap_limit_gas,
-                "unit": self.config.unit_dash,
-                "description": "Restrict gas demand from grid."
-            },
-            {
-                "name": "cap_limit_gas",
-                "value": self.config.cap_limit_gas,
-                "unit": "MWh/year",
-                "description": "Maximum annual energy drawn from the gas grid."
-            },
-            {
-                "name": "enable_supply_biomass",
-                "value": self.config.enable_supply_biomass,
-                "unit": self.config.unit_dash,
-                "description": "Restrict available biomass."
-            },
-            {
-                "name": "enable_supply_limit_biomass",
-                "value": self.config.enable_supply_limit_biomass,
-                "unit": self.config.unit_dash,
-                "description": "Enable limit annual biomass import."
-            },
-            {
-                "name": "supply_limit_biomass",
[…]
