+++ districtgenerator/classes/KPIs.py
@@ -283,14 +283,14 @@ class KPIs:
     def calc_annual_cost_total(self, scenario, decentral_device_data, district, physics):
 
         # Count occurrences of "BOI", "HP", and "CHP" in the 'heater' column
-        heater_counts = scenario['heater'].value_counts()
-        heater_counts["TES"] = heater_counts.sum()
+        counts = scenario['heater'].value_counts()
 
-        # Sum the values in the 'PV', 'STC', 'EV', and 'BAT' columns
-        heater_counts["PV"] = scenario['PV'].sum()
-        heater_counts["STC"] = scenario['STC'].sum()
-        heater_counts["EV"] = scenario['EV'].sum()
-        heater_counts["BAT"] = scenario['BAT'].sum()
+        # Sum the values in the 'TES', 'PV', 'STC', 'EV', and 'BAT' columns
+        counts["TES"] = scenario['f_TES'].apply(lambda x: 1 if x > 0 else 0).sum()
+        counts["PV"] = scenario['f_PV'].apply(lambda x: 1 if x > 0 else 0).sum()
+        counts["STC"] = scenario['f_STC'].apply(lambda x: 1 if x > 0 else 0).sum()
+        counts["EV"] = scenario['EV'].sum()
+        counts["BAT"] = scenario['f_BAT'].apply(lambda x: 1 if x > 0 else 0).sum()
 
         capacities = {}
         for n in range(len(district)):
@@ -310,23 +310,23 @@ class KPIs:
             for dev in ["BOI", "HP", "CHP", "PV", "STC", "EV", "BAT", "TES"]:
                 calc_annual_investment[dev] = {}
                 try:
-                    if heater_counts[dev] > 0:
+                    if counts[dev] > 0:
                         calc_annual_investment[dev] = self.calc_annual_cost_device(decentral_device_data[dev],
                                                                                 decentral_device_data["inv_data"],
                                                                                 capacities[n][dev])
                     else: calc_annual_investment[dev] = 0
                 except:
                     calc_annual_investment[dev] = 0
-                    heater_counts[dev] = 0
+                    counts[dev] = 0
 
-        self.annual_investment_total = (calc_annual_investment["BOI"] * heater_counts["BOI"]
-                                        + calc_annual_investment["HP"] * heater_counts["HP"]
-                                        + calc_annual_investment["CHP"] * heater_counts["CHP"]
-                                        + calc_annual_investment["PV"] * heater_counts["PV"]
-                                        + calc_annual_investment["STC"] * heater_counts["STC"]
-                                        + calc_annual_investment["EV"] * heater_counts["EV"]
-                                        + calc_annual_investment["BAT"] * heater_counts["BAT"]
-                                        + calc_annual_investment["BAT"] * heater_counts["BAT"])
+        self.annual_investment_total = (calc_annual_investment["BOI"] * counts["BOI"]
+                                        + calc_annual_investment["HP"] * counts["HP"]
+                                        + calc_annual_investment["CHP"] * counts["CHP"]
+                                        + calc_annual_investment["PV"] * counts["PV"]
+                                        + calc_annual_investment["STC"] * counts["STC"]
+                                        + calc_annual_investment["EV"] * counts["EV"]
+                                        + calc_annual_investment["BAT"] * counts["BAT"]
+                                        + calc_annual_investment["BAT"] * counts["BAT"])
 
     def calc_annual_cost_device(self, dev, param, cap):
         """
@@ -418,7 +418,7 @@ class KPIs:
[…]
