### 2025-04-08 e677e92
Fiware post for EnergyInfo and EnergyScenario in udp_optimizer
