### 2025-04-11 2d3a0b0
Implementation of reading configs from dataclasses instead of JSON files.
