### 2025-04-12 1ee64b0
Completely eliminated reading from JSON files
