+++ districtgenerator/classes/solar.py
@@ -453,7 +453,7 @@ class Sun:
         # Return total radiation on a tilted surface
         return totalRadTiltSurface
 
-    def calcPVAndSTCProfile(self, time, site, area_roof, beta=35, gamma=0, usageFactorPV=0.4,
+    def calcPVAndSTCProfile(self, time, site, area_roof, devices, beta=35, gamma=0, usageFactorPV=0.4,
                             usageFactorSTC=0.2):
         """
         Computation of power profiles for photovoltaic (PV) collectors and solar thermal collectors (STC).
@@ -506,13 +506,13 @@ class Sun:
         # profile of the ambient temperature
         temperatureProfile = site["T_e"]
 
-        devices = {}
-        with open(os.path.join(self.filePath, 'decentral_device_data.json')) as json_file:
-            jsonData = json.load(json_file)
-            for subData in jsonData:
-                devices[subData["abbreviation"]] = {}
-                for subsubData in subData["specifications"]:
-                    devices[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
+        #devices = {}
+        #with open(os.path.join(self.filePath, 'decentral_device_data.json')) as json_file:
+        #    jsonData = json.load(json_file)
+        #    for subData in jsonData:
+        #        devices[subData["abbreviation"]] = {}
+        #        for subsubData in subData["specifications"]:
+        #            devices[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
 
         # calculate time variant PV efficiency
         eta_PV = np.zeros(time["timeSteps"])
