+++ districtgenerator/classes/optimizer.py
@@ -44,7 +44,7 @@ class Optimizer:
         Identifier of the currently regarded cluster.
     """
 
-    def __init__(self, data, cluster):
+    def __init__(self, data, cluster, config):
         """
         Constructor of Optimizer class.
 
@@ -60,15 +60,15 @@ class Optimizer:
         None.
         """
 
-        self.srcPath = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+        self.srcPath = os.path.dirname(os.path.dirname(os.path.abspath(__file__))) 
         self.filePath = os.path.join(self.srcPath, 'data')
-        self.optiSettings = self.loadGurobiSettings()
+        self.optiSettings = self.loadGurobiSettings(config)
         self.model = None
         self.data = data
         self.cluster = cluster
         self.initializeModel()
 
-    def loadGurobiSettings(self):
+    def loadGurobiSettings(self, config):
         """
         Load the optimization settings for the gurobi solver.
 
@@ -77,8 +77,11 @@ class Optimizer:
         optiSettings : dictionary
             Optimization settings for the gurobi solver.
         """
-
-        gurobiSettings = json.load(open(os.path.join(self.filePath, 'gurobi_settings.json')))
+        conf = config[0]
+        gurobiSettings = {}
+        for attr, value in conf.__dict__.items():
+            gurobiSettings[attr] = value
+        #gurobiSettings = json.load(open(os.path.join(self.filePath, 'gurobi_settings.json')))
 
         return gurobiSettings
 
