### 2025-04-12 b59aa3a
Small fix in KPIs.py
