### 2025-04-17 RWTH-EBC/districtgenerator#66 @jtock
I was able to find functions which are still undocumented. E.g. load_profiles(...) in dhw_stochastical.py or calc(...) in heating_profile_5R1C.py.
Please check again and make sure every function is covered.

E.g. calcHeatLoad(...) in envelope.py still has some comments in German.
