### 2025-04-17 RWTH-EBC/districtgenerator#67 @jtock
Thank you for the information!

I still see some necessary improvements in the documentation of the assumptions. You added to the README that some building information is created statistically. Please specify that even more. I suggest adding a chapter to the documentation explaining the assumptions. Which exactly do you use / what is created statistically? How is it done? Is there a reference?
I see that some parts of that are given as inline code comments, which is helpful, but not well placed. Please transfer this information to the documentation.
