### 2025-04-17 RWTH-EBC/districtgenerator#62 @jtock
You added the result types / files to the README. Please add occ which is still missing.

In the scenario files there are some more files next to example. There are a lot more input variables given as in the example.
id;building;year;retrofit;construction_type;night_setback;area;heater;PV;STC;EV;BAT;f_TES;f_BAT;f_EV;f_PV;f_STC;gamma_PV;ev_charging
Those input variables are neither explained in the documentation nor in the examples. Please add these function at both places.
