### 2025-04-17 RWTH-EBC/districtgenerator#64 @jtock
Dear @JoelSchoelzel, great that you added some more tests. With those you test for datatypes which is not checking yet wether the year is e.g. 1990 or (a wrongly 'calculated') 0.

Please create a test neighborhood scenario for which you can calculate the expected results. You can compare those expected results in the tests with the calculated ones.
With the test you want to know if the tool is behaving as you expect it. This can be effected e.g. by changes a developer makes or by changing dependencies.

You can find the JOSS guidelines to tests here: https://joss.readthedocs.io/en/latest/review_criteria.html#tests
