### 2025-04-23 cac07ea
Adding Norm Description to calcHeatLoad
