### 2025-04-28 292088c
Move to Pydantic configs instead of dataclasses
