+++ districtgenerator/data_handling/config.py
@@ -1,22 +1,22 @@
 from dataclasses import dataclass, field
-from typing import Any, Dict, List
-# dict = {
-#     "location_config": {
-#         "time_zone": 1, #hgrrg
-# }
+from typing import Any, Dict, List, ClassVar, Set
+from pydantic_settings import BaseSettings, SettingsConfigDict
 
-# dictonary class?
 
-@dataclass
-class LocationConfig:
+class LocationConfig(BaseSettings):
     timeZone: float = 1
     albedo: float = 0.2
     TRYYear: str = 'TRY2015'
     TRYType: str = 'Jahr'
     zip: str = '52078'
 
-    ALLOWED_TRY_YEARS = {"TRY2015", "TRY2045"}
-    ALLOWED_TRY_TYPES = {"Jahr", "Somm", "Wint"}
+    ALLOWED_TRY_YEARS: ClassVar[Set[str]] = {"TRY2015", "TRY2045"}
+    ALLOWED_TRY_TYPES: ClassVar[Set[str]] ={"Jahr", "Somm", "Wint"}
+
+    model_config = SettingsConfigDict(
+        env_file= ".locationconfig",
+        extra="forbid" 
+    )
 
     def __post_init__(self):
         if self.TRYYear not in self.ALLOWED_TRY_YEARS:
@@ -26,8 +26,7 @@ class LocationConfig:
         if not (0.0 <= self.albedo <= 1.0):
             raise ValueError("albedo must be between 0.0 and 1.0.")
 
-@dataclass
-class TimeConfig:
+class TimeConfig(BaseSettings):
     timeResolution: int = 3600
     clusterLength: int = 604800
     clusterNumber: int = 4
@@ -37,9 +36,12 @@ class TimeConfig:
     holidays2045: list = field(default_factory=lambda: [1, 97, 100, 121, 138, 149, 159, 276, 305, 358, 359, 360, 365],)
     initial_day_2015: list = field(default_factory=lambda: [4])
     initial_day_2045: list = field(default_factory=lambda: [6])
+    model_config = SettingsConfigDict(
+        env_file= ".timeconfig",
+        extra="forbid" 
+    )
 
-@dataclass
-class DesignBuildingConfig:
+class DesignBuildingConfig(BaseSettings):
     T_set_min: float = 20.0
     T_set_min_night: float = 18.0
     T_set_max: float = 23.0
@@ -53,9 +55,12 @@ class DesignBuildingConfig:
     retrofit_long: list = field(default_factory=lambda: ['tabula_standard', 'tabula_retrofit', 'tabula_adv_retrofit'])
     dhwload: list = field(default_factory=lambda: [4662.1, 4662.1, 4662.1, 3999.8])
     mean_drawoff_vol_per_day: list = field(default_factory=lambda: [40, 40, 40, 40])
+    model_config = SettingsConfigDict(
+        env_file= ".designbuildingconfig",
+        extra="forbid" 
+    )
 
-@dataclass
-class EcoConfig:
+class EcoConfig(BaseSettings):
     price_supply_el: float = 0.32
     revenue_feed_in_el: float = 0.0811
     price_supply_gas: float = 0.12
@@ -67,16 +72,22 @@ class EcoConfig:
     co2_biom: float = 0.35
     co2_waste: float = 0.0
     co2_hydrogen: float = 0.0
+    model_config = SettingsConfigDict(
+        env_file= ".ecoconfig",
+        extra="forbid" 
+    )
 
-@dataclass
-class PhysicsConfig:
+class PhysicsConfig(BaseSettings):
     rho_air: float = 1.2 # kg/m3
     c_p_air: float = 1000.0
     rho_water: float = 1000.0
[…]
