+++ districtgenerator/classes/KPIs.py
@@ -434,13 +434,13 @@ class KPIs:
         with open(os.path.join(filePath, 'eco_data.json')) as json_file:
             jsonData = json.load(json_file)
 
-        CO2_factor_el_grid = jsonData[3]["value"]  # Emi_elec_grid
-        CO2_factor_gas = jsonData[4]["value"]      # Emi_gas
-        CO2_factor_pv = jsonData[5]["value"]       # Emi_pv
+        CO2_factor_el_grid = next(item["value"] for item in jsonData if item["name"] == "co2_el_grid")  # Emi_elec_grid
+        CO2_factor_gas = next(item["value"] for item in jsonData if item["name"] == "co2_gas")      # Emi_gas
+        CO2_factor_pv = next(item["value"] for item in jsonData if item["name"] == "co2_PV")       # Emi_pv
 
         # change unit from [Wh] to [kWh] and consider time resolution --> in function "calculateEnergyExchangeGCP"
-        co2_dem_grid = self.W_dem_GCP_year * CO2_factor_el_grid
-        co2_gas = self.Gas_year * CO2_factor_gas
+        co2_dem_grid = self.W_dem_GCP_year * CO2_factor_el_grid / 1000    # in t/a
+        co2_gas = self.Gas_year * CO2_factor_gas / 1000                   # in t/a
 
         # caused CO2 emissions by PV [kg]
         co2_pv = 0
@@ -449,7 +449,7 @@ class KPIs:
                 try:
                     co2_pv += np.sum(self.inputData["district"][id]["generationPV_cluster"][c, :]
                                  * data.time["timeResolution"] / 3600 / 1000) * CO2_factor_pv \
-                          * self.inputData["clusterWeights"][self.inputData["clusters"][c]]
+                          * self.inputData["clusterWeights"][self.inputData["clusters"][c]]  / 1000     # in t/a
                 except KeyError:
                     co2_pv += 0
 
@@ -1036,13 +1036,13 @@ class KPIs:
         RH_values = tuple(struktur["Reihenhaus"].values())
         B_values = tuple(struktur["Block"].values())
 
-        certificate.drawString((table_width * 2 / 5) + (((table_width / 5) - (EFH_values[0] - 5) * 6.8) / 2) - 8,
+        certificate.drawString((table_width * 2 / 5) + ((table_width / 5) / 2) + 17 - (len(str(EFH_values[0])) * 6.5),
                                table_top - 14 - 18, str(EFH_values[0]))
-        certificate.drawString((table_width * 3 / 5) + (((table_width / 5) - (MFH_values[0] - 5) * 6.8) / 2) - 8,
+        certificate.drawString((table_width * 3 / 5) + ((table_width / 5) / 2) + 17 - (len(str(MFH_values[0])) * 6.5),
                                table_top - 14 - 18, str(MFH_values[0]))
-        certificate.drawString((table_width * 4 / 5) + (((table_width / 5) - (RH_values[0] - 5) * 6.8) / 2) - 8,
+        certificate.drawString((table_width * 4 / 5) + ((table_width / 5) / 2) + 17 - (len(str(RH_values[i])) * 6.5),
                                table_top - 14 - 18, str(RH_values[0]))
-        certificate.drawString((table_width) + (((table_width / 5) - (B_values[0] - 5) * 6.8) / 2) - 8,
+        certificate.drawString((table_width) + ((table_width / 5) / 2) + 17 - (len(str(B_values[i])) * 6.5),
                                table_top - 14 - 18, str(B_values[0]))
 
         for i in range(1, len(EFH_values)):
@@ -1149,56 +1149,61 @@ class KPIs:
             certificate.drawString(85, height - 36 - 6, "Liste der Gebäude")
 
         else:
-
[…]
