### 2025-05-02 f548845
update docstrings
