### 2025-05-02 RWTH-EBC/districtgenerator#67 @JoelSchoelzel
@jtock
All statistic data are from the TABULA webtool that we just load in using the TEASER tool. I would clarify this with links in the readme and documentation referring to the other tools with the corresponding sources and links, ok?
