### 2025-05-02 54b3052
update docstrings
