+++ docs/conf.py
@@ -8,7 +8,7 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../'))
+sys.path.insert(0, os.path.abspath('../'))
 
 project = 'DistrictGenerator'
 copyright = '2024, Joel Schoelzel'
