+++ README.md
@@ -3,7 +3,6 @@
 # DistrictGenerator
 
 [![License](http://img.shields.io/:license-mit-blue.svg)](http://doge.mit-license.org)
-[![Documentation](https://rwth-ebc.github.io/districtgenerator/master/docs/doc.svg)](https://rwth-ebc.github.io/districtgenerator/master/docs/README.html)
 
 Through the DistrictGenerator, we introduce an python-based open-source tool aimed at urban planners, energy suppliers,
 housing associations, engineering firms, architectural professionals, as well as academic and research institutions. 
@@ -52,7 +51,7 @@ and secondly run:
 pip install -e districtgenerator
 ```
 
-Once you have installed the DistrictGenerator, you can check the [examples](./examples/) 
+Once you have installed the DistrictGenerator, you can check the [examples](docs/EXAMPLES.md) 
 to learn how to use the different components.
 
 ### Minimum manual required input data
@@ -66,11 +65,11 @@ The minimal input data set was defined following the [TABULA archetype approach]
 - retrofit: retrofit state according to TABULA (0: existing state, 1: usual refurbishment, 2: advanced refurbishment)
 - area: reference floor area (given in square meters)
 
-The example.csv file can be used as [template](./districtgenerator/data/scenarios/example.csv).
+The example.csv file can be used as [template](../districtgenerator/data/scenarios/example.csv).
 
 ### Additional input data
 
-In the folder [data](./districtgenerator/data) 
+In the folder [data](https://github.com/RWTH-EBC/districtgenerator/tree/JOSS_submission/districtgenerator/data) 
 further data can be found. Default values are already stored there.
 Optionally, the following data can be changed to:
 - design_building_data.json: Maximum and minimum indoor temperature and the room ventilation rate
@@ -83,11 +82,11 @@ in which the main characteristics of the climate are the same. Each zone is repr
 - physics_data.json
 - dhw_stochstical.xlsx
 
-The weather data can be found in [weather](./districtgenerator/data/weather).
+The weather data can be found in [weather](https://github.com/RWTH-EBC/districtgenerator/tree/JOSS_submission/districtgenerator/data/weather).
 
 ## Structure of the DistrictGenerator
 
-![Library Structure](./docs/img/Struktur_Quartiersgenerator.png)
+![Library Structure](img/Struktur_Quartiersgenerator.png)
 
 ## Workflow of the DistrictGenerator
 
@@ -101,21 +100,27 @@ not limited by the program. Optionally, the site of the district, the time
 resolution of the profiles and the test reference year (TRY) for weather data 
 can be modified.
 
-![Library Structure](./docs/img/Workflow_DistrictGenerator.png)
+![Library Structure](img/Workflow_DistrictGenerator.png)
 
-To obtain a fully parameterized building model, the TEASER tool performs a data 
-enrichment with data from the TABULA WebTool and uses statistical and normative 
-information about the building stock. Finally, the TEASER python package determines 
-the geometry and material properties of the buildings. As the TABULA WebTool 
-defines archetypal building properties for type, age class and retrofit level, the 
+To obtain a fully parameterized building model, the [TEASER tool](https://rwth-ebc.github.io/TEASER/main/docs/index.html) 
+performs a data enrichment with data from the [TABULA WebTool](https://webtool.building-typology.eu/#bm) 
+that provides statistical and normative information about the building stock. 
+Finally, the TEASER python package determines the geometry and material properties of the buildings. 
+As the TABULA WebTool defines archetypal building properties for type, age class and retrofit level, the 
 generated districts are composed of representative buildings, making them ideal 
-for representative analyses or scalability studies. The richardsonpy tool generates 
-occupancy profiles that fluctuates pseudo-randomly throughout the day, mirroring the 
[…]
