### 2025-05-05 RWTH-EBC/districtgenerator#67 @jtock
Dear @JoelSchoelzel,
thank you for the suggestion. The solution seems fine for me as long as it gets clear which assumptions are covered by the tools you use.
