+++ districtgenerator.egg-info/SOURCES.txt
@@ -1,46 +1,0 @@
-LICENSE.md
-README.md
-setup.py
-districtgenerator/__init__.py
-districtgenerator.egg-info/PKG-INFO
-districtgenerator.egg-info/SOURCES.txt
-districtgenerator.egg-info/dependency_links.txt
-districtgenerator.egg-info/requires.txt
-districtgenerator.egg-info/top_level.txt
-districtgenerator/classes/KPIs.py
-districtgenerator/classes/__init__.py
-districtgenerator/classes/datahandler.py
-districtgenerator/classes/envelope.py
-districtgenerator/classes/optimizer.py
-districtgenerator/classes/plots.py
-districtgenerator/classes/profils.py
-districtgenerator/classes/solar.py
-districtgenerator/classes/system.py
-districtgenerator/classes/users.py
-districtgenerator/data_handling/EHDO_generator.py
-districtgenerator/data_handling/__init__.py
-districtgenerator/data_handling/building_data_generator.py
-districtgenerator/data_handling/central_device_config.py
-districtgenerator/data_handling/central_device_generator.py
-districtgenerator/data_handling/config.py
-districtgenerator/data_handling/decentral_device_config.py
-districtgenerator/data_handling/decentral_device_data_generator.py
-districtgenerator/data_handling/design_building_data_generator.py
-districtgenerator/data_handling/eco_data_generator.py
-districtgenerator/data_handling/gurobi_generator.py
-districtgenerator/data_handling/heat_grid_data_generator.py
-districtgenerator/data_handling/json_adapter.py
-districtgenerator/data_handling/main_data_generator.py
-districtgenerator/data_handling/physics_data_generator.py
-districtgenerator/data_handling/site_data_generator.py
-districtgenerator/data_handling/time_data_generator.py
-districtgenerator/functions/__init__.py
-districtgenerator/functions/change_resolution.py
-districtgenerator/functions/clustering_medoid.py
-districtgenerator/functions/heating_profile_5R1C.py
-districtgenerator/functions/k_medoids.py
-districtgenerator/functions/load_params_central_devices.py
-districtgenerator/functions/opti_central.py
-districtgenerator/functions/opti_dimensioning_central_devices.py
-districtgenerator/functions/wind_turbines.py
-tests/test_examples.py
