+++ districtgenerator/data_handling/config.py
@@ -1,6 +1,12 @@
 from dataclasses import dataclass, field
-from typing import Any, Dict, List, ClassVar, Set
+import os
+from typing import Any, Dict, List, ClassVar, Set, Optional
+from pydantic import BaseModel
 from pydantic_settings import BaseSettings, SettingsConfigDict
+from districtgenerator.data_handling.central_device_config import CentralDeviceConfig
+from districtgenerator.data_handling.decentral_device_config import DecentralDeviceConfig
+
+
 
 
 class LocationConfig(BaseSettings):
@@ -15,7 +21,7 @@ class LocationConfig(BaseSettings):
 
     model_config = SettingsConfigDict(
         env_file= ".locationconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
     def __post_init__(self):
@@ -38,7 +44,7 @@ class TimeConfig(BaseSettings):
     initial_day_2045: list = field(default_factory=lambda: [6])
     model_config = SettingsConfigDict(
         env_file= ".timeconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class DesignBuildingConfig(BaseSettings):
@@ -57,7 +63,7 @@ class DesignBuildingConfig(BaseSettings):
     mean_drawoff_vol_per_day: list = field(default_factory=lambda: [40, 40, 40, 40])
     model_config = SettingsConfigDict(
         env_file= ".designbuildingconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class EcoConfig(BaseSettings):
@@ -74,7 +80,7 @@ class EcoConfig(BaseSettings):
     co2_hydrogen: float = 0.0
     model_config = SettingsConfigDict(
         env_file= ".ecoconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class PhysicsConfig(BaseSettings):
@@ -84,7 +90,7 @@ class PhysicsConfig(BaseSettings):
     c_p_water: float = 4.18
     model_config = SettingsConfigDict(
         env_file= ".physicsconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class GurobiConfig(BaseSettings):
@@ -98,7 +104,7 @@ class GurobiConfig(BaseSettings):
     DualReductions: int = 0
     model_config = SettingsConfigDict(
         env_file= ".gurobiconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class HeatGridConfig(BaseSettings):
@@ -107,7 +113,7 @@ class HeatGridConfig(BaseSettings):
     delta_T_heatTransfer: float = 5 # Temperature difference in heat exchangers (K)
     model_config = SettingsConfigDict(
         env_file= ".heatgridconfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
 class EHDOConfig(BaseSettings):
@@ -164,49 +170,38 @@ class EHDOConfig(BaseSettings):
     unit_dash: str = "-"         # used for cases where unit is a dash
     model_config = SettingsConfigDict(
         env_file= ".ehdoonfig",
-        extra="forbid" 
+        extra="allow" 
     )
 
-class BuildingConfig(BaseSettings):
-    buildings: List[Dict[str, Any]]
-    heater_types: List[str]
-    night_setback: bool = False
-    PV: bool = False
-    STC: bool = False
-    EV: bool = False
-    BAT: bool = False
-    storage: Dict[str, float] = field(
-        default_factory=lambda: {
-            "f_TES": 35.0,
-            "f_BAT": 1.0,
-            "f_EV": 6000.0,
-            "f_PV": 0.4,
-            "f_STC": 0.04
-        }
+
+class GlobalConfig(BaseModel):
[…]
