+++ requirements.txt
@@ -14,3 +14,5 @@ seaborn==0.13.2
 OpenDHW>=0.2.2
 setuptools~=75.8.0
 reportlab~=3.6.13
+pydantic-settings >= 2.2.1
+pydantic >= 2.8.2
