### 2025-05-06 RWTH-EBC/districtgenerator#64 @JoelSchoelzel
Dear @jtock, thank you for the advice! Another problem is that we use richardsonpy as an external tool, which is based on stochastic methods. From this, the occupancy in the buildings and the household electricity demand are calculated. Both have a direct influence on the domestic hot water demand and the heating demand.
After discussion in the team, we realized that we can switch off the randomness in certain functions for the tests, but not for all functions, because these are located directly in the external tool richardsonpy.
Our proposal is that we determine expected values and possible deviations for domestic hot water, domestic electricity and heating demand and use them in the tests for verification. In your opinion, is this an acceptable procedure for checking the output of the district generator?
