### 2025-05-06 1a3ff8a
update readme and docs
