### 2025-05-09 RWTH-EBC/districtgenerator#66 @JoelSchoelzel
Dear @jtock, we checked again and every function is now covered
