### 2025-05-09 RWTH-EBC/districtgenerator#62 @JoelSchoelzel
Dear @jtock, we updated the documentation by fixing the links and adding the other points. Regarding the last point: this information had been pusht from another branch that are not necessary for this version. We updated the gitignore that such conflicts we be avoided in the future.
