### 2025-05-12 4977df7
Changed .env.CONFIG import to only use pydantic
