+++ examples/e4_generate_buildings.py
@@ -13,7 +13,7 @@ from districtgenerator.classes import *
 
 def example4_generate_buildings():
     # Initialize District
-    data = Datahandler(scenario_name = "example")
+    data = Datahandler(scenario_name = "example_set_1")
 
     # Generate Environment for the District
     data.generateEnvironment()
