+++ districtgenerator.egg-info/SOURCES.txt
@@ -17,23 +17,10 @@ districtgenerator/classes/profils.py
 districtgenerator/classes/solar.py
 districtgenerator/classes/system.py
 districtgenerator/classes/users.py
-districtgenerator/data_handling/EHDO_generator.py
 districtgenerator/data_handling/__init__.py
-districtgenerator/data_handling/building_data_generator.py
 districtgenerator/data_handling/central_device_config.py
-districtgenerator/data_handling/central_device_generator.py
 districtgenerator/data_handling/config.py
 districtgenerator/data_handling/decentral_device_config.py
-districtgenerator/data_handling/decentral_device_data_generator.py
-districtgenerator/data_handling/design_building_data_generator.py
-districtgenerator/data_handling/eco_data_generator.py
-districtgenerator/data_handling/gurobi_generator.py
-districtgenerator/data_handling/heat_grid_data_generator.py
-districtgenerator/data_handling/json_adapter.py
-districtgenerator/data_handling/main_data_generator.py
-districtgenerator/data_handling/physics_data_generator.py
-districtgenerator/data_handling/site_data_generator.py
-districtgenerator/data_handling/time_data_generator.py
 districtgenerator/functions/__init__.py
 districtgenerator/functions/change_resolution.py
 districtgenerator/functions/clustering_medoid.py
