+++ districtgenerator/classes/plot.py
@@ -1,0 +1,826 @@
+import matplotlib.pyplot as plt
+import numpy as np
+import os
+
+def el_buildings(plotData, data):
+
+    num_clusters = plotData["num_clusters"]
+    num_buildings = plotData["num_buildings"]
+
+    time_steps = plotData["time_steps"]
+    time = plotData["time"]
+
+    color_map = {
+        "Grid": "grey",
+        "PV": "yellow",
+        "CHP": "green",
+        "FC": "purple",
+        "BAT discharge": "blue",
+        "EV discharge": "cyan",
+        "Demand": "red",
+        "EV charge": "orange",
+        "HP": "magenta",
+        "EH": "brown",
+        "BAT charge": "lightblue",
+        "Feed-in": "black"
+    }
+
+    for n in range(num_buildings):
+
+        fig, axes = plt.subplots(num_clusters, 1, figsize=(12, 4 * num_clusters), sharex=True)
+
+        if num_clusters == 1:
+            axes = [axes]
+
+        # these two lists will collect _all_ the non‐zero
+        # handles & labels from _all_ clusters:
+        all_handles = []
+        all_labels  = []
+
+        for c in range(num_clusters):
+            ax = axes[c]
+
+            # positive (sources/generation)
+            buy = np.array(plotData["resultsOptimization"][c][n]["res_load"])
+
+            PV_power = safe_array(plotData, c, n, "PV", "P_el", time_steps)
+            CHP = safe_array(plotData, c, n, "CHP", "P_el", time_steps)
+            FC = safe_array(plotData, c, n, "FC", "P_el", time_steps)
+            dch_BAT = safe_array(plotData, c, n, "BAT", "dch", time_steps)
+            dch_EV = safe_array(plotData, c, n, "EV", "dch", time_steps)
+
+            # negative (demands/sinks)
+            feed = np.array(plotData["resultsOptimization"][c][n]["res_inj"])
+            demand = np.array(plotData["resultsOptimization"][c][n]["Demand"]["P_el"])
+
+            ch_EV = safe_array(plotData, c, n, "EV", "ch", time_steps)
+            HP = safe_array(plotData, c, n, "HP", "P_el", time_steps)
+            EH = safe_array(plotData, c, n, "EH", "P_el", time_steps)
+            ch_BAT = safe_array(plotData, c, n, "BAT", "ch", time_steps)
+
+            # SOURCES
+            sources = [buy, PV_power, CHP, FC, dch_BAT, dch_EV]
+            source_labels = ["Grid", "PV", "CHP", "FC", "BAT discharge", "EV discharge"]
+            source_colors = [color_map[label] for label in source_labels]
+
+            # SINKS
+            sinks = [demand, ch_EV, HP, EH, ch_BAT, feed]
+            sink_labels = ["Demand", "EV charge", "HP", "EH", "BAT charge", "Feed-in"]
+            sink_colors = [color_map[label] for label in sink_labels]
+
+            # plot everything (even the all‐zero ones) for stacking consistency
+            src_pc = ax.stackplot(time, sources, colors=source_colors, alpha=0.8)
+            snk_pc = ax.stackplot(time, [-s for s in sinks], colors=sink_colors, alpha=0.8)
+
+            # pick out just the non‐zero ones _in this cluster_
+            src_handles = []
+            src_labels = []
+            for arr, pc, lbl in zip(sources, src_pc, source_labels):
+                if np.any(arr > 0):
+                    src_handles.append(pc)
+                    src_labels.append(lbl)
+
+            snk_handles = []
+            snk_labels = []
+            for arr, pc, lbl in zip(sinks, snk_pc, sink_labels):
+                if np.any(arr > 0):
+                    snk_handles.append(pc)
[…]
