+++ districtgenerator/classes/KPIs.py
@@ -48,7 +48,8 @@ class KPIs:
         self.Gas_year = None
         self.dcf_year = None
         self.scf_year = None
-        self.annual_investment_total = None
+        self.annual_fixed_costs_decentral = None
+        self.annual_fixed_costs_central = None
         self.totalarea_residential = None
         self.totalarea_non_residential = None
         self.totalheatload = None
@@ -88,7 +89,7 @@ class KPIs:
         self.calculateEnergyExchangeWithinDistrict(data)
         self.calculateAutonomy()
         self.calculateCoverFactors(data)
-        self.calc_annual_cost_total(data.scenario, data.decentral_device_data, data.district, data.physics)
+        self.calc_annual_cost_total(data)
 
     def prepareData(self, data):
         """
@@ -280,16 +281,21 @@ class KPIs:
             self.scf_year += self.supplyCoverFactor[c] * (self.inputData["clusterWeights"][self.inputData["clusters"][c]]
                                                           / sum_ClusterWeights)
 
-    def calc_annual_cost_total(self, scenario, decentral_device_data, district, physics):
+    def calc_annual_cost_total(self, data):
 
-        # Count occurrences of "BOI", "HP", and "CHP" in the 'heater' column
+        scenario = data.scenario
+        decentral_device_data = data.decentral_device_data
+        district = data.district
+        physics = data.physics
+
+        # Count occurrences in the 'heater' column
         counts = scenario['heater'].value_counts()
 
         # Sum the values in the 'TES', 'PV', 'STC', 'EV', and 'BAT' columns
-        counts["TES"] = scenario['f_TES'].apply(lambda x: 1 if x > 0 else 0).sum()
+        counts["TES"] = scenario.apply(lambda row: 1 if (row['f_TES'] > 0 and row['heater'] != 'heat_grid') else 0,axis=1).sum()
         counts["PV"] = scenario['f_PV'].apply(lambda x: 1 if x > 0 else 0).sum()
         counts["STC"] = scenario['f_STC'].apply(lambda x: 1 if x > 0 else 0).sum()
-        counts["EV"] = scenario['EV'].sum()
+        counts["EV"] = sum((lambda ev: len(ev) if any(x > 0 for x in ev) else 0)(d["user"].ev_capacity)for d in district)
         counts["BAT"] = scenario['f_BAT'].apply(lambda x: 1 if x > 0 else 0).sum()
 
         capacities = {}
@@ -305,28 +311,29 @@ class KPIs:
             capacities[n]["TES"] = (district[n]["capacities"]["TES"] / physics["rho_water"] / physics["c_p_water"] /
                                     decentral_device_data["TES"]["T_diff_max"] * 3600)
 
-        for n in range(len(district)):
-            calc_annual_investment = {}
-            for dev in ["BOI", "HP", "CHP", "PV", "STC", "EV", "BAT", "TES"]:
-                calc_annual_investment[dev] = {}
-                try:
-                    if counts[dev] > 0:
-                        calc_annual_investment[dev] = self.calc_annual_cost_device(decentral_device_data[dev],
-                                                                                decentral_device_data["inv_data"],
-                                                                                capacities[n][dev])
-                    else: calc_annual_investment[dev] = 0
-                except:
-                    calc_annual_investment[dev] = 0
-                    counts[dev] = 0
+        calc_annual_investment = {dev: 0 for dev in ["BOI", "HP", "CHP", "PV", "STC", "EV", "BAT", "TES"]}
 
-        self.annual_investment_total = (calc_annual_investment["BOI"] * counts["BOI"]
-                                        + calc_annual_investment["HP"] * counts["HP"]
-                                        + calc_annual_investment["CHP"] * counts["CHP"]
[…]

[…]
