### 2025-05-13 3b2b7a2
Add deterministic testing
