+++ examples/e5_generate_demands.py
@@ -12,6 +12,7 @@ Then choose 'Modify Run Configuration' and tick 'Run with Python Console'.
 from districtgenerator.classes import *
 import pandas as pd
 import matplotlib.pyplot as plt
+import os
 
 def example5_generate_demands(rng=None):
 
@@ -46,7 +47,9 @@ def example5_generate_demands(rng=None):
 
     # We can now use the profiles for exemplary analyses like monthly demands or peak loads.
     # We plot the district space heat demand in kWh
-    exemplary_plot(data)
+    # Only show plot if not running under test
+    if "PYTEST_CURRENT_TEST" not in os.environ:
+        exemplary_plot(data)
 
     print("The demand profiles are generated and saved in the directory 'districtgenerator/results/demands/'.")
     print("In data.district[0].user for example you find the demand profiles.")
