### 2025-05-15 RWTH-EBC/districtgenerator#64 @RawadHamze
Dear @jtock ,

We’ve introduced deterministic behavior for testing purposes to ensure reproducibility while keeping the stochastic nature of the model intact for actual simulations. In the test, each building receives a fixed RNG seed to stabilize stochastic outputs. Electricity expectations are sourced from Stromspiegel; DHW and heat demand are based on averages from multiple runs. A ±10% tolerance accounts for natural variability. These changes are available in the branch 64-joss-review---testing. Let us know if further clarification is needed.
