### 2025-05-19 0a44570
correct internal gains
