### 2025-05-21 RWTH-EBC/districtgenerator#66 @jtock
Dear @JoelSchoelzel,
while checking #64 I found this docstring, explaining parameters which are not part of the function.
Please make sure you have correct docstrings, that new users are able to understand the functionality.

def generateBuildings(self):
        """
        Loads building envelope and user data for each building in the district.

        ----------

        Parameters
        ----------
            - building features (dict): Contains building-specific data.
            - number_of_floors (int): The number of floors in the building.
            - height_of_floors (float): The height of each floor in the building.

        Returns
        ----------
            - Creates and assigns building envelope data using the `Envelope` class.
            - Generates and assigns user profiles for each building using the `Users` class.
            - Computes design heat loads and assigns to each building.
        """
