### 2025-05-21 RWTH-EBC/districtgenerator#64 @RawadHamze
Dear @jtock,

Thank you very much for your detailed feedback and your helpful input.

Unfortunately, your suggestion does not solve the problem. The parameter randomize_appliances in richardsonpy only controls whether the available appliances inside a dwelling are randomized or not. While setting this to False helps in making the appliance configuration deterministic, it does not address the main source of non-determinism: the occupancy profile.

The hourly occupancy profile generated by richardsonpy (specifically, the presence of active occupants, i.e., those who are at home and not sleeping) is inherently stochastic. This behavior is embedded within the model and cannot be externally controlled. As a result, this randomness propagates into several other demand profiles that are directly dependent on occupant presence and activity.
To clarify the impact:
- Lighting demand is only generated when at least one occupant is active and present at home.
- Domestic hot water (DHW) usage is similarly driven by occupant presence and activity patterns.
- Internal heat gains depend on the number of active occupants, as each person is considered a heat source.
- Electricity demand for appliances also relies on occupant activity, since appliances are only used when people are home and active.

Therefore, even if the appliance set is deterministic, the timing and magnitude of appliance use and all other usage patterns remain influenced by the stochastic occupancy behavior.

To allow for reproducible testing despite this, We implemented a workaround within the Districtgenerator. We introduced a fixed seed that controls all variables we internally randomize. However, this seeding cannot affect the occupancy modeling in richardsonpy.
With this approach, we are able to fix the parameters we control and define acceptable output ranges for heating, electricity, and DHW profiles. This enables us to test the tool’s outputs meaningfully and check its functionality—even though the core occupancy model remains non-deterministic.

I hope this explanation makes everything clear. These tests should allow us to validate the Districtgenerator in a way that is consistent with the JOSS testing guidelines.
