### 2025-05-26 842949a
added insulation calculation in envelope.py
