+++ districtgenerator/classes/envelope.py
@@ -188,22 +188,6 @@ class Envelope:
         None.
         """
 
-        #TODO:
-        # We need to load the u-value properties from FIWARE to the prj, so they can be used here
-        # Idea: if there is a u-value (thermalTransmittance) in the FIWARE, use it instead of calculating it
-        # this also means, that the calculation has to "stay" to be a back-up
-        # BUT: we need to check if the u-value is correct, so we need to compare it with the calculated one (at least for now)
-        # .
-        # So: For first testing use e.g. the e4 until generate_buildings. -> so it doesn't calculate load profiles
-        # 1. implement the u-value import into the the example.csv, so adjusting needed in optimizer -> new branch
-        # 2. implement the import from the csv in the Datahandler (i think it is self.scenario, line 119), maybe there is no additional step necessary here
-        # 2. put the if / then / else decisions in the loadComponentProperties (the u-values are calculated for opaque
-        #       surfaces in line 409 and for the window in line 424)
-        # -> if there is a FIWARE u-value, use it. But save the calculated u-value in the prj or print it out, so we can compare the values
-        # .
-        # (after comparing the values)
-        # 3. check if the u-value is correct, e.g. compare with TABULA values (hgo will have a look at this)
-
         material_bind = prj.data.material_bind
         element_bind = prj.data.element_bind
 
@@ -254,7 +238,7 @@ class Envelope:
 
         comp = "wall"
         # WALLS: Materials and U-value
-        # TODO: thermalTransmittanceFacade
+        # thermalTransmittanceFacade
         for name, elem in element_bind.items():
             if "OuterWall" in name:
                 if elem["building_age_group"][0] <= self.construction_year <= \
@@ -275,7 +259,7 @@ class Envelope:
 
         comp = "roof"
         # ROOF: Materials and U-value
-        # TODO: thermalTransmittanceRoof
+        # thermalTransmittanceRoof
         for name, elem in element_bind.items():
             if "Rooftop" in name:
                 if elem["building_age_group"][0] <= self.construction_year <= \
@@ -296,7 +280,7 @@ class Envelope:
 
         comp = "floor"
         # FLOOR: Materials and U-value
-        # TODO: thermalTransmittanceFloor
+        # thermalTransmittanceFloor
         for name, elem in element_bind.items():
             if "GroundFloor" in name:
                 if elem["building_age_group"][0] <= self.construction_year <= \
@@ -317,7 +301,7 @@ class Envelope:
 
         comp = "intWall"
         # INTERNAL WALL: Materials and U-value
-        # TODO: We don't have that in the database, solution pending
+        # We don't have that in the database, solution pending
         for name, elem in element_bind.items():
             if "InnerWall" in name:
                 dummy = min(2015,
@@ -339,7 +323,7 @@ class Envelope:
 
         comp = "ceiling"
         # CEILING: Materials and U-value
-        # TODO: thermalTransmittanceCeiling
+        # thermalTransmittanceCeiling
         for name, elem in element_bind.items():
             if "Ceiling" in name:
                 dummy = min(2015,
@@ -361,7 +345,7 @@ class Envelope:
 
         comp = "intFloor"
         # INTERNAL FLOOR: Materials and U-value
-        # Todo: We don't have that in the database, solution pending
+        # We don't have that in the database, solution pending
         for name, elem in element_bind.items():
             if "Floor" in name:
                 dummy = min(2015,
@@ -413,9 +397,6 @@ class Envelope:
                 self.rho["opaque"][x],
                 self.cp["opaque"][x]
             )
-            # TODO: They calculate the u-value here, so this is where it need to be replaced
[…]
