### 2025-05-29 3d4545b
Changes in optimizer
