+++ examples/e8_scenario_evaluation - Kopie.py
@@ -28,9 +28,6 @@ def example8_scenario_evaluation():
     # Create a certificate (PDF) which summarizes the district parameters and calculated KPIs
     data.KPIs.create_certificate(data=data, result_path=data.resultPath)
 
-    # TODO: erzeuge Energieausweis
-
-
     print("Congratulations! You calculated an optimized device operation for the selected neighborhood!")
 
     return data
