### 2025-06-04 199abe8
added: some todos and questions regarding pv integration into districtgenerator
