### 2025-06-04 3f5ab4e
Add validation and test improvements
