+++ districtgenerator/classes/datahandler.py
@@ -147,15 +147,31 @@ class Datahandler():
                                     + self.scenario_name + ".csv",
                                     header=0, delimiter=";")
 
-        # initialize buildings for scenario
-        for id in self.scenario["id"]:
-            # create empty dict for observed building
-            building = {}
+        # Check for missing values in required columns
+        required_columns = ["id", "building", "year", "retrofit", "area"]
+        missing_columns = [col for col in required_columns if col not in self.scenario.columns]
+        if missing_columns:
+            raise ValueError(f"[ERROR] Missing required columns in input: {', '.join(missing_columns)}")
 
-            # store features of the observed building
-            building["buildingFeatures"] = self.scenario.loc[id]
+        for index, row in self.scenario.iterrows():
+            # Check for missing or unrealistic values
+            if pd.isnull(row["area"]) or row["area"] <= 50:
+                raise ValueError(
+                    f"[ERROR] Invalid area value at row {index}: '{row['area']}' – must be greater than 50 m².")
 
-            # append building to district
+            if pd.isnull(row["year"]) or row["year"] < 1800 or row["year"] > 2025:
+                raise ValueError(
+                    f"[ERROR] Unrealistic construction year at row {index}: '{row['year']}' – must be between 1800 and 2025.")
+
+            if pd.isnull(row["retrofit"]) or row["retrofit"] not in [0, 1, 2]:
+                raise ValueError(
+                    f"[ERROR] Invalid retrofit value at row {index}: '{row['retrofit']}' – expected 0, 1, or 2.")
+
+            if pd.isnull(row["building"]):
+                raise ValueError(f"[ERROR] Missing building type at row {index}.")
+
+            # Create building dictionary
+            building = {"buildingFeatures": row}
             self.district.append(building)
 
 
@@ -163,19 +179,17 @@ class Datahandler():
         """
         Loads building envelope and user data for each building in the district.
 
-        ----------
+        For each building defined in the scenario, the following steps are performed:
+        - The number of floors and floor height are estimated.
+        - A TEASER building model is created and its physical envelope is defined.
+        - Design heat loads are calculated.
+
+    ----------
 
         Parameters
-        ----------
-            - building features (dict): Contains building-specific data.
-            - number_of_floors (int): The number of floors in the building.
-            - height_of_floors (float): The height of each floor in the building.
-
-        Returns
-        ----------
-            - Creates and assigns building envelope data using the `Envelope` class.
-            - Generates and assigns user profiles for each building using the `Users` class.
-            - Computes design heat loads and assigns to each building.
+    ----------
+            - rng : Random object, optional
+                A seeded random number generator instance. If None, Python's default random is used.
         """
 
         rg = rng or rd
