### 2025-06-09 cd38a72
thickness saving split
