### 2025-06-10 #7657 @jtock
Dear @AdamRJensen and @JoelSchoelzel, the districtgenerator satisfies the JOSS requirements for me. I have now closed my checklist. The process can continue from my perspective.
