### 2025-06-23 d282ed8
Changes in optimizer
