### 2025-06-24 3310fc5
fix: implemented multiprocessing (testing required)
