+++ districtgenerator/classes/datahandler.py
@@ -9,6 +9,7 @@ import copy
 import datetime
 from threading import Thread
 from time import sleep
+import multiprocessing
 
 import numpy as np
 import time
@@ -249,7 +250,8 @@ class Datahandler:
         self.select_plz_data()
         # load weather data
         # select the correct file depending on the TRY weather station location
-        weatherData = np.loadtxt(os.path.join(self.filePath, "weather", "TRY_" + self.site["TRYYear"][-4:] + "_" + self.site["TRYType"])
+        weatherData = np.loadtxt(
+            os.path.join(self.filePath, "weather", "TRY_" + self.site["TRYYear"][-4:] + "_" + self.site["TRYType"])
             + "/"
             + self.site["TRYYear"] + "_"
             + str(self.site["Location"]) + "_" + str(self.site["TRYType"])
@@ -309,24 +311,26 @@ class Datahandler:
 
         # extract the needed values
         self.site["altitude"] = filtered_data.iloc[0]['Altitude']
-        self.site["location"] = [filtered_data.iloc[0]['Latitude'],filtered_data.iloc[0]['Longitude']]
-        self.site["T_ne"] = filtered_data.iloc[0]['T_ne'] # norm outside temperature for calculating the design heat load
-        self.site["T_me"] = filtered_data.iloc[0]['T_me'] # mean annual temperature for calculating the design heat load
+        self.site["location"] = [filtered_data.iloc[0]['Latitude'], filtered_data.iloc[0]['Longitude']]
+        self.site["T_ne"] = filtered_data.iloc[0][
+            'T_ne']  # norm outside temperature for calculating the design heat load
+        self.site["T_me"] = filtered_data.iloc[0][
+            'T_me']  # mean annual temperature for calculating the design heat load
 
         # Calculate solar irradiance per surface direction - S, W, N, E, Roof represented by angles gamma and beta
         global sun
         sun = Sun(filePath=self.filePath)
         self.site["SunRad"] = sun.getSolarGains(initialTime=0,
-                                        timeDiscretization=self.time["timeResolution"],
-                                        timeSteps=self.time["timeSteps"],
-                                        timeZone=self.site["timeZone"],
-                                        location=self.site["location"],
-                                        altitude=self.site["altitude"],
-                                        beta=[90, 90, 90, 90, 0],
-                                        gamma=[0, 90, 180, 270, 0],
-                                        beamRadiation=self.site["SunDirect"],
-                                        diffuseRadiation=self.site["SunDiffuse"],
-                                        albedo=self.site["albedo"])
+                                                timeDiscretization=self.time["timeResolution"],
+                                                timeSteps=self.time["timeSteps"],
+                                                timeZone=self.site["timeZone"],
+                                                location=self.site["location"],
+                                                altitude=self.site["altitude"],
+                                                beta=[90, 90, 90, 90, 0],
+                                                gamma=[0, 90, 180, 270, 0],
+                                                beamRadiation=self.site["SunDirect"],
+                                                diffuseRadiation=self.site["SunDiffuse"],
+                                                albedo=self.site["albedo"])
 
     def initializeBuildings(self):
         """
@@ -391,10 +395,12 @@ class Datahandler:
                 self.district.append(building)
 
                 # Count number of builidings to predict the approximate calculation time
-                if building["buildingFeatures"]["building"] == 'SFH' or building["buildingFeatures"]["building"] == 'TH':
-                    num_sfh +=1
-                elif building["buildingFeatures"]["building"] == 'MFH' or building["buildingFeatures"]["building"] == 'AB':
-                    num_mfh +=1
[…]
