### 2025-06-24 dbf5acf
fix: implemented multiprocessing (testing required)
