### 2025-06-24 #7657 @JoelSchoelzel
Dear @AdamRJensen, we now completed all the tasks from your list above. The DOI generated by Zenodo is 10.5281/zenodo.15729957. The published version has the number v0.1.1.
Can you already tell us when the paper with DOI will be published approximately? We are currently writing the final report of the project in which the neighborhood generator was developed and would like to reference the JOSS paper.
Best regards
Joel
