### 2025-06-24 4d05db7
fix: fixed multiprocessing for local calculation
