### 2025-06-24 0327a36
fix: fixed multiprocessing for local calculation
