+++ districtgenerator/classes/datahandler.py
@@ -366,7 +366,7 @@ class Datahandler:
                 building["gmlId"] = building["buildingFeatures"]["gmlId"]
 
                 if not self.scenario.thermalTransmittanceFacade.empty:
-                    self.u_values = (self.scenario.thermalTransmittanceFacade.iloc[id],
+                    building["buildingFeatures"]["thermalTransmittance"] = (self.scenario.thermalTransmittanceFacade.iloc[id],
                                                         self.scenario.thermalTransmittanceRoof.iloc[id],
                                                         self.scenario.thermalTransmittanceFloor.iloc[id],
                                                         self.scenario.thermalTransmittanceWindow.iloc[id])
@@ -792,72 +792,158 @@ class Datahandler:
                           design_building_data=self.design_building_data,
                           file_path=self.filePath)
 
-            # Define the path for the new PV values log
-            pv_log_path = os.path.join(self.filePath, "logs", "pv_values_log.csv")
+            # Parse |-separated strings into lists
+            if not pd.isna(building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"]):
+                roof_areas = [float(area) for area in
+                              building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"].split("|")]
+                roof_inclinations = [float(beta) for beta in building["buildingFeatures"]["roofInclination"].split("|")]
+                cardinal_directions = [float(gamma) for gamma in
+                                       building["buildingFeatures"]["cardinalDirection"].split("|")]
+                roof_shapes = building["buildingFeatures"]["roofShape"].split("|")
 
-            pv_row = {
-            "ID": building["buildingFeatures"]["gmlId"],
-            }
+                # Initialize total potentials
+                total_potentialPV = None
+                total_potentialSTC = None
 
-            # Create a new dictionary for PV values logging
-            pv_row.update({
-                "calculated_area_roof": building["envelope"].A["opaque"]["roof"],  # Calculated value
-                "actual_area_roof": building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"],  # Actual value
-                "calculated_beta": 35,  # Calculated value
-                "actual_beta": building["buildingFeatures"]["roofInclination"],  # Actual value
-                "calculated_gamma": building["buildingFeatures"]["gamma_PV"],  # Calculated value
-                "actual_gamma": building["buildingFeatures"]["cardinalDirection"],  # Actual value
-                "roofShape": building["buildingFeatures"]["roofShape"]
-            })
+                pv_rows = []
+                individual_profiles = []
 
-            try:
-                # Attempt to read the existing PV values log
-                df_existing_pv = pd.read_csv(pv_log_path)
-                # Concatenate the new row with the existing DataFrame
-                df_new_pv = pd.concat([df_existing_pv, pd.DataFrame([pv_row])], ignore_index=True)
-            except FileNotFoundError:
-                # If the file does not exist, create a new DataFrame
-                df_new_pv = pd.DataFrame([pv_row])
+                pv_log_path = os.path.join(self.filePath, "logs", "pv_values_log.csv")
 
-            # Write the updated DataFrame to the new CSV log
-            df_new_pv.to_csv(pv_log_path, index=False)
+                # Calculate and save for each roof surface
+                for i in range(len(roof_areas)):
+                    # Create PV log entry for this roof
+                    pv_row = {
+                        "ID": f"{building['buildingFeatures']['gmlId']}_roof_{i + 1}",
+                        "calculated_area_roof": building["envelope"].A["opaque"]["roof"],  # Calculated value
+                        "actual_area_roof": roof_areas[i],  # Actual value
+                        "calculated_beta": 35,  # Calculated value
+                        "actual_beta": roof_inclinations[i],  # Actual value
[…]
