### 2025-07-06 f83c8a1
Removed AI-X Heat changes
