### 2025-07-06 d0643b8
Removed AI-X Heat changes
