+++ districtgenerator/classes/solar.py
@@ -453,8 +453,8 @@ class Sun:
         # Return total radiation on a tilted surface
         return totalRadTiltSurface
 
-    def calcPVAndSTCProfile(self, time, site, area_roof, devices, beta=35, gamma=0, usageFactorPV=0.4,
-                            usageFactorSTC=0.2):
+    def calcPVAndSTCProfile(self, time, site, area_roof, beta=35, gamma=0, usageFactorPV=0.4,
+                            usageFactorSTC=0.2, devices=None):
         """
         Computation of power profiles for photovoltaic (PV) collectors and solar thermal collectors (STC).
 
@@ -505,14 +505,6 @@ class Sun:
 
         # profile of the ambient temperature
         temperatureProfile = site["T_e"]
-
-        #devices = {}
-        #with open(os.path.join(self.filePath, 'decentral_device_data.json')) as json_file:
-        #    jsonData = json.load(json_file)
-        #    for subData in jsonData:
-        #        devices[subData["abbreviation"]] = {}
-        #        for subsubData in subData["specifications"]:
-        #            devices[subData["abbreviation"]][subsubData["name"]] = subsubData["value"]
 
         # calculate time variant PV efficiency
         eta_PV = np.zeros(time["timeSteps"])
