+++ districtgenerator/data_handling/Archive/main_data_generator.py
@@ -1,113 +1,0 @@
-from districtgenerator.data_handling.config import LocationConfig, BuildingConfig, TimeConfig, DesignBuildingConfig, EcoConfig, PhysicsConfig, EHDOConfig, GurobiConfig, HeatGridConfig
-from districtgenerator.data_handling.site_data_generator import SiteDataGenerator
-from districtgenerator.data_handling.time_data_generator import TimeDataGenerator
-from districtgenerator.data_handling.design_building_data_generator import DesignBuildingDataGenerator
-from districtgenerator.data_handling.building_data_generator import BuildingDataGenerator
-from districtgenerator.data_handling.eco_data_generator import EcoDataGenerator
-from districtgenerator.data_handling.physics_data_generator import PhysicsDataGenerator
-from districtgenerator.data_handling.EHDO_generator import EHDOGenerator
-from districtgenerator.data_handling.gurobi_generator import GurobiGenerator
-from districtgenerator.data_handling.heat_grid_data_generator import HeatGridDataGenerator
-from districtgenerator.data_handling.central_device_config import CentralDeviceConfig
-from districtgenerator.data_handling.central_device_generator import CentralDeviceDataGenerator
-from districtgenerator.data_handling.decentral_device_config import DecentralDeviceConfig
-from districtgenerator.data_handling.decentral_device_data_generator import DecentralDeviceDataGenerator
-from districtgenerator.data_handling.json_adapter import JSONDataAdapter
-from typing import Dict, Any
-import json
-import pandas as pd
-import os
-
-class DataGenerator:
-    def __init__(self,
-                 location_config: LocationConfig,
-                 building_config: BuildingConfig,
-                 time_config: TimeConfig = None,
-                 design_building_config: DesignBuildingConfig = None,
-                 eco_config: EcoConfig = None,
-                 EHDO_config: EHDOConfig = None,
-                 physics_config: PhysicsConfig = None,
-                 gurobi_config: GurobiConfig = None,
-                 heat_grid_config: HeatGridConfig = None,
-                 central_device_config: CentralDeviceConfig = None,
-                 decentral_device_config = None):
-        self.location_config = location_config
-        self.building_config = building_config
-        self.time_config = time_config if time_config is not None else TimeConfig()
-        self.design_building_config = design_building_config if design_building_config is not None else DesignBuildingConfig()
-        self.eco_config = eco_config if eco_config is not None else EcoConfig()
-        self.physics_config = physics_config if physics_config is not None else PhysicsConfig()
-        self.EHDO_config = EHDO_config if EHDO_config is not None else EHDOConfig()  
-        self.gurobi_config = gurobi_config if gurobi_config is not None else GurobiConfig()
-        self.heat_grid_config = heat_grid_config if heat_grid_config is not None else HeatGridConfig()
-        self.central_device_config = central_device_config if central_device_config is not None else CentralDeviceConfig()
-        self.decentral_device_config = decentral_device_config if decentral_device_config is not None else DecentralDeviceConfig()
-        self.decentral_device_data_gen = DecentralDeviceDataGenerator(self.decentral_device_config)
-
-        self.site_data_gen = SiteDataGenerator(self.location_config)
-        self.time_data_gen = TimeDataGenerator(self.time_config)
-        self.design_data_gen = DesignBuildingDataGenerator(self.design_building_config)
-        self.eco_data_gen = EcoDataGenerator(self.eco_config)
-        self.building_data_gen = BuildingDataGenerator(self.building_config)
[…]
