+++ examples/e7_optimal_dimensioning_energyHub.py
@@ -14,7 +14,7 @@ def example7_optiEnergyCentral_EHDO():
 
     # We directly generate a complete district.
     data.generateDistrictComplete(calcUserProfiles=False, saveUserProfiles=False,
-                                  saveGenProfiles=False)
+                                  saveGenProfiles=True)
 
     # As last step we use the EHDO tool to get an optimized energy central for neighborhoods. EHDO is a tool for
     # planning and designing complex energy systems. The central feature is coupling of different
