+++ districtgenerator/classes/datahandler.py
@@ -626,7 +626,7 @@ class Datahandler:
 
     def saveProfiles(self, name, elec, dhw, occ, gains, car, nb_flats, nb_occ, heatload, bivalent, heatlimit, thick_req, path):
         """
-        Save profiles to csv.
+        Save profiles to separate parquet files.
 
         Parameters
         ----------
@@ -640,36 +640,65 @@ class Datahandler:
         None.
         """
 
+        # Create the directory with the specified name
+        directory_path = os.path.join(path, name)
+        os.makedirs(directory_path, exist_ok=True)
 
-        other_data = [nb_flats, str(nb_occ)[1:-1], heatload, bivalent, heatlimit]
+        # Create DataFrames directly from the input variables
+        elec_df = pd.DataFrame(elec, columns=['elec'])
+        dhw_df = pd.DataFrame(dhw, columns=['dhw'])
+        occ_df = pd.DataFrame(occ, columns=['occ'])
+        gains_df = pd.DataFrame(gains, columns=['gains'])
+        car_df = pd.DataFrame(car, columns=['car'])
+        nb_flats_df = pd.DataFrame([nb_flats], columns=['nb_flats'])
+        nb_occ_df = pd.DataFrame([nb_occ], columns=['nb_occ'])
+        heatload_df = pd.DataFrame([heatload], columns=['heatload'])
+        bivalent_df = pd.DataFrame([bivalent], columns=['bivalent'])
+        heatlimit_df = pd.DataFrame([heatlimit], columns=['heatlimit'])
 
-        data_dict = {
-            'elec': (elec, "Electricity demand in W"),
-            'dhw': (dhw, "Drinking hot water in W"),
-            'occ': (occ, "Occupancy of persons"),
-            'gains': (gains, "Internal gains in W"),
-            'car': (car, "Electricity demand of EV in W"),
-            'other': (other_data, "Number of flats , "
-                                  "Number of occupants, "
-                                  "Design heat load in W, "
-                                  "Bivalent heat load in W, "
-                                  "Heat limit heat load in W"),
-        }
-
+        # If thick_req is provided, create DataFrames for insulation
         if thick_req:
-            data_dict['wall_ins'] = ([thick_req[0]], "Insulation thickness required in m (wall)")
-            data_dict['roof_ins'] = ([thick_req[1]], "Insulation thickness required in m (roof)")
-            data_dict['floor_ins'] = ([thick_req[2]], "Insulation thickness required in m (floor)")
+            wall_ins_df = pd.DataFrame([thick_req[0]], columns=['wall_ins'])
+            roof_ins_df = pd.DataFrame([thick_req[1]], columns=['roof_ins'])
+            floor_ins_df = pd.DataFrame([thick_req[2]], columns=['floor_ins'])
 
-        excel_file = os.path.join(path, name + '.xlsx')
-        with pd.ExcelWriter(excel_file, engine='xlsxwriter') as writer:
-            for sheet_name, (data, header) in data_dict.items():
-                df = pd.DataFrame(data)
-                df.to_excel(writer, sheet_name=sheet_name, index=False, header=header)
+        # Define file paths for each DataFrame
+        elec_file = os.path.join(directory_path, 'elec.parquet')
+        dhw_file = os.path.join(directory_path, 'dhw.parquet')
+        occ_file = os.path.join(directory_path, 'occ.parquet')
+        gains_file = os.path.join(directory_path, 'gains.parquet')
+        car_file = os.path.join(directory_path, 'car.parquet')
+        nb_flats_file = os.path.join(directory_path, 'nb_flats.parquet')
+        nb_occ_file = os.path.join(directory_path, 'nb_occ.parquet')
[…]
