### 2025-07-08 da2340d
Removed concat
