### 2025-07-09 1494d2c
addded comparison of real and calc data for pv roof. Currently bugfixing
