+++ districtgenerator/classes/datahandler.py
@@ -793,30 +793,36 @@ class Datahandler:
                           file_path=self.filePath)
 
             # Parse |-separated strings into lists
-            if not pd.isna(building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"]):
+            if "surfaceAreaSuitableForSolarPV" in building["buildingFeatures"] and not pd.isna(building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"]):
                 roof_areas = [float(area) for area in
                               building["buildingFeatures"]["surfaceAreaSuitableForSolarPV"].split("|")]
                 roof_inclinations = [float(beta) for beta in building["buildingFeatures"]["roofInclination"].split("|")]
-                cardinal_directions = [float(gamma) for gamma in
-                                       building["buildingFeatures"]["cardinalDirection"].split("|")]
+                cardinal_directions = [
+                    (float(gamma) - 180) if gamma != '-1' else 0
+                    for gamma in building["buildingFeatures"]["cardinalDirection"].split("|")
+                ]
                 roof_shapes = building["buildingFeatures"]["roofShape"].split("|")
-
+#
                 # Initialize total potentials
                 total_potentialPV = None
                 total_potentialSTC = None
-
+#
                 pv_rows = []
                 individual_profiles = []
-
+                individual_profiles_prj = []
+#
                 pv_log_path = os.path.join(self.filePath, "logs", "pv_values_log.csv")
-
+#
                 # Calculate and save for each roof surface
                 for i in range(len(roof_areas)):
-                    # Create PV log entry for this roof
+
+                     # Create PV log entry for this roof
                     pv_row = {
                         "ID": f"{building['buildingFeatures']['gmlId']}_roof_{i + 1}",
                         "calculated_area_roof": building["envelope"].A["opaque"]["roof"],  # Calculated value
                         "actual_area_roof": roof_areas[i],  # Actual value
+                        "calculated_groundfloor": building["envelope"].A["opaque"]["floor"],
+                        "actual_groundfloor": building["buildingFeatures"]["area"],  # Actual value
                         "calculated_beta": 35,  # Calculated value
                         "actual_beta": roof_inclinations[i],  # Actual value
                         "calculated_gamma": building["buildingFeatures"]["gamma_PV"],  # Calculated value
@@ -824,8 +830,8 @@ class Datahandler:
                         "roofShape": roof_shapes[i]
                     }
                     pv_rows.append(pv_row)
-
-                    potentialPV, potentialSTC = \
+#
+                    potentialPV_prj, potentialSTC_prj = \
                         sun.calcPVAndSTCProfile(time=self.time,
                                                 site=self.site,
                                                 area_roof=roof_areas[i],
@@ -834,10 +840,21 @@ class Datahandler:
                                                 usageFactorPV=building["buildingFeatures"]["f_PV"],
                                                 usageFactorSTC=building["buildingFeatures"]["f_STC"],
                                                 devices=self.decentral_device_data)
-
+                    potentialPV, potentialSTC = \
+                        sun.calcPVAndSTCProfile(time=self.time,
+                                            site=self.site,
+                                            area_roof=building["envelope"].A["opaque"]["roof"], # todo: Dachfläche passt zu Datenplattform?
+                                            # In Germany, this is a roof pitch between 30 and 35 degrees
+                                            beta=[35],
+                                            # surface azimuth angles (Orientation to the south: 0°)
+                                            gamma=[building["buildingFeatures"]["gamma_PV"]],
+                                            usageFactorPV=building["buildingFeatures"]["f_PV"],
[…]
