### 2025-07-09 369c2be
Editorial suggested changes to JOSS paper (#79)
