+++ joss/paper.md
@@ -33,28 +33,28 @@ bibliography: paper.bib
 
 # Summary
 
-The `Districtgenerator` enables an automated generation of time-resolved load profiles for residential districts. 
+The `Districtgenerator` Python tool enables an automated generation of time-resolved load profiles for residential districts. 
 The profiles generated for each building within the district are the demands for electricity, space heating demand, domestic hot water, and the occupancy profiles. 
 Additionally, a heating load calculation is carried out for each building.
 The `Districtgenerator` is conceptualized using a minimum amount of input data. 
-Thus, the tool is valuable for researchers and planners to obtain information  needed, for example, for the energy system or energy management system design in an early planning phase of a district.
+Thus, the tool is valuable for researchers and planners to obtain information  needed, for example, for the energy system or energy management system design during the early planning phase of a new district.
 
 # Statement of need
 
-An essential part of facing global warming is to reduce the carbon footprint of the building sector, especially with a growing world population and a trend of urbanization in mind [@IEA_2021].
-The `Districtgenerator` offers knowledge, shows synergy potentials and supports scalability regarding building and district solutions.  
+An essential part of mitigating global warming is reducing the carbon footprint of the building sector, especially considering the growing world population and the trend towards urbanization [@IEA_2021].
+The `Districtgenerator` tool offers knowledge, shows synergy potentials, and supports scalability regarding building and district solutions.  
 
-* **Knowledge** of the energy demands is fundamental for the proper design and operation of any energy system.
-With the integration of distributed energy resources (DER), district energy systems and building energy systems get more complex.
-The energy supply of highly variable generators and user-dependent consumption need to be aligned. 
-Thus, detailed time-resolved profiles of energy demand and generation are required for a successful and future-oriented energy system design and operation. 
+* **Knowledge** of energy demands is fundamental for the proper design and operation of any energy system.
+With the integration of distributed energy resources (DER), district energy systems and building energy systems become more complex.
+The energy supply from highly variable generators and user-dependent consumption need to be aligned. 
+Thus, detailed time-resolved profiles of energy generation and demand are required for a successful and future-oriented energy system design and operation. 
 
 * **Synergy** potentials between buildings of different usage structures can be exploited through joint concepts.
 While joint heating concepts, such as local heating networks, have been the focus of research for a long time, joint electricity concepts have only recently emerged.
 For example, the European Union proposed and introduced the concept of energy communities [@REDII_2018; @IEMD_2019].
 According to the Renewable Energy Directive [@REDII_2018], a district can also form an energy community where energy is shared or traded Peer-to-Peer.
-By this, district concepts can lead to financial benefits for consumers and prosumers. Moreover, its expected that self-consumption increases, 
-peaks are evened out and the superordinate grid might be decongested. [@bose_reinforcement_2021]
+By this, district concepts can lead to financial benefits for consumers and prosumers. Moreover, it is expected that self-consumption increases, 
+peaks are evened out, and the superordinate grid might be decongested. [@bose_reinforcement_2021]
 
 * **Scalability** is a key issue for a resource-efficient transformation of the energy system, especially as numerous existing districts will have to be retrofitted and newly equipped in the next few years [@IEA_2021].
 The `Districtgenerator` enables its users to easily gain information about districts in order to identify promising concepts and conduct comparative evaluations.
@@ -64,16 +64,16 @@ Thus, the data applies to a wide range of residential districts and can be inclu
 Also existing tools facilitate during the strategic and preliminary planning phases the identification of measures that 
[…]
