### 2025-07-11 46df3f9
Updates JOSS paper
Moves figure to avoid large empty space in paper
