### 2025-07-13 #7657 @JoelSchoelzel
@kyleniemeyer I reviewed and merged it
